Choose best matches

Find a set of vertices pairs in the order of goodness of matching according to a specified measure.

best_matches(A, B, measure, num, x, match_corr)

Arguments

A
A matrix or an igraph object. Adjacency matrix of \(G_1\).

B
A matrix or an igraph object. Adjacency matrix of \(G_2\).

measure
A character. Measure for computing goodness of matching.

num
An integer. Number of pairs of best matched vertices needed.

x
A vector of logical. TRUE indicates the corresponding vertex in \(G_1\) is of interest in finding best matched vertices. Length of vector should be the number of vertices of graphs.

match_corr
A matrix or data frame. Graph matching correspondence between \(G_1\) and \(G_2\) with the first column represents indices in \(G_1\) and the second column represents indices in \(G_2\).

Value

best_matches returns a data frame with the indices of best matched vertices in \(G_1\) named A_best and the indices of best matched vertices in \(G_2\) named B_best.

Examples

cgnp_pair <- sample_correlated_gnp_pair(n = 50, corr =  0.3, p =  0.5)
g1 <- cgnp_pair$graph1
g2 <- cgnp_pair$graph2
seeds <- 1:50 <= 10
nonseeds <- !seeds
match <- graph_match_FW(g1, g2, seeds)

# Application: select best matched seeds from non seeds as new seeds, and do the
# graph matching iteratively to get higher matching accuracy
best_matches(g1, g2, "row_perm_stat", num = 5, x = nonseeds, match$corr)

#>   A_best B_best
#> 1     32     26
#> 2     43     43
#> 3     31     17
#> 4     23     33
#> 5     19     35