Center adjacency matrix

Center the adjacency matrix, including center the adjacency matrix to entries equal to -1 or 1, center the adjacency matrix by using Universal Singular Value Thresholding.

center_graph(A, scheme, use_splr = TRUE)

Arguments

A
A matrix or an igraph object. Adjacency matrix.

scheme
A number or a character. Type of matrix to return. If approximatedly center the adjacency matrix using Universal Singular Value Thresholding, specify scheme with the number of singular values used to estimate the expectation of adjacency matrix. Suitable choice of scheme can be obtained with rank of order logn, where n is the dimension of A, expecially in the setting of latent space graph models.

use_splr
A boolean indicating whether to use the splrMatrix object when storing the centered graph. Defaults to TRUE.

Value

center_graph returns a centered adjacency matrix. 'Naive' scheme returns the original adjacency matrix. 'Center' scheme returns a centered adjacency matrix with entries equal to -1 or 1 where 1 corresponds to an edge. Scheme specified by a number returns a centered adjacency matrix, which is calcalated by: A - Q, where Q is an approximate estimation of the expectation of A by using Singular Value Thresholding.

Examples

A <- sample_correlated_gnp_pair(n = 10, corr = .5, p = .5)$graph1
center_graph(A, scheme = "naive")

#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                          
#>  [1,] . . . 1 1 1 1 1 . .
#>  [2,] . . 1 1 . 1 . 1 . .
#>  [3,] . 1 . 1 1 1 . . 1 .
#>  [4,] 1 1 1 . . 1 . 1 1 1
#>  [5,] 1 . 1 . . 1 1 . . .
#>  [6,] 1 1 1 1 1 . . 1 . 1
#>  [7,] 1 . . . 1 . . . 1 .
#>  [8,] 1 1 . 1 . 1 . . 1 .
#>  [9,] . . 1 1 . . 1 1 . 1
#> [10,] . . . 1 . 1 . . 1 .

center_graph(A, scheme = "center")

#> Sparse part
#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                          
#>  [1,] . . . 2 2 2 2 2 . .
#>  [2,] . . 2 2 . 2 . 2 . .
#>  [3,] . 2 . 2 2 2 . . 2 .
#>  [4,] 2 2 2 . . 2 . 2 2 2
#>  [5,] 2 . 2 . . 2 2 . . .
#>  [6,] 2 2 2 2 2 . . 2 . 2
#>  [7,] 2 . . . 2 . . . 2 .
#>  [8,] 2 2 . 2 . 2 . . 2 .
#>  [9,] . . 2 2 . . 2 2 . 2
#> [10,] . . . 2 . 2 . . 2 .
#> plus left factor
#> 10 x 1 Matrix of class "dgeMatrix"
#>       [,1]
#>  [1,]   -1
#>  [2,]   -1
#>  [3,]   -1
#>  [4,]   -1
#>  [5,]   -1
#>  [6,]   -1
#>  [7,]   -1
#>  [8,]   -1
#>  [9,]   -1
#> [10,]   -1
#> times right factor transpose
#> 10 x 1 Matrix of class "dgeMatrix"
#>       [,1]
#>  [1,]    1
#>  [2,]    1
#>  [3,]    1
#>  [4,]    1
#>  [5,]    1
#>  [6,]    1
#>  [7,]    1
#>  [8,]    1
#>  [9,]    1
#> [10,]    1

center_graph(A, scheme = 2)

#> Sparse part
#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                          
#>  [1,] . . . 1 1 1 1 1 . .
#>  [2,] . . 1 1 . 1 . 1 . .
#>  [3,] . 1 . 1 1 1 . . 1 .
#>  [4,] 1 1 1 . . 1 . 1 1 1
#>  [5,] 1 . 1 . . 1 1 . . .
#>  [6,] 1 1 1 1 1 . . 1 . 1
#>  [7,] 1 . . . 1 . . . 1 .
#>  [8,] 1 1 . 1 . 1 . . 1 .
#>  [9,] . . 1 1 . . 1 1 . 1
#> [10,] . . . 1 . 1 . . 1 .
#> plus left factor
#> 10 x 2 Matrix of class "dgeMatrix"
#>            [,1]        [,2]
#>  [1,] 1.6052764 -0.34741168
#>  [2,] 1.5297273 -0.61136564
#>  [3,] 1.6869922  1.15956139
#>  [4,] 2.2225141 -0.00999191
#>  [5,] 1.2308767 -0.35506731
#>  [6,] 2.1797139 -0.66872171
#>  [7,] 0.8437653  0.80762453
#>  [8,] 1.7599590  1.15670323
#>  [9,] 1.4932886 -1.46075405
#> [10,] 1.1489778  0.79875193
#> times right factor transpose
#> 10 x 2 Matrix of class "dgeMatrix"
#>             [,1]         [,2]
#>  [1,] -0.3128525 -0.129703177
#>  [2,] -0.2981287 -0.228248124
#>  [3,] -0.3287781  0.432912312
#>  [4,] -0.4331460 -0.003730394
#>  [5,] -0.2398857 -0.132561340
#>  [6,] -0.4248047 -0.249661522
#>  [7,] -0.1644415  0.301519699
#>  [8,] -0.3429986  0.431845241
#>  [9,] -0.2910272 -0.545360011
#> [10,] -0.2239244  0.298207189

center_graph(A, scheme = 1)

#> Sparse part
#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                          
#>  [1,] . . . 1 1 1 1 1 . .
#>  [2,] . . 1 1 . 1 . 1 . .
#>  [3,] . 1 . 1 1 1 . . 1 .
#>  [4,] 1 1 1 . . 1 . 1 1 1
#>  [5,] 1 . 1 . . 1 1 . . .
#>  [6,] 1 1 1 1 1 . . 1 . 1
#>  [7,] 1 . . . 1 . . . 1 .
#>  [8,] 1 1 . 1 . 1 . . 1 .
#>  [9,] . . 1 1 . . 1 1 . 1
#> [10,] . . . 1 . 1 . . 1 .
#> plus left factor
#> 10 x 1 Matrix of class "dgeMatrix"
#>             [,1]
#>  [1,] -0.3128525
#>  [2,] -0.2981287
#>  [3,] -0.3287781
#>  [4,] -0.4331460
#>  [5,] -0.2398857
#>  [6,] -0.4248047
#>  [7,] -0.1644415
#>  [8,] -0.3429986
#>  [9,] -0.2910272
#> [10,] -0.2239244
#> times right factor transpose
#> 10 x 1 Matrix of class "dgeMatrix"
#>            [,1]
#>  [1,] 0.3128525
#>  [2,] 0.2981287
#>  [3,] 0.3287781
#>  [4,] 0.4331460
#>  [5,] 0.2398857
#>  [6,] 0.4248047
#>  [7,] 0.1644415
#>  [8,] 0.3429986
#>  [9,] 0.2910272
#> [10,] 0.2239244