Get Permutation

Get an m-by-n permutation matrix according to the mapping correspondence.

get_perm(m, n, corr)

Arguments

m
An integer. Order of \(G_1\).

n
An integer. Order of \(G_2\).

corr
A matrix or a data frame. Matching correspondence with the first and second columns correspond to indices in \(G_1\) and \(G_2\) respectively.

Value

get_perm returns an m-by-n sparse permutation matrix or whose submatrix is a permutation matrix if only parts of nodes from both graphs get matched or in the case of matching graphs of different order.

Examples

# returns a permutation matrix: m=n, all the nodes get matched
corr <- data.frame(corr_A = c(1,2,3,4), corr_B = c(1,4,2,3))
get_perm(4, 4, corr)

#> 4 x 4 sparse Matrix of class "dgCMatrix"
#>             
#> [1,] 1 . . .
#> [2,] . . . 1
#> [3,] . 1 . .
#> [4,] . . 1 .

# submatrix is a permutation matrix: parts of graphs get matched
get_perm(5, 6, corr)

#> 5 x 6 sparse Matrix of class "dgCMatrix"
#>                 
#> [1,] 1 . . . . .
#> [2,] . . . 1 . .
#> [3,] . 1 . . . .
#> [4,] . . 1 . . .
#> [5,] . . . . . .