Spectral Graph Matching Methods: Umeyama Algorithm

Spectral Graph Matching Methods: Umeyama Algorithm

graph_match_Umeyama(A, B, similarity = NULL, seeds = NULL, alpha = 0.5)

Arguments

A
A matrix, igraph object, or list of either.

B
A matrix, igraph object, or list of either.

similarity
A matrix. An n-by-n matrix containing vertex similaities.

seeds
A vector of integers or logicals, a matrix or a data frame. If the seed pairs have the same indices in both graphs then seeds can be a vector. If not, seeds must be a matrix. or a data frame, with the first column being the indices of \(G_1\) and the second column being the corresponding indices of \(G_2\).

alpha
A number betwen 0 and 1. Bigger alpha means putting more importance on the information in network topology over other information such as similarity scores.

Value

graph_match_Umeyama returns a list of graph matching results, including the graph matching formula, a data frame containing the matching correspondence between \(G_1\) and \(G_2\) named corr_A and corr_B and the number of seeds.

References

S. Umeyama (1988), An eigendecomposition approach to weighted graph matching problems. IEEE TPAMI. USA, pages 695-703.

Examples

# match G_1 & G_2 using Umeyama algorithm
G <- sample_correlated_gnp_pair(10, .9, .5)
g1 <- G$graph1
g2 <- G$graph2
startm <- matrix(0, 10, 10)
diag(startm)[1:4] <- 1
graph_match_Umeyama(g1, g2, startm, alpha = .3)

#> $call
#> graph_match_Umeyama(A = g1, B = g2, similarity = startm, alpha = 0.3)
#> 
#> $corr
#>    corr_A corr_B
#> 1       1      1
#> 2       2      2
#> 3       3      3
#> 4       4      4
#> 5       5      5
#> 6       6      6
#> 7       7      7
#> 8       8      8
#> 9       9      9
#> 10     10     10
#> 
#> $ns
#> [1] 0
#>