Graph Match Methods

Match two given graphs, returns a list of graph matching results, including matching correspondence vector of \(G_2\) with respect to \(G_1\), doubly stochastic matrix and permutation matrix.

graph_match_FW(
  A,
  B,
  seeds = NULL,
  start = "convex",
  max_iter = 20,
  similarity = NULL,
  lap_method = NULL
)

gm_indefinite(
  A,
  B,
  seeds = NULL,
  start = "convex",
  max_iter = 20,
  similarity = NULL,
  lap_method = NULL
)

graph_match_convex(
  A,
  B,
  seeds = NULL,
  start = "bari",
  max_iter = 100,
  similarity = NULL,
  tol = 1e-05,
  lap_method = NULL
)

graph_match_PATH(
  A,
  B,
  similarity = NULL,
  seeds = NULL,
  alpha = 0.5,
  epsilon = 1
)

graph_match_percolation(A, B, seeds, r = 2)

graph_match_ExpandWhenStuck(A, B, seeds, r = 2)

graph_match_soft_percolation(A, B, seeds, r = 2, max_iter = 100)

graph_match_IsoRank(
  A,
  B,
  similarity,
  seeds = NULL,
  alpha = 0.5,
  max_iter = 50,
  method = "greedy"
)

graph_match_Umeyama(A, B, similarity = NULL, seeds = NULL, alpha = 0.5)

Arguments

A
A matrix or an igraph object. Adjacency matrix of \(G_1\).

B
A matrix or an igraph object. Adjacency matrix of \(G_2\).

seeds
A vector of integers or logicals, a matrix or a data frame. If there is no error in seeds input can be a vector of seed indices in \(G_1\). Or if there exists error in seeds, input in the form of a matrix or a data frame, with the first column being the indices of \(G_1\) and the second column being the corresponding indices of \(G_2\).

start
A matrix or a character. Any nns-by-nns matrix or character value like "bari" or "convex" to initialize the starting matrix.

max_iter
A number. Maximum number of replacing matches equals to max_iter times number of total vertices of \(G_1\).

similarity
A matrix. An n-by-n matrix containing vertex similaities.

lap_method
Choice for lap method.

tol
A number. Tolerance of edge disagreements.

alpha
A number betwen 0 and 1. Bigger alpha means putting more importance on the information in network topology over other information such as similarity scores

epsilon
A small number

r
A number. Threshold of neighboring pair scores.

method
A character. Choice of method to extract mapping from score matrix, including greedy method and the Hungarian algorithm.

Value

graph_match_FW returns a list of graph matching results, including the graph matching formula, a data frame containing the matching correspondence between \(G_1\) and \(G_2\) named corr_A and corr_B and the number of seeds.

graph_match_convex returns a list of graph matching results, including the graph matching formula, a data frame containing the matching correspondence between \(G_1\) and \(G_2\) named corr_A and corr_B and the number of seeds.

graph_match_PATH returns a list of graph matching results, including the graph matching formula, a data frame containing the matching correspondence between \(G_1\) and \(G_2\) named corr_A and corr_B, the number of seeds if any, the permutation matrix and the doubly stochastic matrix before projection onto the permutation set.

graph_match_percolation returns a list of graph matching results, including the graph matching formula, a data frame containing the matching correspondence between \(G_1\) and \(G_2\) named corr_A and corr_B, the number of seeds and the order of nodes getting matched.

graph_match_ExpandWhenStuck returns a list of graph matching results, including the graph matching formula, a data frame containing the matching correspondence between \(G_1\) and \(G_2\) named corr_A and corr_B, the number of seeds and the order of nodes getting matched.

graph_match_soft_percolation returns a list of graph matching results, including the graph matching formula, a data frame containing the matching correspondence between \(G_1\) and \(G_2\) named corr_A and corr_B, the number of seeds and the order of nodes getting matched.

graph_match_IsoRank returns a list of graph matching results, including the graph matching formula, a data frame containing the matching correspondence between \(G_1\) and \(G_2\) named corr_A and corr_B and the number of seeds. If choose the greedy method to extract mapping, the order of nodes getting matched will also be returned.

graph_match_Umeyama returns a list of graph matching results, including the graph matching formula, a data frame containing the matching correspondence between \(G_1\) and \(G_2\) named corr_A and corr_B and the number of seeds.

References

M. Zaslavskiy, F. Bach and J. Vert (2009), A Path following algorithm for the graph matching problem. IEEE Trans Pattern Anal Mach Intell, pages 2227-2242.

L. Yartseva and M. Grossglauser (2013), On the performance of percolation graph matching. COSN, Boston, MA, USA, pages 119–130.

E. Kazemi, S. H. Hassani, and M. Grossglauser (2015), Growing a graph matching from a handful of seeds. Proc. of the VLDB Endowment, 8(10):1010–1021.

R. Singh, J. Xu, B. Berger (2008), Global alignment of multiple protein interaction networks with application to functional orthology detection. Proc Natl Acad Sci. USA, pages 12763-12768.

S. Umeyama (1988), An eigendecomposition approach to weighted graph matching problems. IEEE TPAMI. USA, pages 695-703.

Examples

cgnp_pair <- sample_correlated_gnp_pair(n = 10, corr =  0.3, p =  0.5)
g1 <- cgnp_pair$graph1
g2 <- cgnp_pair$graph2
# match G_1 & G_2 with no seeds
graph_match_FW(g1, g2)

#> $call
#> graph_match_FW(A = g1, B = g2)
#> 
#> $corr
#>    corr_A corr_B
#> 1       1      4
#> 2       2      1
#> 3       3      9
#> 4       4      7
#> 5       5      6
#> 6       6      5
#> 7       7      2
#> 8       8      8
#> 9       9      3
#> 10     10     10
#> 
#> $ns
#> [1] 0
#> 
#> $P
#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                          
#>  [1,] . . . 1 . . . . . .
#>  [2,] 1 . . . . . . . . .
#>  [3,] . . . . . . . . 1 .
#>  [4,] . . . . . . 1 . . .
#>  [5,] . . . . . 1 . . . .
#>  [6,] . . . . 1 . . . . .
#>  [7,] . 1 . . . . . . . .
#>  [8,] . . . . . . . 1 . .
#>  [9,] . . 1 . . . . . . .
#> [10,] . . . . . . . . . 1
#> 
#> $D
#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                          
#>  [1,] . . . 1 . . . . . .
#>  [2,] 1 . . . . . . . . .
#>  [3,] . . . . . . . . 1 .
#>  [4,] . . . . . . 1 . . .
#>  [5,] . . . . . 1 . . . .
#>  [6,] . . . . 1 . . . . .
#>  [7,] . 1 . . . . . . . .
#>  [8,] . . . . . . . 1 . .
#>  [9,] . . 1 . . . . . . .
#> [10,] . . . . . . . . . 1
#> 
#> $num_iter
#> [1] 2
#> 

# match G_1 & G_2 with some known node pairs as seeds
seeds <- 1:10 <= 3
graph_match_FW(g1, g2, seeds, start = "bari")

#> $call
#> graph_match_FW(A = g1, B = g2, seeds = seeds, start = "bari")
#> 
#> $corr
#>    corr_A corr_B
#> 1       1      1
#> 2       2      2
#> 3       3      3
#> 4       4      4
#> 5       5      5
#> 6       6      8
#> 7       7      7
#> 8       8      9
#> 9       9      6
#> 10     10     10
#> 
#> $ns
#> [1] 3
#> 
#> $P
#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                          
#>  [1,] 1 . . . . . . . . .
#>  [2,] . 1 . . . . . . . .
#>  [3,] . . 1 . . . . . . .
#>  [4,] . . . 1 . . . . . .
#>  [5,] . . . . 1 . . . . .
#>  [6,] . . . . . . . 1 . .
#>  [7,] . . . . . . 1 . . .
#>  [8,] . . . . . . . . 1 .
#>  [9,] . . . . . 1 . . . .
#> [10,] . . . . . . . . . 1
#> 
#> $D
#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                          
#>  [1,] 1 . . . . . . . . .
#>  [2,] . 1 . . . . . . . .
#>  [3,] . . 1 . . . . . . .
#>  [4,] . . . 1 . . . . . .
#>  [5,] . . . . 1 . . . . .
#>  [6,] . . . . . . . 1 . .
#>  [7,] . . . . . . 1 . . .
#>  [8,] . . . . . . . . 1 .
#>  [9,] . . . . . 1 . . . .
#> [10,] . . . . . . . . . 1
#> 
#> $num_iter
#> [1] 3
#> 

# match G_1 & G_2 with some incorrect seeds
hard_seeds <- matrix(c(4,6,5,4),2)
seeds <- rbind(as.matrix(check_seeds(seeds, nv = 10)$seeds),hard_seeds)
graph_match_FW(g1, g2, seeds, start = "convex")

#> $call
#> graph_match_FW(A = g1, B = g2, seeds = seeds, start = "convex")
#> 
#> $corr
#>    corr_A corr_B
#> 1       1      1
#> 2       2      2
#> 3       3      3
#> 4       4      5
#> 5       5     10
#> 6       6      4
#> 7       7      7
#> 8       8      9
#> 9       9      8
#> 10     10      6
#> 
#> $ns
#> [1] 5
#> 
#> $P
#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                          
#>  [1,] 1 . . . . . . . . .
#>  [2,] . 1 . . . . . . . .
#>  [3,] . . 1 . . . . . . .
#>  [4,] . . . . 1 . . . . .
#>  [5,] . . . . . . . . . 1
#>  [6,] . . . 1 . . . . . .
#>  [7,] . . . . . . 1 . . .
#>  [8,] . . . . . . . . 1 .
#>  [9,] . . . . . . . 1 . .
#> [10,] . . . . . 1 . . . .
#> 
#> $D
#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                          
#>  [1,] 1 . . . . . . . . .
#>  [2,] . 1 . . . . . . . .
#>  [3,] . . 1 . . . . . . .
#>  [4,] . . . . 1 . . . . .
#>  [5,] . . . . . . . . . 1
#>  [6,] . . . 1 . . . . . .
#>  [7,] . . . . . . 1 . . .
#>  [8,] . . . . . . . . 1 .
#>  [9,] . . . . . . . 1 . .
#> [10,] . . . . . 1 . . . .
#> 
#> $num_iter
#> [1] 2
#> 

 gp_list <- replicate(3, sample_correlated_gnp_pair(100, .3, .5), simplify = FALSE)
 A <- lapply(gp_list, function(gp)gp[[1]])
 B <- lapply(gp_list, function(gp)gp[[2]])
 match <- graph_match_FW(A, B, seeds = 1:10, start = "bari", max_iter = 20)
 match$corr

#>     corr_A corr_B
#> 1        1      1
#> 2        2      2
#> 3        3      3
#> 4        4      4
#> 5        5      5
#> 6        6      6
#> 7        7      7
#> 8        8      8
#> 9        9      9
#> 10      10     10
#> 11      11     11
#> 12      12     12
#> 13      13     13
#> 14      14     14
#> 15      15     15
#> 16      16     16
#> 17      17     17
#> 18      18     18
#> 19      19     19
#> 20      20     20
#> 21      21     21
#> 22      22     22
#> 23      23     23
#> 24      24     24
#> 25      25     25
#> 26      26     26
#> 27      27     27
#> 28      28     28
#> 29      29     29
#> 30      30     30
#> 31      31     31
#> 32      32     32
#> 33      33     33
#> 34      34     34
#> 35      35     35
#> 36      36     36
#> 37      37     37
#> 38      38     38
#> 39      39     39
#> 40      40     40
#> 41      41     41
#> 42      42     42
#> 43      43     43
#> 44      44     44
#> 45      45     45
#> 46      46     46
#> 47      47     47
#> 48      48     48
#> 49      49     49
#> 50      50     50
#> 51      51     51
#> 52      52     52
#> 53      53     53
#> 54      54     54
#> 55      55     55
#> 56      56     56
#> 57      57     57
#> 58      58     58
#> 59      59     59
#> 60      60     60
#> 61      61     61
#> 62      62     62
#> 63      63     63
#> 64      64     64
#> 65      65     65
#> 66      66     66
#> 67      67     67
#> 68      68     68
#> 69      69     69
#> 70      70     70
#> 71      71     71
#> 72      72     72
#> 73      73     73
#> 74      74     74
#> 75      75     75
#> 76      76     76
#> 77      77     77
#> 78      78     78
#> 79      79     79
#> 80      80     80
#> 81      81     81
#> 82      82     82
#> 83      83     83
#> 84      84     84
#> 85      85     85
#> 86      86     86
#> 87      87     87
#> 88      88     88
#> 89      89     89
#> 90      90     90
#> 91      91     91
#> 92      92     92
#> 93      93     93
#> 94      94     94
#> 95      95     95
#> 96      96     96
#> 97      97     97
#> 98      98     98
#> 99      99     99
#> 100    100    100

seeds <- 1:10 <= 3
graph_match_convex(g1, g2, seeds)

#> $call
#> graph_match_convex(A = g1, B = g2, seeds = seeds)
#> 
#> $corr
#>    corr_A corr_B
#> 1       1      1
#> 2       2      2
#> 3       3      3
#> 4       4      9
#> 5       5      5
#> 6       6      8
#> 7       7      7
#> 8       8      4
#> 9       9      6
#> 10     10     10
#> 
#> $ns
#> [1] 3
#> 
#> $P
#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                          
#>  [1,] 1 . . . . . . . . .
#>  [2,] . 1 . . . . . . . .
#>  [3,] . . 1 . . . . . . .
#>  [4,] . . . . . . . . 1 .
#>  [5,] . . . . 1 . . . . .
#>  [6,] . . . . . . . 1 . .
#>  [7,] . . . . . . 1 . . .
#>  [8,] . . . 1 . . . . . .
#>  [9,] . . . . . 1 . . . .
#> [10,] . . . . . . . . . 1
#> 
#> $D
#> Sparse part
#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                                                                              
#>  [1,] 1 . . .          .          .          .          .          .         
#>  [2,] . 1 . .          .          .          .          .          .         
#>  [3,] . . 1 .          .          .          .          .          .         
#>  [4,] . . . 0.25379158 0.23624274 0.02987928 0.15261265 .          0.18530399
#>  [5,] . . . 0.10171569 0.35773462 0.04090133 0.05781589 0.22191666 .         
#>  [6,] . . . 0.02761104 0.04456679 0.11069092 0.33427950 0.42650294 .         
#>  [7,] . . . 0.13685503 0.17125671 0.15678282 0.36094589 0.03970156 0.11901052
#>  [8,] . . . 0.32405512 0.11981019 .          0.07889860 0.15029744 0.20039768
#>  [9,] . . . .          0.01404013 0.33504886 .          0.11625465 0.40057264
#> [10,] . . . 0.14052407 0.04090133 0.31124932 .          0.02987928 0.07926769
#>                 
#>  [1,] .         
#>  [2,] .         
#>  [3,] .         
#>  [4,] 0.12672228
#>  [5,] 0.20446833
#>  [6,] 0.04090133
#>  [7,] .         
#>  [8,] 0.11109349
#>  [9,] 0.11863624
#> [10,] 0.38273084
#> plus left factor
#> 10 x 1 sparse Matrix of class "dgCMatrix"
#>                 
#>  [1,] .         
#>  [2,] .         
#>  [3,] .         
#>  [4,] 0.01544748
#>  [5,] 0.01544748
#>  [6,] 0.01544748
#>  [7,] 0.01544748
#>  [8,] 0.01544748
#>  [9,] 0.01544748
#> [10,] 0.01544748
#> times right factor transpose
#> 10 x 1 sparse Matrix of class "dgCMatrix"
#>                
#>  [1,] .        
#>  [2,] .        
#>  [3,] .        
#>  [4,] 0.1428571
#>  [5,] 0.1428571
#>  [6,] 0.1428571
#>  [7,] 0.1428571
#>  [8,] 0.1428571
#>  [9,] 0.1428571
#> [10,] 0.1428571
#> 
#> $num_iter
#> [1] 100
#> 

hard_seeds <- matrix(c(4,6,5,4),2)
seeds <- rbind(as.matrix(check_seeds(seeds, 10)$seeds),hard_seeds)
if (FALSE) {
# for some reason this fails in check
graph_match_convex(g1, g2, seeds)
}

# match G_1 & G_2 using PATH algorithm
graph_match_PATH(g1, g2)

#> $call
#> graph_match_PATH(A = g1, B = g2)
#> 
#> $corr
#>    corr_A corr_B
#> 1       1      1
#> 2       2      2
#> 3       3      3
#> 4       4      4
#> 5       5      5
#> 6       6      6
#> 7       7      7
#> 8       8      8
#> 9       9      9
#> 10     10     10
#> 
#> $ns
#> [1] 0
#> 
#> $P
#> 10 x 10 diagonal matrix of class "ddiMatrix"
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    1    .    .    .    .    .    .    .    .     .
#>  [2,]    .    1    .    .    .    .    .    .    .     .
#>  [3,]    .    .    1    .    .    .    .    .    .     .
#>  [4,]    .    .    .    1    .    .    .    .    .     .
#>  [5,]    .    .    .    .    1    .    .    .    .     .
#>  [6,]    .    .    .    .    .    1    .    .    .     .
#>  [7,]    .    .    .    .    .    .    1    .    .     .
#>  [8,]    .    .    .    .    .    .    .    1    .     .
#>  [9,]    .    .    .    .    .    .    .    .    1     .
#> [10,]    .    .    .    .    .    .    .    .    .     1
#> 
#> $D
#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                                                                              
#>  [1,] 0.7295158 .         .         0.2704842 . .         .         .        
#>  [2,] 0.2704842 0.7295158 .         .         . .         .         .        
#>  [3,] .         .         0.7295158 .         . .         .         .        
#>  [4,] .         .         .         0.7295158 . .         0.2704842 .        
#>  [5,] .         .         .         .         1 .         .         .        
#>  [6,] .         .         .         .         . 0.7295158 .         0.2704842
#>  [7,] .         0.2704842 .         .         . .         0.7295158 .        
#>  [8,] .         .         .         .         . 0.2704842 .         0.7295158
#>  [9,] .         .         0.2704842 .         . .         .         .        
#> [10,] .         .         .         .         . .         .         .        
#>                  
#>  [1,] .         .
#>  [2,] .         .
#>  [3,] 0.2704842 .
#>  [4,] .         .
#>  [5,] .         .
#>  [6,] .         .
#>  [7,] .         .
#>  [8,] .         .
#>  [9,] 0.7295158 .
#> [10,] .         1
#> 
#> $iter
#> [1] 2
#> 
#> $lambda
#> [1] 1
#> 

# match G_1 & G_2 using percolation graph matching method
seeds <- 1:5
graph_match_percolation(g1, g2, seeds, r = 2)

#> $call
#> graph_match_percolation(A = g1, B = g2, seeds = seeds, r = 2)
#> 
#> $corr
#>    corr_A corr_B
#> 1       1      1
#> 2       2      2
#> 3       3      3
#> 4       4      4
#> 5       5      5
#> 6       6      8
#> 7       7      7
#> 8       8      9
#> 9       9      6
#> 10     10     10
#> 
#> $ns
#> [1] 5
#> 
#> $order
#>  [1]  1  2  3  4  5  9 10  8  7  6
#> 

# match G_1 & G_2 using Expand When Stuck graph matching method
seeds <- 1:5
graph_match_ExpandWhenStuck(g1, g2, seeds, r = 2)

#> $call
#> graph_match_ExpandWhenStuck(A = g1, B = g2, seeds = seeds, r = 2)
#> 
#> $corr
#>    corr_A corr_B
#> 1       1      1
#> 2       2      2
#> 3       3      3
#> 4       4      4
#> 5       5      5
#> 6       6      8
#> 7       7      7
#> 8       8      9
#> 9       9      6
#> 10     10     10
#> 
#> $ns
#> [1] 5
#> 
#> $order
#>  [1]  1  2  3  4  5  9 10  8  7  6
#> 

# match G_1 & G_2 using soft percolation graph matching method
seeds <- 1:5
graph_match_soft_percolation(g1, g2, seeds, r = 2, max_iter = 2)

#> $call
#> graph_match_soft_percolation(A = g1, B = g2, seeds = seeds, r = 2, 
#>     max_iter = 2)
#> 
#> $corr
#>    corr_A corr_B
#> 1       1      1
#> 2       2      2
#> 3       3      3
#> 4       4      4
#> 5       5      5
#> 10      6      7
#> 9       7      8
#> 8       8      6
#> 7       9      9
#> 6      10     10
#> 
#> $ns
#> [1] 5
#> 
#> $order
#>  [1]  1  2  3  4  5 10  9  8  7  6
#> 

# match G_1 & G_2 using IsoRank algorithm
startm <- matrix(0, 10, 10)
diag(startm)[1:4] <- 1
GM_IsoRank <- graph_match_IsoRank(g1, g2, startm, alpha = .3, method = "greedy")

# match G_1 & G_2 using Umeyama algorithm
G <- sample_correlated_gnp_pair(10, .9, .5)
G1 <- G$graph1
G2 <- G$graph2
GM_U <- graph_match_Umeyama(G1, G2, startm, alpha = .3)