Initialization of the start matrix

Initialize the start matrix for graph matching iteration.

init_start(start, nns, ns = 0, soft_seeds = NULL, ...)

Arguments

start
A matrix or a character. Any nns-by-nns doubly stochastic matrix or start method like "bari", "convex" or "rds" to initialize the start matrix.

nns
An integer. Number of non-seeds.

ns
An integer. Number of seeds.

soft_seeds
A vector, a matrix or a data frame. If there is no error in the soft seeds, input can be a vector of soft seed indices in \(G_1\). Or if there is error in soft seeds, input should be in the form of a matrix or a data frame, with the first column being the indices of \(G_1\) and the second column being the corresponding indices of \(G_2\). Note that if there are seeds in graphs, seeds should be put before non-seeds.

...
Arguments passed to other start functions

Value

init_start returns a nns-by-nns doubly stochastic matrix as the start matrix in the graph matching iteration. If conduct a soft seeding graph matching, returns a nns-by-nns doubly stochastic matrix with 1's corresponding to the soft seeds and values at the other places are derived by different start method.

Examples

ss <- matrix(c(5, 4, 4, 3), nrow = 2)
# initialize start matrix without soft seeds
init_start(start = "bari", nns = 5)

#> Sparse part
#> 5 x 5 sparse Matrix of class "dsCMatrix"
#>               
#> [1,] . . . . .
#> [2,] . . . . .
#> [3,] . . . . .
#> [4,] . . . . .
#> [5,] . . . . .
#> plus left factor
#> 5 x 1 Matrix of class "dgeMatrix"
#>      [,1]
#> [1,]    1
#> [2,]    1
#> [3,]    1
#> [4,]    1
#> [5,]    1
#> times right factor transpose
#> 5 x 1 Matrix of class "dgeMatrix"
#>      [,1]
#> [1,]  0.2
#> [2,]  0.2
#> [3,]  0.2
#> [4,]  0.2
#> [5,]  0.2

init_start(start = "rds", nns = 3)

#> 3 x 3 Matrix of class "dgeMatrix"
#>            [,1]       [,2]       [,3]
#> [1,] 0.54119625 0.04863885 0.41016490
#> [2,] 0.02729525 0.43510871 0.53759604
#> [3,] 0.43150886 0.51625227 0.05223887

init_start(start = "rds_perm_bari", nns = 5)

#> Sparse part
#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>                                                       
#> [1,] .         .         .         0.7933691 .        
#> [2,] .         .         .         .         0.7933691
#> [3,] .         0.7933691 .         .         .        
#> [4,] .         .         0.7933691 .         .        
#> [5,] 0.7933691 .         .         .         .        
#> plus left factor
#> 5 x 1 Matrix of class "dgeMatrix"
#>           [,1]
#> [1,] 0.2066309
#> [2,] 0.2066309
#> [3,] 0.2066309
#> [4,] 0.2066309
#> [5,] 0.2066309
#> times right factor transpose
#> 5 x 1 Matrix of class "dgeMatrix"
#>      [,1]
#> [1,]  0.2
#> [2,]  0.2
#> [3,]  0.2
#> [4,]  0.2
#> [5,]  0.2

# initialize start matrix with soft seeds
init_start(start = "bari", nns = 5, ns = 3, soft_seeds = c(5, 7, 8))

#> Sparse part
#> 5 x 5 sparse Matrix of class "dsCMatrix"
#>               
#> [1,] . . . . .
#> [2,] . 1 . . .
#> [3,] . . . . .
#> [4,] . . . 1 .
#> [5,] . . . . 1
#> plus left factor
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>       
#> [1,] 1
#> [2,] .
#> [3,] 1
#> [4,] .
#> [5,] .
#> times right factor transpose
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>         
#> [1,] 0.5
#> [2,] .  
#> [3,] 0.5
#> [4,] .  
#> [5,] .  

init_start(start = "rds", nns = 5, soft_seeds = ss)

#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>                                        
#> [1,] 0.43166405 0.1849583 . . 0.3833776
#> [2,] 0.09693483 0.4355168 . . 0.4675483
#> [3,] 0.47140113 0.3795249 . . 0.1490740
#> [4,] .          .         1 . .        
#> [5,] .          .         . 1 .        

init_start(start = "rds_perm_bari", nns = 5, soft_seeds = ss)

#> Sparse part
#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>                                       
#> [1,] 0.2971942 .         . . .        
#> [2,] .         0.2971942 . . .        
#> [3,] .         .         . . 0.2971942
#> [4,] .         .         1 . .        
#> [5,] .         .         . 1 .        
#> plus left factor
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>               
#> [1,] 0.7028058
#> [2,] 0.7028058
#> [3,] 0.7028058
#> [4,] .        
#> [5,] .        
#> times right factor transpose
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>               
#> [1,] 0.3333333
#> [2,] 0.3333333
#> [3,] .        
#> [4,] .        
#> [5,] 0.3333333

# initialize start matrix for convex graph matching
cgnp_pair <- sample_correlated_gnp_pair(n = 10, corr =  0.3, p =  0.5)
g1 <- cgnp_pair$graph1
g2 <- cgnp_pair$graph2
seeds <- 1:10 <= 2
init_start(start = "convex", nns = 8, A = g1, B = g2, seeds = seeds)

#> Sparse part
#> 6 x 6 sparse Matrix of class "dgCMatrix"
#>                                                                      
#> [1,] 0.18509628 0.19033837 .         0.06066119 0.12061754 0.12173266
#> [2,] 0.17649921 0.26465315 .         0.06565256 0.17598035 0.07469973
#> [3,] 0.07624274 0.06565256 0.1353609 .          .          0.40383162
#> [4,] 0.06565256 .          0.1898150 0.28174611 0.06066119 .         
#> [5,] 0.19676831 0.08589846 0.1281181 0.09919783 0.17649921 0.06565256
#> [6,] 0.18787512 0.25709358 .         .          0.18981502 0.17649921
#> plus left factor
#> 6 x 1 sparse Matrix of class "dgCMatrix"
#>                
#> [1,] 0.03716605
#> [2,] 0.03716605
#> [3,] 0.03716605
#> [4,] 0.03716605
#> [5,] 0.03716605
#> [6,] 0.03716605
#> times right factor transpose
#> 6 x 1 sparse Matrix of class "dgCMatrix"
#>           
#> [1,] 0.125
#> [2,] 0.125
#> [3,] 0.125
#> [4,] 0.125
#> [5,] 0.125
#> [6,] 0.125

# FW graph matching with incorrect seeds to start at convex start
init_start(start = "convex", nns = 8, ns = 2, soft_seeds = ss, A = g1, B = g2, seeds = seeds)

#> Sparse part
#> 8 x 8 sparse Matrix of class "dgCMatrix"
#>                                                                        
#> [1,] 0.40982767 0.20557868 .          .          .           0.12064897
#> [2,] 0.33059791 0.29692752 0.05547322 0.13662409 0.042912116 .         
#> [3,] .          0.05778114 0.21483643 .          0.007139059 0.35334998
#> [4,] .          .          0.19392873 0.32710025 0.138797959 0.01349164
#> [5,] 0.13324303 0.19809102 .          0.03445091 0.360396782 0.12950904
#> [6,] 0.03599750 0.24162164 .          0.05547322 0.201462599 0.25716878
#> [7,] 0.05547322 .          0.20035301 0.21979983 0.227992865 .         
#> [8,] 0.03486065 .          0.31410999 0.20525309 .           0.10453298
#>                            
#> [1,] 0.112706639 0.12993942
#> [2,] 0.019666643 0.11779850
#> [3,] 0.005253483 0.36163991
#> [4,] 0.305382805 .         
#> [5,] 0.035204673 0.08780593
#> [6,] 0.121062628 0.06591502
#> [7,] 0.275082460 .         
#> [8,] 0.104342051 0.21560261
#> plus left factor
#> 8 x 1 sparse Matrix of class "dgCMatrix"
#>                
#> [1,] 0.02129862
#> [2,] .         
#> [3,] .         
#> [4,] 0.02129862
#> [5,] 0.02129862
#> [6,] 0.02129862
#> [7,] 0.02129862
#> [8,] 0.02129862
#> times right factor transpose
#> 8 x 1 sparse Matrix of class "dgCMatrix"
#>               
#> [1,] .        
#> [2,] .        
#> [3,] 0.1666667
#> [4,] 0.1666667
#> [5,] 0.1666667
#> [6,] 0.1666667
#> [7,] 0.1666667
#> [8,] 0.1666667

#