Matching performance summary

Get a summary of the matching result and measures of the matching performance based on several evaluation metrics associated with nodes and edges of two graphs.

match_report(match, A, B, true_label = NULL, directed = NULL)

edge_match_info(corr, A, B, directed = NULL)

Arguments

match
Graph matching result see graph match methods.

A
A matrix or an igraph object. Adjacency matrix of \(G_1\).

B
A matrix or an igraph object. Adjacency matrix of \(G_2\).

true_label
A vector. NULL if the true correspondence between two graphs is unknown. A vector indicating the true correspondence in the second graph if the true correspondence is known.#'

directed
Whether the graphs should be treated as directed or undirected. NULL defaults to !isSymmetric(A).

corr
Correspondence data.frame as given by match$corr

Value

match_report returns the match object evaluation metrics including number of matches, true matches, and a data frame with edge correctness information. edge_match_info returns this data frame with columns for number of common edges, missing edges, extra edges, and common non-edges, and Frobenius norm.

Details

For multilayered graphs information is given per layer. For weighted graphs the counts are based on non-zero entries. Equality of weights is not tested. If you want to ignore seeds in the edge match info you must remove them from corr/match$corr.

TODO

support weighted? loops? ...?

Examples

graphs <- sample_correlated_gnp_pair(10, .5, .3)
A <- graphs$graph1
B <- graphs$graph2
res <- graph_match_percolation(A, B, 1:4)
match_report(res, A, B)

#> Call: 
#> graph_match_percolation(A = A, B = B, seeds = 1:4)
#> 
#> # Matches: 1
#>                           
#> common_edges      4.000000
#> missing_edges     2.000000
#> extra_edges       2.000000
#> common_non_edges 37.000000
#> fnorm             2.828427
#> 

gp_list <- replicate(3,
   sample_correlated_gnp_pair(100, .8, .3),
   simplify = FALSE)
A <- lapply(gp_list, function(gp)gp[[1]])
B <- lapply(gp_list, function(gp)gp[[2]])
corr <- data.frame(corr_A = 1:100, corr_B = 1:100)
edge_match_info(corr, A, B)

#>   layer common_edges missing_edges extra_edges common_non_edges    fnorm
#> 1     1         1282           193         204             3271 28.17801
#> 2     2         1281           203         221             3245 29.12044
#> 3     3         1293           194         204             3259 28.21347