Lists of Matrices

Basically a list with matrix components that are all the same dimension. Mostly for internal igraphmatch use. It can do various things like arithmetic and indexing, all of which are done component-wise.

matrix_list(ml)

# S4 method for matrix_list,matrix_list
%*%(x, y)

# S4 method for matrix_list
t(x)

# S4 method for matrix_list
dim(x)

# S4 method for matrix_list,ANY,ANY,ANY
[(x, i = 1:nrow(x[[1]]), j = 1:ncol(x[[1]]), drop = FALSE)

# S4 method for matrix_list,ANY,missing,ANY
[(x, i, drop = FALSE)

# S4 method for matrix_list,missing,ANY,ANY
[(x, j, drop = FALSE)

# S4 method for matrix_list,missing,missing,ANY
[(x, drop = FALSE)

# S4 method for matrix_list,ANY
%*%(x, y)

# S4 method for ANY,matrix_list
%*%(x, y)

# S4 method for matrix_list,logical
sum(x, na.rm = FALSE)

# S4 method for matrix_list,ANY
^(e1, e2)

# S4 method for matrix_list,matrix_list
-(e1, e2)

# S4 method for matrix_list,matrix_list
+(e1, e2)

# S4 method for matrix_list,matrix_list
*(e1, e2)

# S4 method for matrix_list,matrix_list
/(e1, e2)

# S4 method for matrix_list,ANY
-(e1, e2)

# S4 method for matrix_list,ANY
+(e1, e2)

# S4 method for matrix_list,ANY
*(e1, e2)

# S4 method for matrix_list,ANY
/(e1, e2)

# S4 method for ANY,matrix_list
-(e1, e2)

# S4 method for ANY,matrix_list
+(e1, e2)

# S4 method for ANY,matrix_list
*(e1, e2)

# S4 method for ANY,matrix_list
/(e1, e2)

# S4 method for matrix_list,missing
-(e1, e2)

# S4 method for matrix_list
names(x) <- value

Arguments

ml

A list of matrices

x

As in Matrix

y

As in Matrix

i

As in Matrix

j

As in Matrix

drop

As in Matrix

na.rm

As in Matrix

e1

As in Matrix

e2

As in Matrix