Measure functions

Measures for computing the goodness of matching for each vertex.

row_cor(g1, g2)

row_diff(g1, g2)

row_perm_stat(g1, g2, exact = TRUE)

Arguments

g1
A matrix or an igraph object. Adjacency matrix of \(G_1\).

g2
A matrix or an igraph object. Adjacency matrix of \(G_2\) after adjusting rows and columns according to the correlation of matching between two graphs.

exact
A logical. If g1 and g2 are binary, then set exact=TRUE, if g1 and g2 are weighted graphs, then set exact=FALSE.

Value

row_cor returns a vector, each element is 1 minus the row correlation value for the corresponding vertex.

row_diff returns a vector, each element is the row difference value for the corresponding vertex.

row_perm_stat returns a vector, each element is the row permutation statistics value for the corresponding vertex.

Examples

cgnp_pair <- sample_correlated_gnp_pair(n = 50, corr =  0.3,p =  0.5)
g1 <- cgnp_pair$graph1
g2 <- cgnp_pair$graph2
match <- graph_match_FW(g1, g2)
g2m <- g2[match$corr$corr_B, match$corr$corr_B]
g1 <- g1[]
row_cor(g1, g2m)

#>  [1] 0.6419280 0.4926108 0.6114222 0.7706458 0.5996796 0.5246525 0.5600000
#>  [8] 0.5681818 0.6114222 0.5651372 0.5516036 0.5467529 0.5516036 0.6410256
#> [15] 0.7221688 0.5809099 0.5585852 0.3879069 0.6568144 0.2995272 0.6484480
#> [22] 0.4900874 0.4563292 0.5917517 0.6545754 0.5723314 0.6887336 0.6114222
#> [29] 0.5529756 0.6944950 0.6008439 0.7211538 0.7190997 0.3219189 0.4783280
#> [36] 0.5612812 0.5681818 0.5542575 0.6322504 0.6353016 0.5585852 0.8636364
#> [43] 0.6251711 0.6114222 0.4800000 0.5570540 0.5246525 0.7389163 0.6545754
#> [50] 0.2848863

row_diff(g1, g2m)

#>  [1] 16 12 15 19 15 13 14 14 15 14 14 14 14 16 18 14 14 11 16  8 17 12 11 14 16
#> [26] 14 17 15 14 18 15 18 18  8 12 14 14 14 16 16 14 22 15 15 12 13 13 18 16  7

row_perm_stat(g1, g2m)

#>  [1] -2.4337141 -3.5493550 -2.7271680 -1.5904150 -2.7153356 -3.2705975
#>  [7] -3.0013940 -3.0130714 -2.7271680 -3.0097238 -2.9964157 -3.0013940
#> [13] -2.9848488 -2.4337141 -1.8300226 -2.9782694 -3.0013940 -3.8747572
#> [19] -2.4443757 -4.7177439 -2.0941446 -3.5848781 -3.8092848 -2.9711668
#> [25] -2.4412646 -2.9711668 -2.1615005 -2.6991002 -2.9964157 -1.7950390
#> [31] -2.6539095 -1.8615975 -1.8656972 -4.6915261 -3.5817370 -2.9964157
#> [37] -2.9848488 -3.0158730 -2.4244024 -2.4071839 -3.0013940 -0.6988596
#> [43] -2.6927266 -2.6927266 -3.5735106 -3.3051261 -3.2705975 -1.8728784
#> [49] -2.4133515 -5.0143949