Sample correlated G(n,p) random graphs

Sample a pair of correlated G(n,p) random graphs with correlation between two graphs being rho and edge probability being p.

sample_correlated_gnp_pair(n, corr, p, permutation = 1:n, ...)

sample_correlated_gnp_pair_w_junk(
  n,
  corr,
  p,
  ncore = n,
  permutation = 1:n,
  ...
)

Arguments

n
An integer. Number of total vertices for the sampled graphs.

corr
A number. The target Pearson correlation between the adjacency matrices of the generated graphs. It must be in open (0,1) interval.

p
A number. Edge probability between two vertices. It must be in open (0,1) interval.

permutation
A numeric vector,permute second graph.

...
Passed to sample_correlated_gnp_pair and sample_correlated_gnp_pair_w_junk.

ncore
An integer. Number of core vertices.

Value

sample_correlated_gnp_pair returns a list of two igraph object, named graph1 and graph2, which are two graphs whose adjacency matrix entries correlated with rho.

sample_correlated_gnp_pair_w_junk returns a list of two igraph object, named graph1 and graph2, which are two graphs whose adjacency matrix entries correlated with rho and with first ncore vertices being core vertices and the rest being junk vertices.

Examples

sample_correlated_gnp_pair(50, 0.3, 0.5)

#> $graph1
#> IGRAPH 0cec6e5 U--- 50 631 -- Erdos renyi (gnp) graph
#> + attr: name (g/c), type (g/c), loops (g/l), p (g/n)
#> + edges from 0cec6e5:
#>  [1]  1-- 2  2-- 3  2-- 4  3-- 4  1-- 5  3-- 5  1-- 6  2-- 6  4-- 7  5-- 7
#> [11]  1-- 8  2-- 8  3-- 8  7-- 8  1-- 9  2-- 9  3-- 9  4-- 9  5-- 9  8-- 9
#> [21]  2--10  4--10  6--10  8--10  1--11  5--11  9--11 10--11  2--12  4--12
#> [31]  5--12  7--12 10--12 11--12  1--13  2--13  3--13  9--13  1--14  2--14
#> [41]  3--14  6--14  8--14 11--14 12--14  1--15  2--15  3--15  4--15  6--15
#> [51]  8--15 11--15 12--15 13--15 14--15  2--16  4--16  5--16  7--16  8--16
#> [61]  1--17  8--17  9--17 13--17 15--17 16--17  2--18  3--18  6--18  7--18
#> [71]  8--18 10--18 17--18  1--19  2--19  3--19  4--19  5--19 10--19 13--19
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH d24b681 U--- 50 622 -- Erdos renyi (gnp) graph
#> + attr: name_1 (g/c), name_2 (g/c), type_1 (g/c), type_2 (g/c), loops_1
#> | (g/l), loops_2 (g/l), p_1 (g/n), p_2 (g/n), name (g/c), type (g/c),
#> | loops (g/l), p (g/n)
#> + edges from d24b681:
#>  [1] 49--50 48--49 47--50 46--50 46--49 46--48 45--49 45--48 45--46 44--50
#> [11] 44--48 44--47 44--46 44--45 43--47 43--46 43--44 42--50 42--49 42--45
#> [21] 42--44 41--50 41--49 41--47 41--46 41--44 41--43 40--49 40--47 40--43
#> [31] 40--42 39--46 39--43 39--40 38--50 38--48 38--46 38--45 38--44 38--43
#> [41] 38--42 38--41 38--40 37--48 37--47 37--42 37--40 37--38 36--47 36--46
#> [51] 36--45 36--44 36--42 36--41 36--37 35--48 35--47 35--45 35--38 35--36
#> + ... omitted several edges
#> 

sample_correlated_gnp_pair_w_junk(50, 0.3, 0.5, 40)

#> $graph1
#> IGRAPH 7f16d2e U--- 50 607 -- 
#> + attr: name_1 (g/c), name_2 (g/c), loops_1 (g/l), loops_2 (g/l), type
#> | (g/c), p (g/n)
#> + edges from 7f16d2e:
#>  [1] 49--50 47--50 47--49 46--49 46--48 45--49 45--46 44--50 44--49 44--48
#> [11] 44--46 43--50 43--45 43--44 42--50 42--45 42--43 41--50 41--47 41--46
#> [21] 41--45 40--48 40--47 40--45 40--43 40--42 40--41 39--50 39--49 39--46
#> [31] 39--45 39--43 39--41 38--50 38--43 37--50 37--49 37--47 37--46 37--44
#> [41] 37--39 36--50 36--47 36--43 36--37 35--50 35--49 35--47 35--44 35--42
#> [51] 35--38 35--36 34--49 34--47 34--45 34--43 34--41 34--39 34--38 33--49
#> [61] 33--41 33--40 32--49 32--45 32--43 32--42 32--39 32--36 32--34 32--33
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH b13fa97 U--- 50 618 -- 
#> + attr: name_1 (g/c), name_2 (g/c), loops_1 (g/l), loops_2 (g/l),
#> | type_1 (g/c), type_2 (g/c), p_1 (g/n), p_2 (g/n), type (g/c), p (g/n)
#> + edges from b13fa97:
#>  [1] 49--50 48--49 47--50 47--48 46--50 46--49 46--48 45--50 44--50 44--48
#> [11] 44--45 43--50 43--48 43--46 43--44 42--50 42--49 42--48 42--45 42--43
#> [21] 41--50 41--49 41--48 41--45 41--44 41--42 40--49 40--48 40--47 40--44
#> [31] 40--43 40--41 39--49 39--48 39--47 39--45 39--44 39--42 39--41 38--49
#> [41] 38--48 38--47 38--46 38--45 38--43 38--41 38--39 37--49 37--48 37--47
#> [51] 37--46 37--44 37--43 37--41 37--39 36--49 36--48 36--47 36--45 36--44
#> [61] 35--45 35--44 35--43 35--42 35--41 35--39 35--38 35--37 35--36 34--50
#> + ... omitted several edges
#>