Sample graphs from edge probability matrix and correlation matrix

Sample a pair of graphs with specified edge probability and correlation between each pair of vertices.

sample_correlated_ieg_pair(
  n,
  p_mat,
  c_mat,
  directed = FALSE,
  loops = FALSE,
  permutation = 1:n
)

sample_correlated_rdpg(X, rho, nc = nrow(X), ...)

Arguments

n
An integer. Number of total vertices for the sampled graphs.

p_mat
An n-by-n matrix. Edge probability matrix, each entry should be in the open (0,1) interval.

c_mat
An n-by-n matrix. The target Pearson correlation matrix, each entry should be in the open (0,1) interval.

directed
Logical scalar, whether to generate directed graphs.

loops
Logical scalar, whether self-loops are allowed in the graph.

permutation
A numeric vector,permute second graph.

X
A matrix. Dot products matrix, each entry must be in open (0,1) interval.

rho
A number. The target Pearson correlation between the adjacency matrices of the generated graphs. It must be in open (0,1) interval.

nc
An integer. Number of core vertices.

...
Passed to sample_correlated_rdpg_pair.

Value

sample_correlated_ieg_pair returns two igraph objects named graph1 and graph2.

sample_correlated_rdpg returns two igraph objects named graph1 and graph2 that are sampled from random dot product graphs model.

Examples

n <- 50
p_mat <- matrix(runif(n^2),n)
c_mat <- matrix(runif(n^2),n)
sample_correlated_ieg_pair(n,p_mat,c_mat)

#> $graph1
#> IGRAPH 776fd9f U--- 50 599 -- 
#> + edges from 776fd9f:
#>   [1] 1-- 2 1-- 3 1-- 4 1-- 5 1-- 9 1--10 1--14 1--19 1--21 1--22 1--26 1--29
#>  [13] 1--30 1--33 1--35 1--41 1--44 1--47 1--49 1--50 2-- 3 2-- 7 2-- 8 2--10
#>  [25] 2--13 2--15 2--18 2--20 2--21 2--22 2--23 2--24 2--25 2--28 2--29 2--36
#>  [37] 2--37 2--40 2--43 2--45 2--46 3-- 4 3-- 9 3--11 3--12 3--13 3--16 3--19
#>  [49] 3--22 3--23 3--25 3--27 3--29 3--31 3--32 3--33 3--34 3--35 3--36 3--39
#>  [61] 3--43 3--44 3--48 4-- 9 4--10 4--12 4--14 4--16 4--18 4--20 4--23 4--24
#>  [73] 4--28 4--30 4--31 4--32 4--34 4--37 4--40 4--41 4--42 4--45 4--46 4--48
#>  [85] 4--49 5-- 9 5--13 5--14 5--18 5--22 5--23 5--24 5--28 5--29 5--30 5--31
#>  [97] 5--34 5--39 5--40 5--43 5--44 5--45 5--46 6--11 6--12 6--13 6--19 6--21
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH 809be3e U--- 50 610 -- 
#> + edges from 809be3e:
#>   [1] 1-- 2 1-- 3 1-- 4 1-- 5 1-- 8 1-- 9 1--10 1--19 1--21 1--22 1--25 1--26
#>  [13] 1--29 1--30 1--32 1--33 1--34 1--35 1--41 1--42 1--44 1--47 1--49 1--50
#>  [25] 2-- 3 2-- 6 2-- 7 2-- 8 2--10 2--13 2--15 2--17 2--18 2--20 2--21 2--23
#>  [37] 2--24 2--25 2--28 2--29 2--36 2--37 2--40 2--41 2--42 2--43 2--44 2--45
#>  [49] 2--46 2--47 3-- 4 3-- 9 3--10 3--12 3--13 3--16 3--19 3--20 3--22 3--23
#>  [61] 3--25 3--27 3--28 3--29 3--31 3--32 3--33 3--34 3--35 3--36 3--39 3--43
#>  [73] 3--44 3--46 3--47 3--48 4-- 6 4--10 4--11 4--12 4--16 4--18 4--23 4--27
#>  [85] 4--28 4--29 4--30 4--31 4--32 4--34 4--37 4--41 4--42 4--45 4--46 5-- 9
#>  [97] 5--10 5--14 5--18 5--22 5--23 5--24 5--28 5--31 5--34 5--37 5--39 5--40
#> + ... omitted several edges
#> 

## sample a pair of igraph objects from random dot
## product graphs model with dimension 3 and scale 8
n <- 50
xdim <- 3
scale <- 8
X <- matrix(rgamma(n*(xdim+1),scale,1),n,xdim+1)
X <- X/rowSums(X)
X <- X[,1:xdim]
sample_correlated_rdpg(X,rho=0.5)

#> $graph1
#> IGRAPH 71a4f1b U--- 50 217 -- 
#> + edges from 71a4f1b:
#>  [1]  1--13  1--15  1--45  1--48  2-- 5  2-- 6  2-- 7  2-- 9  2--11  2--20
#> [11]  2--24  2--25  2--32  2--36  2--42  3-- 4  3-- 6  3-- 7  3--10  3--14
#> [21]  3--16  3--18  3--22  3--23  3--25  3--27  4--12  4--17  4--20  4--35
#> [31]  4--37  4--38  4--43  4--45  4--50  5-- 9  5--13  5--23  5--34  6--29
#> [41]  6--34  7--14  7--17  7--19  7--26  7--29  7--36  7--42  7--46  7--47
#> [51]  7--50  8--13  8--16  8--26  8--30  8--31  8--37  8--41  8--46  9--14
#> [61]  9--28  9--32  9--41  9--47  9--50 10--13 10--24 10--25 10--29 10--30
#> [71] 10--35 10--36 11--13 11--25 11--26 11--28 11--29 11--33 11--45 12--16
#> [81] 12--19 12--27 12--28 12--32 12--33 12--36 12--37 13--14 13--19 13--26
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH 234bd08 U--- 50 213 -- 
#> + edges from 234bd08:
#>  [1]  1-- 3  1--13  1--25  1--27  1--41  2-- 5  2-- 7  2--11  2--18  3-- 4
#> [11]  3-- 6  3-- 7  3--10  3--21  3--22  3--23  3--27  3--41  3--48  3--49
#> [21]  4-- 8  4--10  4--20  4--28  4--34  4--35  4--37  4--42  4--45  4--50
#> [31]  5-- 9  5--13  5--34  6--10  6--11  6--15  6--22  6--29  6--38  7--10
#> [41]  7--14  7--17  7--19  7--28  7--29  7--42  7--44  7--47  8--13  8--16
#> [51]  8--30  8--37  8--41  8--46  9--11  9--14  9--15  9--28  9--34  9--41
#> [61]  9--45  9--47  9--50 10--13 10--24 10--30 10--35 10--42 10--43 10--44
#> [71] 10--46 11--16 11--18 11--25 11--26 11--27 11--29 11--33 11--45 12--16
#> [81] 12--19 12--32 12--33 12--36 12--37 12--38 13--14 13--19 13--25 13--27
#> + ... omitted several edges
#>