Sample graphs pair from stochastic block model

Sample a pair of random graphs from stochastic block model with correlation between two graphs being rho and edge probability being p.

sample_correlated_sbm_pair(n, pref.matrix, block.sizes, rho,
  permutation = 1:n, ...)

sample_correlated_sbm_pair_w_junk(n, pref.matrix, block.sizes, rho,
  core.block.sizes, permutation = 1:n, ...)

Arguments

n
An integer. Number of vertices in the graph.

pref.matrix
The matrix giving the Bernoulli rates. This is a K-by-K matrix, where k is the number of groups. The probability of creating an edge between vertices from groups i and j is given by element i,j. For undirected graphs, this matrix must be symmetric.

block.sizes
A numeric vector. Give the number of vertices in each group. The sum of the vector must match the number of vertices.

rho
A number. The target Pearson correlation between the adjacency matrices of the generated graphs. It must be in open (0,1) interval.

...
Passed to sample_correlated_sbm_pair.

permutation.
A numeric vector, permute second graph.

core.block.size
A numeric vector. Give the number of core vertices in each group. Entries should be smaller than block.sizes and the vector length should be the same as block.sizes.

Value

A list of two igraph object, named graph1 and graph2.

Examples

pm <- cbind( c(.1, .001), c(.001, .05) )
sample_correlated_sbm_pair(1000, pref.matrix=pm, block.sizes=c(300,700), rho=0.5)

#> $graph1
#> IGRAPH 499cb1f U--- 1000 16905 -- Stochastic block-model
#> + attr: name (g/c), loops (g/l)
#> + edges from 499cb1f:
#>  [1]  2-- 3  4-- 7  1-- 8  4-- 9  7-- 9  5--10  3--12 10--13  9--14 11--14
#> [11] 13--14  9--15  5--16  7--16  2--17  7--17 14--17  4--18  5--18  7--18
#> [21] 16--18 13--19 15--19 16--19  1--20  2--20  3--21  5--21 14--21 17--22
#> [31]  5--23  9--23 13--23 16--23  4--24 15--25 21--25 22--25  4--26 17--26
#> [41] 18--26  9--27 16--27 19--27 20--27 14--28 19--28  4--29 26--29  3--30
#> [51]  7--30 27--30  7--31 11--31 21--31 24--31 12--32  3--33  9--33 14--33
#> [61] 17--33 22--33 32--33  2--34  9--34 18--34  3--35  4--35 22--35 30--35
#> [71]  3--36 10--36 32--36 19--37 25--37 34--37  8--38 12--38 24--38  8--39
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH 38301eb U--- 1000 17020 -- Stochastic block-model
#> + attr: name_1 (g/c), name_2 (g/c), loops_1 (g/l), loops_2 (g/l), name
#> | (g/c), loops (g/l)
#> + edges from 38301eb:
#>  [1] 993--1000 993-- 996 992-- 997 989--1000 987--1000 987-- 999 984-- 989
#>  [8] 981-- 991 981-- 982 980-- 991 980-- 983 978-- 994 976-- 985 975-- 981
#> [15] 974-- 983 973-- 996 973-- 993 973-- 985 973-- 978 972-- 994 971-- 975
#> [22] 970-- 995 970-- 981 969-- 975 969-- 973 968-- 996 968-- 990 968-- 980
#> [29] 968-- 978 967-- 997 967-- 992 967-- 984 967-- 982 967-- 971 966-- 999
#> [36] 966-- 990 966-- 967 965--1000 965-- 990 964-- 989 962-- 976 961-- 999
#> [43] 961-- 992 961-- 989 961-- 985 961-- 970 959-- 984 959-- 974 958-- 996
#> + ... omitted several edges
#> 

sample_correlated_sbm_pair_w_junk(1000, pref.matrix=pm, block.sizes=c(300,700), rho=0.5,
core.block.sizes=c(200,500))

#> $graph1
#> IGRAPH fff0fa0 U--- 1000 16667 -- Stochastic block-model
#> + attr: name (g/c), loops (g/l)
#> + edges from fff0fa0:
#>  [1]  1-- 9  1--11  2--11  9--13  7--14 12--15  4--17 10--17 12--17 11--18
#> [11]  8--20 11--21 13--21 14--21 19--21 20--22 12--24 15--24 18--24  7--25
#> [21] 10--25 11--25 12--25 22--25  8--26 16--26 19--26 22--26 24--26 25--26
#> [31]  7--27 12--27 20--28 21--28 11--29  2--30 25--30 13--31 14--31  6--32
#> [41] 10--32 13--32  7--33 10--33 16--34 18--34 21--34  1--35  8--35 12--35
#> [51] 34--35  5--36  6--36  9--36 15--36 25--36  2--37  3--37  8--37 24--37
#> [61]  6--38 11--38 17--38 33--38 34--38 20--39 37--39  5--40 30--40  1--41
#> [71]  9--41 25--41 31--41 32--41 40--41  8--42 17--42 22--42 25--42 31--42
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH 9e508a4 U--- 1000 16763 -- Stochastic block-model
#> + attr: name_1 (g/c), name_2 (g/c), loops_1 (g/l), loops_2 (g/l), name
#> | (g/c), loops (g/l)
#> + edges from 9e508a4:
#>  [1] 994--995 990--995 990--993 990--992 989--993 988--991 987--989 986--998
#>  [9] 986--988 985--991 984--998 983--988 982--983 981--989 981--985 980--986
#> [17] 980--981 978--994 978--985 977--992 975--997 975--991 975--983 974--995
#> [25] 974--993 974--976 973--997 973--978 972--992 971--980 970--987 970--977
#> [33] 969--983 969--978 969--972 965--992 965--989 965--983 965--970 964--991
#> [41] 964--974 963--999 963--990 962--995 962--991 961--998 961--968 961--965
#> [49] 960--985 960--967 959--962 958--999 958--961 957--996 957--988 957--987
#> + ... omitted several edges
#>