Start matrix initialization

initialize the start matrix for graph matching iteration.

bari_start(nns, ns = 0, soft_seeds = NULL)

rds_sinkhorn_start(nns, ns = 0, soft_seeds = NULL, distribution = "runif")

rds_perm_bari_start(nns, ns = 0, soft_seeds = NULL, g = 1, is_splr = TRUE)

rds_from_sim_start(nns, ns = 0, soft_seeds = NULL, sim)

Arguments

nns
An integer. Number of non-seeds.

ns
An integer. Number of hard seeds.

soft_seeds
A vector, a matrix or a data frame. If there is no error in soft seeds, input can be a vector of soft seed indices in \(G_1\). Or if there is error in soft seeds, input in the form of a matrix or a data frame, with the first column being the indices of \(G_1\) and the second column being the corresponding indices of \(G_2\). Note that if there are seeds in graphs, seeds should be put before non-seeds.

distribution
A charactor. Specify the distribution from which the random doubly stochastic matrix is sampled. Should input the name of the function for generating random deviates from that distribution.

g
A number. Specified in the range of [0, 1] to set weights to random permutaion matrix and barycenter matrix.

is_splr
should we return a splr matrix? (default = TRUE)

sim
nns x nns non-negative matrix.

Value

bari_start returns a nns-by-nns matrix with 1's corresponding to the adaptive seeds and being bari-centered at other places.

rds_sinkhorn_start returns a nns-by-nns doubly stochastic matrix with 1's corresponding to adaptive seeds.

rds_perm_bari returns a nns-by-nns doubly stochastic matrix with 1's corresponding to adaptive seeds.

rds_from_sim_start returns a doubly stochastic Matrix given by sinkhorn algorithm applied to a matrix of iid log-normal with mu=sim. Note, this ignores soft seeds.

Examples

## Case without soft seeds
bari_start(3)

#> Sparse part
#> 3 x 3 sparse Matrix of class "dsCMatrix"
#>           
#> [1,] . . .
#> [2,] . . .
#> [3,] . . .
#> plus left factor
#> 3 x 1 Matrix of class "dgeMatrix"
#>      [,1]
#> [1,]    1
#> [2,]    1
#> [3,]    1
#> times right factor transpose
#> 3 x 1 Matrix of class "dgeMatrix"
#>           [,1]
#> [1,] 0.3333333
#> [2,] 0.3333333
#> [3,] 0.3333333

## Case with correct soft seeds and input is a vector
bari_start(nns=5, ns=3, soft_seeds=c(5, 7, 8))

#> Sparse part
#> 5 x 5 sparse Matrix of class "dsCMatrix"
#>               
#> [1,] . . . . .
#> [2,] . 1 . . .
#> [3,] . . . . .
#> [4,] . . . 1 .
#> [5,] . . . . 1
#> plus left factor
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>       
#> [1,] 1
#> [2,] .
#> [3,] 1
#> [4,] .
#> [5,] .
#> times right factor transpose
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>         
#> [1,] 0.5
#> [2,] .  
#> [3,] 0.5
#> [4,] .  
#> [5,] .  

## Case with erroneous soft seeds and the input is a matrix
bari_start(nns=5, soft_seeds=matrix(c(2, 4, 2, 3), nrow=2))

#> Sparse part
#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>               
#> [1,] . . . . .
#> [2,] . 1 . . .
#> [3,] . . . . .
#> [4,] . . 1 . .
#> [5,] . . . . .
#> plus left factor
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>       
#> [1,] 1
#> [2,] .
#> [3,] 1
#> [4,] .
#> [5,] 1
#> times right factor transpose
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>               
#> [1,] 0.3333333
#> [2,] .        
#> [3,] .        
#> [4,] 0.3333333
#> [5,] 0.3333333

## Case without soft seeds
rds_sinkhorn_start(5)

#> 5 x 5 Matrix of class "dgeMatrix"
#>            [,1]       [,2]      [,3]       [,4]       [,5]
#> [1,] 0.01569865 0.47033569 0.1407185 0.15286939 0.22037779
#> [2,] 0.42719808 0.22166156 0.2305148 0.08699909 0.03362651
#> [3,] 0.35623021 0.10291727 0.1651684 0.28589779 0.08978635
#> [4,] 0.09807594 0.03951686 0.2994644 0.23625201 0.32669083
#> [5,] 0.10279713 0.16556861 0.1641340 0.23798172 0.32951852

## Case with soft seeds and the input is a data frame
rds_sinkhorn_start(nns=5,
   soft_seeds = as.data.frame(matrix(c(2, 4, 2, 3), nrow=2)),
   distribution = "rnorm")

#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>                                         
#> [1,] 0.09279355 . . 0.57120528 0.3360012
#> [2,] .          1 . .          .        
#> [3,] 0.64524340 . . 0.04506174 0.3096949
#> [4,] .          . 1 .          .        
#> [5,] 0.26196305 . . 0.38373299 0.3543040

## Case without soft seeds
rds_perm_bari_start(nns=5)

#> Sparse part
#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>                                                       
#> [1,] .         .         .         0.1759884 .        
#> [2,] .         .         .         .         0.1759884
#> [3,] 0.1759884 .         .         .         .        
#> [4,] .         0.1759884 .         .         .        
#> [5,] .         .         0.1759884 .         .        
#> plus left factor
#> 5 x 1 Matrix of class "dgeMatrix"
#>           [,1]
#> [1,] 0.8240116
#> [2,] 0.8240116
#> [3,] 0.8240116
#> [4,] 0.8240116
#> [5,] 0.8240116
#> times right factor transpose
#> 5 x 1 Matrix of class "dgeMatrix"
#>      [,1]
#> [1,]  0.2
#> [2,]  0.2
#> [3,]  0.2
#> [4,]  0.2
#> [5,]  0.2

## Case with soft seeds and the input is a data frame
rds_perm_bari_start(nns=5, ns=0, soft_seeds=as.data.frame(matrix(c(2, 4, 2, 3), nrow=2)))

#> Sparse part
#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>                                       
#> [1,] 0.4146009 . . .         .        
#> [2,] .         1 . .         .        
#> [3,] .         . . 0.4146009 .        
#> [4,] .         . 1 .         .        
#> [5,] .         . . .         0.4146009
#> plus left factor
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>               
#> [1,] 0.5853991
#> [2,] .        
#> [3,] 0.5853991
#> [4,] .        
#> [5,] 0.5853991
#> times right factor transpose
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>               
#> [1,] 0.3333333
#> [2,] .        
#> [3,] .        
#> [4,] 0.3333333
#> [5,] 0.3333333

sim <- Matrix::rsparsematrix(10, 10, .4,
 rand.x = function(n) rep(1,n))
start_sparse <- rds_from_sim_start(10, sim = sim)
start_dense <- rds_from_sim_start(10, sim = as.matrix(sim))