Britain Transportation network

The network reflects the transportation connections in the UK, with five layers representing ferry, rail, metro, coach, and bus. The original graph has 262385 nodes and 461555 connections in five layers in total. A smaller graph was constructed based on a random walk starting from a randomly chosen hub node, a node that has connections in all the layers. The template graph has 53 nodes and 56 connections in total and is an induced subgraph of the original graph. In addition, the dataset also contains a list of candidate matches for each template node, where the true correspondence is guaranteed to among the candidates. The world graph, the template graph and the candidate matches information are stored in a list where two graphs are stored in two lists of igraph objects representing different layers, and the candidate information is stored in a data frame where the ids of the template nodes and the world nodes are in columns named "tem" and "wor".

data(Transportation)

Format

An object of class list of length 3.

References

Gallotti Riccardo and Barthelemy Marc. (2015), The Multilayer Temporal Network of Public Transport in Great Britain. Scientific Data.

Examples

data(Transportation)
template <- Transportation[[1]]
world <- Transportation[[2]]
candidate <- Transportation[[3]]