Check the similarity matrix passed to a matching function

Internal function that checks that a similarity matrix satisfies necessary conditions and modifies it for use in graph matching.

check_sim(sim, seeds, nonseeds, totv1, totv2, for_nonseeds = TRUE)

Arguments

sim

Similarity matrix

seeds

dataframe of seed matches from running check_seeds

nonseeds

dataframe of nonseed nodes from running check_seeds

totv1

total number of vertices in the first graph

totv2

total number of vertices in the second graph

for_nonseeds

Whether the similarities are between non-seed nodes only (default = TRUE), or if similarities among seed nodes are included (FALSE)

Value

Standardized similarity matrix for similarities only between nonseeds across the two graphs, if for_nonseeds = TRUE, or between all nodes, if for_nonseeds = FALSE

Details

The goal here is to be flexible in terms of the dimensions of the similarity matrix passed to gm. This is useful when the graphs have different orders in which case the function accepts matrices with dimensions equal to that of orders of the original graphs or the number of nonseeds.