Frank-Wolfe Graph Matching Methods

Match two given graphs, returns a list of graph matching results, including matching correspondence vector of \(G_2\) with respect to \(G_1\), doubly stochastic matrix and permutation matrix.

graph_match_convex(
  A,
  B,
  seeds = NULL,
  start = "bari",
  max_iter = 100,
  similarity = NULL,
  tol = 1e-05,
  lap_method = NULL
)

graph_match_FW(
  A,
  B,
  seeds = NULL,
  start = "bari",
  max_iter = 20,
  similarity = NULL,
  lap_method = NULL
)

gm_indefinite(
  A,
  B,
  seeds = NULL,
  start = "bari",
  max_iter = 20,
  similarity = NULL,
  lap_method = NULL
)

graph_match_PATH(
  A,
  B,
  similarity = NULL,
  seeds = NULL,
  alpha = 0.5,
  epsilon = 1
)

Arguments

A
A matrix, igraph object, or list of either.

B
A matrix, igraph object, or list of either.

seeds
A vector of integers or logicals, a matrix or a data frame. If the seed pairs have the same indices in both graphs then seeds can be a vector. If not, seeds must be a matrix or a data frame, with the first column being the indices of \(G_1\) and the second column being the corresponding indices of \(G_2\).

start
A matrix or a character. Any nns-by-nns matrix or character value like "bari" or "convex" to initialize the starting matrix.

max_iter
A number. Maximum number of replacing matches equals to max_iter times number of total vertices of \(G_1\).

similarity
A matrix. An n-by-n matrix containing vertex similaities.

tol
A number. Tolerance of edge disagreements.

lap_method
Choice for lap method.

alpha
A number betwen 0 and 1. Bigger alpha means putting more importance on the information in network topology over other information such as similarity scores

epsilon
A small number

Value

graph_match_convex returns a list of graph matching results, including the graph matching formula, a data frame containing the matching correspondence between \(G_1\) and \(G_2\) named corr_A and corr_B and the number of seeds.

graph_match_FW returns a list of graph matching results, including the graph matching formula, a data frame containing the matching correspondence between \(G_1\) and \(G_2\) named corr_A and corr_B and the number of seeds.

graph_match_PATH returns a list of graph matching results, including the graph matching formula, a data frame containing the matching correspondence between \(G_1\) and \(G_2\) named corr_A and corr_B, the number of seeds if any, the permutation matrix and the doubly stochastic matrix before projection onto the permutation set.

References

M. Zaslavskiy, F. Bach and J. Vert (2009), A Path following algorithm for the graph matching problem. IEEE Trans Pattern Anal Mach Intell, pages 2227-2242.

Examples

cgnp_pair <- sample_correlated_gnp_pair(n = 10, corr =  0.3, p =  0.5)
g1 <- cgnp_pair$graph1
g2 <- cgnp_pair$graph2
# match G_1 & G_2 with no seeds
graph_match_FW(g1, g2)

#> $call
#> graph_match_FW(A = g1, B = g2)
#> 
#> $corr
#>    corr_A corr_B
#> 1       1      5
#> 2       2      1
#> 3       3      9
#> 4       4      2
#> 5       5     10
#> 6       6      4
#> 7       7      7
#> 8       8      3
#> 9       9      8
#> 10     10      6
#> 
#> $ns
#> [1] 0
#> 
#> $P
#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                          
#>  [1,] . . . . 1 . . . . .
#>  [2,] 1 . . . . . . . . .
#>  [3,] . . . . . . . . 1 .
#>  [4,] . 1 . . . . . . . .
#>  [5,] . . . . . . . . . 1
#>  [6,] . . . 1 . . . . . .
#>  [7,] . . . . . . 1 . . .
#>  [8,] . . 1 . . . . . . .
#>  [9,] . . . . . . . 1 . .
#> [10,] . . . . . 1 . . . .
#> 
#> $D
#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                          
#>  [1,] . . . . 1 . . . . .
#>  [2,] 1 . . . . . . . . .
#>  [3,] . . . . . . . . 1 .
#>  [4,] . 1 . . . . . . . .
#>  [5,] . . . . . . . . . 1
#>  [6,] . . . 1 . . . . . .
#>  [7,] . . . . . . 1 . . .
#>  [8,] . . 1 . . . . . . .
#>  [9,] . . . . . . . 1 . .
#> [10,] . . . . . 1 . . . .
#> 
#> $num_iter
#> [1] 16
#> 

seeds <- 1:10 <= 3
graph_match_convex(g1, g2, seeds)

#> $call
#> graph_match_convex(A = g1, B = g2, seeds = seeds)
#> 
#> $corr
#>    corr_A corr_B
#> 1       1      1
#> 2       2      2
#> 3       3      3
#> 4       4      4
#> 5       5      5
#> 6       6      8
#> 7       7      7
#> 8       8      6
#> 9       9      9
#> 10     10     10
#> 
#> $ns
#> [1] 3
#> 
#> $P
#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                          
#>  [1,] 1 . . . . . . . . .
#>  [2,] . 1 . . . . . . . .
#>  [3,] . . 1 . . . . . . .
#>  [4,] . . . 1 . . . . . .
#>  [5,] . . . . 1 . . . . .
#>  [6,] . . . . . . . 1 . .
#>  [7,] . . . . . . 1 . . .
#>  [8,] . . . . . 1 . . . .
#>  [9,] . . . . . . . . 1 .
#> [10,] . . . . . . . . . 1
#> 
#> $D
#> Sparse part
#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                                                                              
#>  [1,] 1 . . .          .          .          .          .          .         
#>  [2,] . 1 . .          .          .          .          .          .         
#>  [3,] . . 1 .          .          .          .          .          .         
#>  [4,] . . . 0.26915096 0.28943342 0.08815433 0.09282202 .          0.05550895
#>  [5,] . . . 0.18802438 0.24427048 .          0.06819994 0.21983929 0.08815433
#>  [6,] . . . .          0.06819994 .          0.34228594 0.46513015 .         
#>  [7,] . . . 0.06819994 .          0.14594092 0.31679917 0.08815433 0.14971718
#>  [8,] . . . 0.19495882 0.21816952 0.20938119 0.08815433 0.14034399 0.11276252
#>  [9,] . . . 0.06245991 0.08815433 0.38963407 0.05550895 0.05030260 0.31771049
#> [10,] . . . 0.18097635 0.05554267 0.13065986 .          .          0.23991689
#>                 
#>  [1,] .         
#>  [2,] .         
#>  [3,] .         
#>  [4,] 0.16870068
#>  [5,] 0.15528193
#>  [6,] 0.08815433
#>  [7,] 0.19495882
#>  [8,] .         
#>  [9,] .         
#> [10,] 0.35667460
#> plus left factor
#> 10 x 1 sparse Matrix of class "dgCMatrix"
#>                 
#>  [1,] .         
#>  [2,] .         
#>  [3,] .         
#>  [4,] 0.03622964
#>  [5,] 0.03622964
#>  [6,] 0.03622964
#>  [7,] 0.03622964
#>  [8,] 0.03622964
#>  [9,] 0.03622964
#> [10,] 0.03622964
#> times right factor transpose
#> 10 x 1 sparse Matrix of class "dgCMatrix"
#>                
#>  [1,] .        
#>  [2,] .        
#>  [3,] .        
#>  [4,] 0.1428571
#>  [5,] 0.1428571
#>  [6,] 0.1428571
#>  [7,] 0.1428571
#>  [8,] 0.1428571
#>  [9,] 0.1428571
#> [10,] 0.1428571
#> 
#> $num_iter
#> [1] 14
#> 

hard_seeds <- matrix(c(4,6,5,4),2)
seeds <- rbind(as.matrix(check_seeds(seeds, 10)$seeds),hard_seeds)
if (FALSE) {
# for some reason this fails in check
graph_match_convex(g1, g2, seeds)
}

# match G_1 & G_2 with some known node pairs as seeds
seeds <- 1:10 <= 3
graph_match_FW(g1, g2, seeds, start = "bari")

#> $call
#> graph_match_FW(A = g1, B = g2, seeds = seeds, start = "bari")
#> 
#> $corr
#>    corr_A corr_B
#> 1       1      1
#> 2       2      2
#> 3       3      3
#> 4       4      4
#> 5       5      5
#> 6       6      8
#> 7       7      7
#> 8       8      9
#> 9       9      6
#> 10     10     10
#> 
#> $ns
#> [1] 3
#> 
#> $P
#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                          
#>  [1,] 1 . . . . . . . . .
#>  [2,] . 1 . . . . . . . .
#>  [3,] . . 1 . . . . . . .
#>  [4,] . . . 1 . . . . . .
#>  [5,] . . . . 1 . . . . .
#>  [6,] . . . . . . . 1 . .
#>  [7,] . . . . . . 1 . . .
#>  [8,] . . . . . . . . 1 .
#>  [9,] . . . . . 1 . . . .
#> [10,] . . . . . . . . . 1
#> 
#> $D
#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                          
#>  [1,] 1 . . . . . . . . .
#>  [2,] . 1 . . . . . . . .
#>  [3,] . . 1 . . . . . . .
#>  [4,] . . . 1 . . . . . .
#>  [5,] . . . . 1 . . . . .
#>  [6,] . . . . . . . 1 . .
#>  [7,] . . . . . . 1 . . .
#>  [8,] . . . . . . . . 1 .
#>  [9,] . . . . . 1 . . . .
#> [10,] . . . . . . . . . 1
#> 
#> $num_iter
#> [1] 3
#> 

# match G_1 & G_2 with some incorrect seeds
hard_seeds <- matrix(c(4,6,5,4),2)
seeds <- rbind(as.matrix(check_seeds(seeds, nv = 10)$seeds),hard_seeds)
graph_match_FW(g1, g2, seeds, start = "convex")

#> $call
#> graph_match_FW(A = g1, B = g2, seeds = seeds, start = "convex")
#> 
#> $corr
#>    corr_A corr_B
#> 1       1      1
#> 2       2      2
#> 3       3      3
#> 4       4      5
#> 5       5      8
#> 6       6      4
#> 7       7      7
#> 8       8      9
#> 9       9     10
#> 10     10      6
#> 
#> $ns
#> [1] 5
#> 
#> $P
#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                          
#>  [1,] 1 . . . . . . . . .
#>  [2,] . 1 . . . . . . . .
#>  [3,] . . 1 . . . . . . .
#>  [4,] . . . . 1 . . . . .
#>  [5,] . . . . . . . 1 . .
#>  [6,] . . . 1 . . . . . .
#>  [7,] . . . . . . 1 . . .
#>  [8,] . . . . . . . . 1 .
#>  [9,] . . . . . . . . . 1
#> [10,] . . . . . 1 . . . .
#> 
#> $D
#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                          
#>  [1,] 1 . . . . . . . . .
#>  [2,] . 1 . . . . . . . .
#>  [3,] . . 1 . . . . . . .
#>  [4,] . . . . 1 . . . . .
#>  [5,] . . . . . . . 1 . .
#>  [6,] . . . 1 . . . . . .
#>  [7,] . . . . . . 1 . . .
#>  [8,] . . . . . . . . 1 .
#>  [9,] . . . . . . . . . 1
#> [10,] . . . . . 1 . . . .
#> 
#> $num_iter
#> [1] 4
#> 

 gp_list <- replicate(3, sample_correlated_gnp_pair(20, .3, .5), simplify = FALSE)
 A <- lapply(gp_list, function(gp)gp[[1]])
 B <- lapply(gp_list, function(gp)gp[[2]])
 match <- graph_match_FW(A, B, seeds = 1:10, start = "bari", max_iter = 20)
 match$corr

#>    corr_A corr_B
#> 1       1      1
#> 2       2      2
#> 3       3      3
#> 4       4      4
#> 5       5      5
#> 6       6      6
#> 7       7      7
#> 8       8      8
#> 9       9      9
#> 10     10     10
#> 11     11     11
#> 12     12     14
#> 13     13     13
#> 14     14     19
#> 15     15     18
#> 16     16     16
#> 17     17     17
#> 18     18     15
#> 19     19     12
#> 20     20     20

# match G_1 & G_2 using PATH algorithm
graph_match_PATH(g1, g2)

#> $call
#> graph_match_PATH(A = g1, B = g2)
#> 
#> $corr
#>    corr_A corr_B
#> 1       1      1
#> 2       2      2
#> 3       3      3
#> 4       4      4
#> 5       5      5
#> 6       6      6
#> 7       7      7
#> 8       8      8
#> 9       9      9
#> 10     10     10
#> 
#> $ns
#> [1] 0
#> 
#> $P
#> 10 x 10 diagonal matrix of class "ddiMatrix"
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    1    .    .    .    .    .    .    .    .     .
#>  [2,]    .    1    .    .    .    .    .    .    .     .
#>  [3,]    .    .    1    .    .    .    .    .    .     .
#>  [4,]    .    .    .    1    .    .    .    .    .     .
#>  [5,]    .    .    .    .    1    .    .    .    .     .
#>  [6,]    .    .    .    .    .    1    .    .    .     .
#>  [7,]    .    .    .    .    .    .    1    .    .     .
#>  [8,]    .    .    .    .    .    .    .    1    .     .
#>  [9,]    .    .    .    .    .    .    .    .    1     .
#> [10,]    .    .    .    .    .    .    .    .    .     1
#> 
#> $D
#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                                                                              
#>  [1,] 0.754977 .        .        .        .        .        0.245023 .       
#>  [2,] .        0.754977 .        .        .        .        .        .       
#>  [3,] .        .        0.754977 0.245023 .        .        .        .       
#>  [4,] .        .        .        0.754977 0.245023 .        .        .       
#>  [5,] .        .        .        .        0.754977 0.245023 .        .       
#>  [6,] .        .        .        .        .        0.754977 .        0.245023
#>  [7,] 0.245023 .        .        .        .        .        0.754977 .       
#>  [8,] .        .        0.245023 .        .        .        .        0.754977
#>  [9,] .        .        .        .        .        .        .        .       
#> [10,] .        0.245023 .        .        .        .        .        .       
#>                        
#>  [1,] .        .       
#>  [2,] 0.245023 .       
#>  [3,] .        .       
#>  [4,] .        .       
#>  [5,] .        .       
#>  [6,] .        .       
#>  [7,] .        .       
#>  [8,] .        .       
#>  [9,] 0.754977 0.245023
#> [10,] .        0.754977
#> 
#> $iter
#> [1] 2
#> 
#> $lambda
#> [1] 1
#>