Multiple Graph Match Methods

Match two lists of graphs, returns a list of graph matching results, including matching correspondence vector of \(G_2\) with respect to \(G_1\), doubly stochastic matrix and permutation matrix.

graph_match_FW_multi(
  A,
  B,
  seeds = NULL,
  start = "bari",
  max_iter = 20,
  similarity = NULL,
  lap_method = NULL
)

Arguments

A
A list of graphs (or igraph objects)

B
A list of graphs (or igraph objects)

seeds
A vector of integers or logicals, a matrix or a data frame. If there is no error in seeds input can be a vector of seed indices in \(G_1\). Or if there exists error in seeds, input in the form of a matrix or a data frame, with the first column being the indices of \(G_1\) and the second column being the corresponding indices of \(G_2\).

start
A matrix or a character. Any nns-by-nns matrix or character value like "bari" or "convex" to initialize the starting matrix.

max_iter
An integer. Maximum iteration time.

similarity
matrix of similarity scores

lap_method
Which lap function to use

Value

graph_match_FW_multi returns a list of graph matching results, including match correspondence vector of \(G_2\) with respect to \(G_1\) named corr, doubly stochastic matrix named D, permutation matrix named P based on Frank-Wolfe methodology and iteration time of the algorithm named iter.

Examples

if (FALSE) {
 gp_list <- replicate(3, sample_correlated_gnp_pair(100, .3, .5), simplify = FALSE)
 A <- lapply(gp_list, function(gp)gp[[1]])
 B <- lapply(gp_list, function(gp)gp[[2]])
 match <- graph_match_FW_multi(A, B, seeds = 1:10, start = "bari", max_iter = 20)
 match$corr
}