Graph Match Methods

Match two given graphs, returns a list of graph matching results, including matching correspondence vector of \(G_2\) with respect to \(G_1\), doubly stochastic matrix and permutation matrix.

graph_match_FW(
  A,
  B,
  seeds = NULL,
  start = "convex",
  max_iter = 20,
  similarity = NULL,
  lap_method = NULL
)

gm_indefinite(
  A,
  B,
  seeds = NULL,
  start = "convex",
  max_iter = 20,
  similarity = NULL,
  lap_method = NULL
)

graph_match_convex(
  A,
  B,
  seeds = NULL,
  start = "bari",
  max_iter = 100,
  similarity = NULL,
  tol = 1e-05,
  lap_method = NULL
)

graph_match_PATH(
  A,
  B,
  similarity = NULL,
  seeds = NULL,
  alpha = 0.5,
  epsilon = 1
)

graph_match_percolation(A, B, seeds, r = 2)

graph_match_ExpandWhenStuck(A, B, seeds, r = 2)

graph_match_soft_percolation(A, B, seeds, r = 2, max_iter = 100)

graph_match_IsoRank(
  A,
  B,
  similarity,
  seeds = NULL,
  alpha = 0.5,
  max_iter = 50,
  method = "greedy"
)

graph_match_Umeyama(A, B, similarity = NULL, seeds = NULL, alpha = 0.5)

Arguments

A
A matrix or an igraph object. Adjacency matrix of \(G_1\).

B
A matrix or an igraph object. Adjacency matrix of \(G_2\).

seeds
A vector of integers or logicals, a matrix or a data frame. If there is no error in seeds input can be a vector of seed indices in \(G_1\). Or if there exists error in seeds, input in the form of a matrix or a data frame, with the first column being the indices of \(G_1\) and the second column being the corresponding indices of \(G_2\).

start
A matrix or a character. Any nns-by-nns matrix or character value like "bari" or "convex" to initialize the starting matrix.

max_iter
A number. Maximum number of replacing matches equals to max_iter times number of total vertices of \(G_1\).

similarity
A matrix. An n-by-n matrix containing vertex similaities.

lap_method
Choice for lap method.

tol
A number. Tolerance of edge disagreements.

alpha
A number betwen 0 and 1. Bigger alpha means putting more importance on the information in network topology over other information such as similarity scores

epsilon
A small number

r
A number. Threshold of neighboring pair scores.

method
A character. Choice of method to extract mapping from score matrix, including greedy method and the Hungarian algorithm.

Value

graph_match_FW returns a list of graph matching results, including the graph matching formula, a data frame containing the matching correspondence between \(G_1\) and \(G_2\) named corr_A and corr_B and the number of seeds.

graph_match_convex returns a list of graph matching results, including the graph matching formula, a data frame containing the matching correspondence between \(G_1\) and \(G_2\) named corr_A and corr_B and the number of seeds.

graph_match_PATH returns a list of graph matching results, including the graph matching formula, a data frame containing the matching correspondence between \(G_1\) and \(G_2\) named corr_A and corr_B, the number of seeds if any, the permutation matrix and the doubly stochastic matrix before projection onto the permutation set.

graph_match_percolation returns a list of graph matching results, including the graph matching formula, a data frame containing the matching correspondence between \(G_1\) and \(G_2\) named corr_A and corr_B, the number of seeds and the order of nodes getting matched.

graph_match_ExpandWhenStuck returns a list of graph matching results, including the graph matching formula, a data frame containing the matching correspondence between \(G_1\) and \(G_2\) named corr_A and corr_B, the number of seeds and the order of nodes getting matched.

graph_match_soft_percolation returns a list of graph matching results, including the graph matching formula, a data frame containing the matching correspondence between \(G_1\) and \(G_2\) named corr_A and corr_B, the number of seeds and the order of nodes getting matched.

graph_match_IsoRank returns a list of graph matching results, including the graph matching formula, a data frame containing the matching correspondence between \(G_1\) and \(G_2\) named corr_A and corr_B and the number of seeds. If choose the greedy method to extract mapping, the order of nodes getting matched will also be returned.

graph_match_Umeyama returns a list of graph matching results, including the graph matching formula, a data frame containing the matching correspondence between \(G_1\) and \(G_2\) named corr_A and corr_B and the number of seeds.

References

M. Zaslavskiy, F. Bach and J. Vert (2009), A Path following algorithm for the graph matching problem. IEEE Trans Pattern Anal Mach Intell, pages 2227-2242.

L. Yartseva and M. Grossglauser (2013), On the performance of percolation graph matching. COSN, Boston, MA, USA, pages 119–130.

E. Kazemi, S. H. Hassani, and M. Grossglauser (2015), Growing a graph matching from a handful of seeds. Proc. of the VLDB Endowment, 8(10):1010–1021.

R. Singh, J. Xu, B. Berger (2008), Global alignment of multiple protein interaction networks with application to functional orthology detection. Proc Natl Acad Sci. USA, pages 12763-12768.

S. Umeyama (1988), An eigendecomposition approach to weighted graph matching problems. IEEE TPAMI. USA, pages 695-703.

Examples

cgnp_pair <- sample_correlated_gnp_pair(n = 10, corr =  0.3, p =  0.5)
g1 <- cgnp_pair$graph1
g2 <- cgnp_pair$graph2
# match G_1 & G_2 with no seeds
graph_match_FW(g1, g2)

#> Error in .local(x, i, j, ..., value): RHS 'value' (class dsparseVector) matches 'ANY', but must match matrix class splrMatrix

# match G_1 & G_2 with some known node pairs as seeds
seeds <- 1:10 <= 3
graph_match_FW(g1, g2, seeds, start = "bari")

#> $call
#> graph_match_FW(A = g1, B = g2, seeds = seeds, start = "bari")
#> 
#> $corr
#>    corr_A corr_B
#> 1       1      1
#> 2       2      2
#> 3       3      3
#> 4       4      6
#> 5       5      5
#> 6       6      7
#> 7       7     10
#> 8       8      9
#> 9       9      8
#> 10     10      4
#> 
#> $ns
#> [1] 3
#> 
#> $P
#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                          
#>  [1,] 1 . . . . . . . . .
#>  [2,] . 1 . . . . . . . .
#>  [3,] . . 1 . . . . . . .
#>  [4,] . . . . . 1 . . . .
#>  [5,] . . . . 1 . . . . .
#>  [6,] . . . . . . 1 . . .
#>  [7,] . . . . . . . . . 1
#>  [8,] . . . . . . . . 1 .
#>  [9,] . . . . . . . 1 . .
#> [10,] . . . 1 . . . . . .
#> 
#> $D
#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                          
#>  [1,] 1 . . . . . . . . .
#>  [2,] . 1 . . . . . . . .
#>  [3,] . . 1 . . . . . . .
#>  [4,] . . . . . 1 . . . .
#>  [5,] . . . . 1 . . . . .
#>  [6,] . . . . . . 1 . . .
#>  [7,] . . . . . . . . . 1
#>  [8,] . . . . . . . . 1 .
#>  [9,] . . . . . . . 1 . .
#> [10,] . . . 1 . . . . . .
#> 
#> $num_iter
#> [1] 4
#> 

# match G_1 & G_2 with some incorrect seeds
hard_seeds <- matrix(c(4,6,5,4),2)
seeds <- rbind(as.matrix(check_seeds(seeds, nv = 10)$seeds),hard_seeds)
graph_match_FW(g1, g2, seeds, start = "convex")

#> $call
#> graph_match_FW(A = g1, B = g2, seeds = seeds, start = "convex")
#> 
#> $corr
#>    corr_A corr_B
#> 1       1      1
#> 2       2      2
#> 3       3      3
#> 4       4      5
#> 5       5      6
#> 6       6      4
#> 7       7      8
#> 8       8      7
#> 9       9      9
#> 10     10     10
#> 
#> $ns
#> [1] 5
#> 
#> $P
#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                          
#>  [1,] 1 . . . . . . . . .
#>  [2,] . 1 . . . . . . . .
#>  [3,] . . 1 . . . . . . .
#>  [4,] . . . . 1 . . . . .
#>  [5,] . . . . . 1 . . . .
#>  [6,] . . . 1 . . . . . .
#>  [7,] . . . . . . . 1 . .
#>  [8,] . . . . . . 1 . . .
#>  [9,] . . . . . . . . 1 .
#> [10,] . . . . . . . . . 1
#> 
#> $D
#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                          
#>  [1,] 1 . . . . . . . . .
#>  [2,] . 1 . . . . . . . .
#>  [3,] . . 1 . . . . . . .
#>  [4,] . . . . 1 . . . . .
#>  [5,] . . . . . 1 . . . .
#>  [6,] . . . 1 . . . . . .
#>  [7,] . . . . . . . 1 . .
#>  [8,] . . . . . . 1 . . .
#>  [9,] . . . . . . . . 1 .
#> [10,] . . . . . . . . . 1
#> 
#> $num_iter
#> [1] 2
#> 

 gp_list <- replicate(3, sample_correlated_gnp_pair(100, .3, .5), simplify = FALSE)
 A <- lapply(gp_list, function(gp)gp[[1]])
 B <- lapply(gp_list, function(gp)gp[[2]])
 match <- graph_match_FW(A, B, seeds = 1:10, start = "bari", max_iter = 20)
 match$corr

#>     corr_A corr_B
#> 1        1      1
#> 2        2      2
#> 3        3      3
#> 4        4      4
#> 5        5      5
#> 6        6      6
#> 7        7      7
#> 8        8      8
#> 9        9      9
#> 10      10     10
#> 11      11     11
#> 12      12     12
#> 13      13     13
#> 14      14     14
#> 15      15     15
#> 16      16     16
#> 17      17     17
#> 18      18     18
#> 19      19     19
#> 20      20     20
#> 21      21     21
#> 22      22     22
#> 23      23     23
#> 24      24     24
#> 25      25     25
#> 26      26     26
#> 27      27     27
#> 28      28     28
#> 29      29     29
#> 30      30     30
#> 31      31     31
#> 32      32     32
#> 33      33     33
#> 34      34     34
#> 35      35     35
#> 36      36     36
#> 37      37     37
#> 38      38     38
#> 39      39     39
#> 40      40     40
#> 41      41     41
#> 42      42     42
#> 43      43     43
#> 44      44     44
#> 45      45     45
#> 46      46     46
#> 47      47     47
#> 48      48     48
#> 49      49     49
#> 50      50     50
#> 51      51     51
#> 52      52     52
#> 53      53     53
#> 54      54     54
#> 55      55     55
#> 56      56     56
#> 57      57     57
#> 58      58     58
#> 59      59     59
#> 60      60     60
#> 61      61     61
#> 62      62     62
#> 63      63     63
#> 64      64     64
#> 65      65     65
#> 66      66     66
#> 67      67     67
#> 68      68     68
#> 69      69     69
#> 70      70     70
#> 71      71     71
#> 72      72     72
#> 73      73     73
#> 74      74     74
#> 75      75     75
#> 76      76     76
#> 77      77     77
#> 78      78     78
#> 79      79     79
#> 80      80     80
#> 81      81     81
#> 82      82     82
#> 83      83     83
#> 84      84     84
#> 85      85     85
#> 86      86     86
#> 87      87     87
#> 88      88     88
#> 89      89     89
#> 90      90     90
#> 91      91     91
#> 92      92     92
#> 93      93     93
#> 94      94     94
#> 95      95     95
#> 96      96     96
#> 97      97     97
#> 98      98     98
#> 99      99     99
#> 100    100    100

seeds <- 1:10 <= 3
graph_match_convex(g1, g2, seeds)

#> $call
#> graph_match_convex(A = g1, B = g2, seeds = seeds)
#> 
#> $corr
#>    corr_A corr_B
#> 1       1      1
#> 2       2      2
#> 3       3      3
#> 4       4      8
#> 5       5      5
#> 6       6     10
#> 7       7      6
#> 8       8      7
#> 9       9      9
#> 10     10      4
#> 
#> $ns
#> [1] 3
#> 
#> $P
#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                          
#>  [1,] 1 . . . . . . . . .
#>  [2,] . 1 . . . . . . . .
#>  [3,] . . 1 . . . . . . .
#>  [4,] . . . . . . . 1 . .
#>  [5,] . . . . 1 . . . . .
#>  [6,] . . . . . . . . . 1
#>  [7,] . . . . . 1 . . . .
#>  [8,] . . . . . . 1 . . .
#>  [9,] . . . . . . . . 1 .
#> [10,] . . . 1 . . . . . .
#> 
#> $D
#> Sparse part
#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                                                                              
#>  [1,] 1 . . .          .          .          .          .          .         
#>  [2,] . 1 . .          .          .          .          .          .         
#>  [3,] . . 1 .          .          .          .          .          .         
#>  [4,] . . . 0.13094970 0.19859335 0.15201277 .          0.27517216 .         
#>  [5,] . . . 0.11913853 0.40111065 0.25826019 0.04027245 0.08494896 0.08564368
#>  [6,] . . . 0.08949029 .          0.09100843 0.40847885 0.02120008 0.18984457
#>  [7,] . . . 0.15671849 0.13653036 0.28269271 0.03654878 0.16541532 0.04057926
#>  [8,] . . . 0.03871529 0.04057926 .          0.47895265 0.19189755 0.21410800
#>  [9,] . . . 0.17145926 0.15183484 0.07321282 0.02512172 0.06658902 0.45919896
#> [10,] . . . 0.28290291 0.06072601 0.13218753 .          0.18415138 .         
#>                 
#>  [1,] .         
#>  [2,] .         
#>  [3,] .         
#>  [4,] 0.23264648
#>  [5,] .         
#>  [6,] 0.18935223
#>  [7,] 0.17088954
#>  [8,] 0.02512172
#>  [9,] 0.04195784
#> [10,] 0.32940664
#> plus left factor
#> 10 x 1 sparse Matrix of class "dgCMatrix"
#>                 
#>  [1,] .         
#>  [2,] .         
#>  [3,] .         
#>  [4,] 0.01062554
#>  [5,] 0.01062554
#>  [6,] 0.01062554
#>  [7,] 0.01062554
#>  [8,] 0.01062554
#>  [9,] 0.01062554
#> [10,] 0.01062554
#> times right factor transpose
#> 10 x 1 sparse Matrix of class "dgCMatrix"
#>                
#>  [1,] .        
#>  [2,] .        
#>  [3,] .        
#>  [4,] 0.1428571
#>  [5,] 0.1428571
#>  [6,] 0.1428571
#>  [7,] 0.1428571
#>  [8,] 0.1428571
#>  [9,] 0.1428571
#> [10,] 0.1428571
#> 
#> $num_iter
#> [1] 100
#> 
#> $seq
#> $seq$alpha_seq
#>   [1] 0.1076707 0.5831962 0.8065760 0.8516346 0.8257825 0.9223287 0.9237181
#>   [8] 0.9354109 0.9464000 0.9585199 0.9632732 0.9617794 0.9503583 0.9669643
#>  [15] 0.9752237 0.9786937 0.9760522 0.9786866 0.9822018 0.9869172 0.9818157
#>  [22] 0.9781458 0.9824051 0.9804122 0.9863539 0.9864822 0.9891901 0.9853169
#>  [29] 0.9880383 0.9818069 0.9901659 0.9883028 0.9910934 0.9930368 0.9894673
#>  [36] 0.9839770 0.9906627 0.9935088 0.9921356 0.9918315 0.9872221 0.9921830
#>  [43] 0.9922363 0.9920575 0.9931195 0.9906574 0.9947572 0.9936680 0.9920599
#>  [50] 0.9955960 0.9877446 0.9922737 0.9937892 0.9926500 0.9931093 0.9950715
#>  [57] 0.9933537 0.9937059 0.9934875 0.9941581 0.9935178 0.9916240 0.9938127
#>  [64] 0.9957199 0.9945599 0.9935158 0.9952693 0.9939039 0.9954093 0.9941165
#>  [71] 0.9942263 0.9967243 0.9938397 0.9948680 0.9944109 0.9952822 0.9946550
#>  [78] 0.9962159 0.9946999 0.9951680 0.9955703 0.9946358 0.9950305 0.9962105
#>  [85] 0.9941263 0.9957302 0.9943910 0.9950177 0.9954722 0.9966827 0.9940848
#>  [92] 0.9966135 0.9961353 0.9958781 0.9957906 0.9953761 0.9968574 0.9962957
#>  [99] 0.9970417 0.9966657
#> 
#> $seq$Pseq
#> $seq$Pseq[[1]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . 1 . .
#> [2,] . 1 . . . . .
#> [3,] . . 1 . . . .
#> [4,] . . . 1 . . .
#> [5,] . . . . . 1 .
#> [6,] 1 . . . . . .
#> [7,] . . . . . . 1
#> 
#> $seq$Pseq[[2]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . 1 . . .
#> [2,] . . 1 . . . .
#> [3,] . . . . . 1 .
#> [4,] . . . . . . 1
#> [5,] 1 . . . . . .
#> [6,] . 1 . . . . .
#> [7,] . . . . 1 . .
#> 
#> $seq$Pseq[[3]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . . . 1
#> [2,] . . . . 1 . .
#> [3,] . . . . . 1 .
#> [4,] 1 . . . . . .
#> [5,] . 1 . . . . .
#> [6,] . . . 1 . . .
#> [7,] . . 1 . . . .
#> 
#> $seq$Pseq[[4]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . 1 . . . .
#> [2,] 1 . . . . . .
#> [3,] . . . . . . 1
#> [4,] . . . . . 1 .
#> [5,] . . . 1 . . .
#> [6,] . . . . 1 . .
#> [7,] . 1 . . . . .
#> 
#> $seq$Pseq[[5]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . 1 . . .
#> [2,] . 1 . . . . .
#> [3,] . . . . 1 . .
#> [4,] . . 1 . . . .
#> [5,] . . . . . 1 .
#> [6,] 1 . . . . . .
#> [7,] . . . . . . 1
#> 
#> $seq$Pseq[[6]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . 1 . . . .
#> [2,] . . . . 1 . .
#> [3,] 1 . . . . . .
#> [4,] . 1 . . . . .
#> [5,] . . . . . . 1
#> [6,] . . . . . 1 .
#> [7,] . . . 1 . . .
#> 
#> $seq$Pseq[[7]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . 1 . . . . .
#> [2,] . . . 1 . . .
#> [3,] 1 . . . . . .
#> [4,] . . . . . . 1
#> [5,] . . . . . 1 .
#> [6,] . . 1 . . . .
#> [7,] . . . . 1 . .
#> 
#> $seq$Pseq[[8]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . 1 . . .
#> [2,] . . 1 . . . .
#> [3,] . . . . . 1 .
#> [4,] . . . . 1 . .
#> [5,] 1 . . . . . .
#> [6,] . . . . . . 1
#> [7,] . 1 . . . . .
#> 
#> $seq$Pseq[[9]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . . . 1
#> [2,] . 1 . . . . .
#> [3,] . . . 1 . . .
#> [4,] . . . . 1 . .
#> [5,] . . . . . 1 .
#> [6,] 1 . . . . . .
#> [7,] . . 1 . . . .
#> 
#> $seq$Pseq[[10]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . 1 . . . . .
#> [2,] . . . 1 . . .
#> [3,] . . . . . 1 .
#> [4,] . . 1 . . . .
#> [5,] 1 . . . . . .
#> [6,] . . . . 1 . .
#> [7,] . . . . . . 1
#> 
#> $seq$Pseq[[11]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . 1 . . . .
#> [2,] . . . . . 1 .
#> [3,] . . . . . . 1
#> [4,] . 1 . . . . .
#> [5,] . . . . 1 . .
#> [6,] 1 . . . . . .
#> [7,] . . . 1 . . .
#> 
#> $seq$Pseq[[12]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . 1 . . . . .
#> [2,] . . 1 . . . .
#> [3,] . . . . . 1 .
#> [4,] . . . . . . 1
#> [5,] . . . 1 . . .
#> [6,] 1 . . . . . .
#> [7,] . . . . 1 . .
#> 
#> $seq$Pseq[[13]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . . . 1
#> [2,] . 1 . . . . .
#> [3,] 1 . . . . . .
#> [4,] . . 1 . . . .
#> [5,] . . . . . 1 .
#> [6,] . . . . 1 . .
#> [7,] . . . 1 . . .
#> 
#> $seq$Pseq[[14]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . 1 . . . . .
#> [2,] . . . . 1 . .
#> [3,] . . . . . 1 .
#> [4,] . . 1 . . . .
#> [5,] . . . 1 . . .
#> [6,] 1 . . . . . .
#> [7,] . . . . . . 1
#> 
#> $seq$Pseq[[15]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . 1 . . . .
#> [2,] . . . 1 . . .
#> [3,] . . . . . . 1
#> [4,] . 1 . . . . .
#> [5,] . . . . . 1 .
#> [6,] 1 . . . . . .
#> [7,] . . . . 1 . .
#> 
#> $seq$Pseq[[16]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . 1 . . .
#> [2,] . 1 . . . . .
#> [3,] . . . . . 1 .
#> [4,] . . . . 1 . .
#> [5,] 1 . . . . . .
#> [6,] . . 1 . . . .
#> [7,] . . . . . . 1
#> 
#> $seq$Pseq[[17]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . . . 1
#> [2,] . 1 . . . . .
#> [3,] 1 . . . . . .
#> [4,] . . 1 . . . .
#> [5,] . . . . . 1 .
#> [6,] . . . 1 . . .
#> [7,] . . . . 1 . .
#> 
#> $seq$Pseq[[18]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . 1 . . . . .
#> [2,] . . . . . 1 .
#> [3,] . . . . . . 1
#> [4,] . . 1 . . . .
#> [5,] . . . . 1 . .
#> [6,] 1 . . . . . .
#> [7,] . . . 1 . . .
#> 
#> $seq$Pseq[[19]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . . . 1
#> [2,] . . 1 . . . .
#> [3,] . . . . . 1 .
#> [4,] . 1 . . . . .
#> [5,] . . . 1 . . .
#> [6,] 1 . . . . . .
#> [7,] . . . . 1 . .
#> 
#> $seq$Pseq[[20]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . 1 . . . . .
#> [2,] . . 1 . . . .
#> [3,] 1 . . . . . .
#> [4,] . . . 1 . . .
#> [5,] . . . . . 1 .
#> [6,] . . . . 1 . .
#> [7,] . . . . . . 1
#> 
#> $seq$Pseq[[21]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . 1 . . .
#> [2,] . 1 . . . . .
#> [3,] . . . . . 1 .
#> [4,] . . . . 1 . .
#> [5,] 1 . . . . . .
#> [6,] . . . . . . 1
#> [7,] . . 1 . . . .
#> 
#> $seq$Pseq[[22]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . 1 . . . . .
#> [2,] . . 1 . . . .
#> [3,] . . . . . . 1
#> [4,] . . . 1 . . .
#> [5,] . . . . . 1 .
#> [6,] 1 . . . . . .
#> [7,] . . . . 1 . .
#> 
#> $seq$Pseq[[23]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . . . 1
#> [2,] . 1 . . . . .
#> [3,] . . . . . 1 .
#> [4,] . . 1 . . . .
#> [5,] 1 . . . . . .
#> [6,] . . . . 1 . .
#> [7,] . . . 1 . . .
#> 
#> $seq$Pseq[[24]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . 1 . . . .
#> [2,] . 1 . . . . .
#> [3,] . . . . . 1 .
#> [4,] . . . . 1 . .
#> [5,] . . . 1 . . .
#> [6,] 1 . . . . . .
#> [7,] . . . . . . 1
#> 
#> $seq$Pseq[[25]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . 1 . . .
#> [2,] 1 . . . . . .
#> [3,] . . . . . . 1
#> [4,] . . 1 . . . .
#> [5,] . . . . . 1 .
#> [6,] . 1 . . . . .
#> [7,] . . . . 1 . .
#> 
#> $seq$Pseq[[26]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . 1 . . . . .
#> [2,] . . 1 . . . .
#> [3,] . . . . 1 . .
#> [4,] . . . . . . 1
#> [5,] . . . . . 1 .
#> [6,] 1 . . . . . .
#> [7,] . . . 1 . . .
#> 
#> $seq$Pseq[[27]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . . . 1
#> [2,] . . . 1 . . .
#> [3,] 1 . . . . . .
#> [4,] . 1 . . . . .
#> [5,] . . . . . 1 .
#> [6,] . . . . 1 . .
#> [7,] . . 1 . . . .
#> 
#> $seq$Pseq[[28]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . 1 . .
#> [2,] . 1 . . . . .
#> [3,] . . . . . 1 .
#> [4,] . . 1 . . . .
#> [5,] . . . 1 . . .
#> [6,] 1 . . . . . .
#> [7,] . . . . . . 1
#> 
#> $seq$Pseq[[29]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . . . 1
#> [2,] . . 1 . . . .
#> [3,] 1 . . . . . .
#> [4,] . . . . 1 . .
#> [5,] . . . . . 1 .
#> [6,] . 1 . . . . .
#> [7,] . . . 1 . . .
#> 
#> $seq$Pseq[[30]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . 1 . . .
#> [2,] . 1 . . . . .
#> [3,] . . . . . . 1
#> [4,] . . 1 . . . .
#> [5,] . . . . . 1 .
#> [6,] 1 . . . . . .
#> [7,] . . . . 1 . .
#> 
#> $seq$Pseq[[31]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . 1 . . . . .
#> [2,] . . . . 1 . .
#> [3,] . . . . . 1 .
#> [4,] . . . 1 . . .
#> [5,] 1 . . . . . .
#> [6,] . . 1 . . . .
#> [7,] . . . . . . 1
#> 
#> $seq$Pseq[[32]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . 1 . . . .
#> [2,] 1 . . . . . .
#> [3,] . . . . . 1 .
#> [4,] . 1 . . . . .
#> [5,] . . . 1 . . .
#> [6,] . . . . 1 . .
#> [7,] . . . . . . 1
#> 
#> $seq$Pseq[[33]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . 1 . . . . .
#> [2,] . . 1 . . . .
#> [3,] . . . . 1 . .
#> [4,] . . . . . . 1
#> [5,] . . . . . 1 .
#> [6,] 1 . . . . . .
#> [7,] . . . 1 . . .
#> 
#> $seq$Pseq[[34]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . . . 1
#> [2,] . . . 1 . . .
#> [3,] . . . . . 1 .
#> [4,] . 1 . . . . .
#> [5,] . . . . 1 . .
#> [6,] 1 . . . . . .
#> [7,] . . 1 . . . .
#> 
#> $seq$Pseq[[35]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . 1 . . .
#> [2,] . . 1 . . . .
#> [3,] 1 . . . . . .
#> [4,] . . . . 1 . .
#> [5,] . . . . . 1 .
#> [6,] . 1 . . . . .
#> [7,] . . . . . . 1
#> 
#> $seq$Pseq[[36]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . . . 1
#> [2,] . 1 . . . . .
#> [3,] . . . . . 1 .
#> [4,] . . 1 . . . .
#> [5,] . . . 1 . . .
#> [6,] 1 . . . . . .
#> [7,] . . . . 1 . .
#> 
#> $seq$Pseq[[37]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . 1 . . . .
#> [2,] . 1 . . . . .
#> [3,] 1 . . . . . .
#> [4,] . . . . . . 1
#> [5,] . . . . . 1 .
#> [6,] . . . . 1 . .
#> [7,] . . . 1 . . .
#> 
#> $seq$Pseq[[38]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . 1 . .
#> [2,] . 1 . . . . .
#> [3,] . . . . . 1 .
#> [4,] . . 1 . . . .
#> [5,] 1 . . . . . .
#> [6,] . . . 1 . . .
#> [7,] . . . . . . 1
#> 
#> $seq$Pseq[[39]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . 1 . . .
#> [2,] . . . . 1 . .
#> [3,] . . . . . . 1
#> [4,] . 1 . . . . .
#> [5,] . . . . . 1 .
#> [6,] 1 . . . . . .
#> [7,] . . 1 . . . .
#> 
#> $seq$Pseq[[40]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . . . 1
#> [2,] . . 1 . . . .
#> [3,] . . . . . 1 .
#> [4,] . . . 1 . . .
#> [5,] 1 . . . . . .
#> [6,] . 1 . . . . .
#> [7,] . . . . 1 . .
#> 
#> $seq$Pseq[[41]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . 1 . . . . .
#> [2,] 1 . . . . . .
#> [3,] . . . . . . 1
#> [4,] . . . . 1 . .
#> [5,] . . . . . 1 .
#> [6,] . . 1 . . . .
#> [7,] . . . 1 . . .
#> 
#> $seq$Pseq[[42]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . 1 . .
#> [2,] . 1 . . . . .
#> [3,] . . . . . 1 .
#> [4,] . . 1 . . . .
#> [5,] . . . 1 . . .
#> [6,] 1 . . . . . .
#> [7,] . . . . . . 1
#> 
#> $seq$Pseq[[43]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . . . 1
#> [2,] . . 1 . . . .
#> [3,] 1 . . . . . .
#> [4,] . 1 . . . . .
#> [5,] . . . . . 1 .
#> [6,] . . . . 1 . .
#> [7,] . . . 1 . . .
#> 
#> $seq$Pseq[[44]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . 1 . . .
#> [2,] . 1 . . . . .
#> [3,] . . . . . . 1
#> [4,] . . . . 1 . .
#> [5,] . . . . . 1 .
#> [6,] 1 . . . . . .
#> [7,] . . 1 . . . .
#> 
#> $seq$Pseq[[45]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . . . 1
#> [2,] . . . 1 . . .
#> [3,] . . . . . 1 .
#> [4,] . . 1 . . . .
#> [5,] 1 . . . . . .
#> [6,] . 1 . . . . .
#> [7,] . . . . 1 . .
#> 
#> $seq$Pseq[[46]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . 1 . . . . .
#> [2,] . . 1 . . . .
#> [3,] . . . . 1 . .
#> [4,] . . . 1 . . .
#> [5,] . . . . . 1 .
#> [6,] 1 . . . . . .
#> [7,] . . . . . . 1
#> 
#> $seq$Pseq[[47]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . 1 . . . .
#> [2,] . . . . . 1 .
#> [3,] 1 . . . . . .
#> [4,] . 1 . . . . .
#> [5,] . . . 1 . . .
#> [6,] . . . . . . 1
#> [7,] . . . . 1 . .
#> 
#> $seq$Pseq[[48]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . 1 . . .
#> [2,] . 1 . . . . .
#> [3,] . . . . . 1 .
#> [4,] . . 1 . . . .
#> [5,] 1 . . . . . .
#> [6,] . . . . 1 . .
#> [7,] . . . . . . 1
#> 
#> $seq$Pseq[[49]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . 1 . . . .
#> [2,] . 1 . . . . .
#> [3,] . . . . . . 1
#> [4,] . . . 1 . . .
#> [5,] . . . . . 1 .
#> [6,] 1 . . . . . .
#> [7,] . . . . 1 . .
#> 
#> $seq$Pseq[[50]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . 1 . . . . .
#> [2,] . . 1 . . . .
#> [3,] 1 . . . . . .
#> [4,] . . . . . . 1
#> [5,] . . . . . 1 .
#> [6,] . . . . 1 . .
#> [7,] . . . 1 . . .
#> 
#> $seq$Pseq[[51]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . . . 1
#> [2,] . 1 . . . . .
#> [3,] . . . . . 1 .
#> [4,] . . 1 . . . .
#> [5,] . . . 1 . . .
#> [6,] 1 . . . . . .
#> [7,] . . . . 1 . .
#> 
#> $seq$Pseq[[52]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . 1 . . .
#> [2,] . . . . 1 . .
#> [3,] 1 . . . . . .
#> [4,] . . . . . . 1
#> [5,] . . . . . 1 .
#> [6,] . 1 . . . . .
#> [7,] . . 1 . . . .
#> 
#> $seq$Pseq[[53]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . 1 . . .
#> [2,] . 1 . . . . .
#> [3,] . . . . . 1 .
#> [4,] . . . . 1 . .
#> [5,] 1 . . . . . .
#> [6,] . . 1 . . . .
#> [7,] . . . . . . 1
#> 
#> $seq$Pseq[[54]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . 1 . . . . .
#> [2,] 1 . . . . . .
#> [3,] . . . . . . 1
#> [4,] . . 1 . . . .
#> [5,] . . . . . 1 .
#> [6,] . . . . 1 . .
#> [7,] . . . 1 . . .
#> 
#> $seq$Pseq[[55]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . . . 1
#> [2,] . . 1 . . . .
#> [3,] . . . . . 1 .
#> [4,] . 1 . . . . .
#> [5,] . . . 1 . . .
#> [6,] 1 . . . . . .
#> [7,] . . . . 1 . .
#> 
#> $seq$Pseq[[56]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . 1 . . . .
#> [2,] . 1 . . . . .
#> [3,] 1 . . . . . .
#> [4,] . . . . 1 . .
#> [5,] . . . . . 1 .
#> [6,] . . . 1 . . .
#> [7,] . . . . . . 1
#> 
#> $seq$Pseq[[57]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . 1 . . . . .
#> [2,] . . . . 1 . .
#> [3,] . . . . . . 1
#> [4,] . . 1 . . . .
#> [5,] . . . . . 1 .
#> [6,] 1 . . . . . .
#> [7,] . . . 1 . . .
#> 
#> $seq$Pseq[[58]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . 1 . .
#> [2,] . . 1 . . . .
#> [3,] . . . . . 1 .
#> [4,] . . . 1 . . .
#> [5,] 1 . . . . . .
#> [6,] . 1 . . . . .
#> [7,] . . . . . . 1
#> 
#> $seq$Pseq[[59]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . 1 . . .
#> [2,] . 1 . . . . .
#> [3,] . . . . . . 1
#> [4,] . . . . 1 . .
#> [5,] . . . . . 1 .
#> [6,] 1 . . . . . .
#> [7,] . . 1 . . . .
#> 
#> $seq$Pseq[[60]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . . . 1
#> [2,] 1 . . . . . .
#> [3,] . . . . . 1 .
#> [4,] . . 1 . . . .
#> [5,] . . . 1 . . .
#> [6,] . 1 . . . . .
#> [7,] . . . . 1 . .
#> 
#> $seq$Pseq[[61]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . . . 1
#> [2,] . 1 . . . . .
#> [3,] 1 . . . . . .
#> [4,] . . . 1 . . .
#> [5,] . . . . . 1 .
#> [6,] . . . . 1 . .
#> [7,] . . 1 . . . .
#> 
#> $seq$Pseq[[62]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . 1 . . .
#> [2,] . . 1 . . . .
#> [3,] . . . . . 1 .
#> [4,] . 1 . . . . .
#> [5,] . . . . 1 . .
#> [6,] 1 . . . . . .
#> [7,] . . . . . . 1
#> 
#> $seq$Pseq[[63]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . 1 . . . . .
#> [2,] . . 1 . . . .
#> [3,] . . . . 1 . .
#> [4,] . . . . . . 1
#> [5,] . . . . . 1 .
#> [6,] 1 . . . . . .
#> [7,] . . . 1 . . .
#> 
#> $seq$Pseq[[64]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . 1 . . .
#> [2,] . 1 . . . . .
#> [3,] . . . . . 1 .
#> [4,] . . . . 1 . .
#> [5,] 1 . . . . . .
#> [6,] . . 1 . . . .
#> [7,] . . . . . . 1
#> 
#> $seq$Pseq[[65]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . . . 1
#> [2,] . . 1 . . . .
#> [3,] 1 . . . . . .
#> [4,] . . . 1 . . .
#> [5,] . . . . . 1 .
#> [6,] . 1 . . . . .
#> [7,] . . . . 1 . .
#> 
#> $seq$Pseq[[66]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . 1 . .
#> [2,] . 1 . . . . .
#> [3,] . . . . . 1 .
#> [4,] . . 1 . . . .
#> [5,] . . . 1 . . .
#> [6,] 1 . . . . . .
#> [7,] . . . . . . 1
#> 
#> $seq$Pseq[[67]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . 1 . . . .
#> [2,] . . . 1 . . .
#> [3,] . . . . . . 1
#> [4,] . 1 . . . . .
#> [5,] . . . . . 1 .
#> [6,] 1 . . . . . .
#> [7,] . . . . 1 . .
#> 
#> $seq$Pseq[[68]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . 1 . . . . .
#> [2,] 1 . . . . . .
#> [3,] . . . . . . 1
#> [4,] . . 1 . . . .
#> [5,] . . . . . 1 .
#> [6,] . . . . 1 . .
#> [7,] . . . 1 . . .
#> 
#> $seq$Pseq[[69]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . 1 . . .
#> [2,] . 1 . . . . .
#> [3,] . . . . . 1 .
#> [4,] . . . . 1 . .
#> [5,] 1 . . . . . .
#> [6,] . . 1 . . . .
#> [7,] . . . . . . 1
#> 
#> $seq$Pseq[[70]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . 1 . . . . .
#> [2,] . . 1 . . . .
#> [3,] . . . . . 1 .
#> [4,] . . . . . . 1
#> [5,] . . . 1 . . .
#> [6,] 1 . . . . . .
#> [7,] . . . . 1 . .
#> 
#> $seq$Pseq[[71]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . . . 1
#> [2,] . 1 . . . . .
#> [3,] 1 . . . . . .
#> [4,] . . . . 1 . .
#> [5,] . . . . . 1 .
#> [6,] . . . 1 . . .
#> [7,] . . 1 . . . .
#> 
#> $seq$Pseq[[72]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . 1 . . .
#> [2,] . . . . . 1 .
#> [3,] . . . . . . 1
#> [4,] . . 1 . . . .
#> [5,] 1 . . . . . .
#> [6,] . 1 . . . . .
#> [7,] . . . . 1 . .
#> 
#> $seq$Pseq[[73]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . . . 1
#> [2,] . 1 . . . . .
#> [3,] . . . . 1 . .
#> [4,] . . 1 . . . .
#> [5,] . . . . . 1 .
#> [6,] 1 . . . . . .
#> [7,] . . . 1 . . .
#> 
#> $seq$Pseq[[74]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . 1 . . . .
#> [2,] . . . 1 . . .
#> [3,] 1 . . . . . .
#> [4,] . 1 . . . . .
#> [5,] . . . . . 1 .
#> [6,] . . . . 1 . .
#> [7,] . . . . . . 1
#> 
#> $seq$Pseq[[75]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . 1 . . . . .
#> [2,] . . 1 . . . .
#> [3,] . . . . . 1 .
#> [4,] . . . . . . 1
#> [5,] . . . 1 . . .
#> [6,] 1 . . . . . .
#> [7,] . . . . 1 . .
#> 
#> $seq$Pseq[[76]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . 1 . . .
#> [2,] . . . . 1 . .
#> [3,] . . . . . 1 .
#> [4,] . . 1 . . . .
#> [5,] 1 . . . . . .
#> [6,] . 1 . . . . .
#> [7,] . . . . . . 1
#> 
#> $seq$Pseq[[77]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . 1 . . . .
#> [2,] . 1 . . . . .
#> [3,] . . . . . . 1
#> [4,] . . . . 1 . .
#> [5,] . . . . . 1 .
#> [6,] 1 . . . . . .
#> [7,] . . . 1 . . .
#> 
#> $seq$Pseq[[78]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . 1 . . . . .
#> [2,] . . . . . 1 .
#> [3,] 1 . . . . . .
#> [4,] . . 1 . . . .
#> [5,] . . . 1 . . .
#> [6,] . . . . 1 . .
#> [7,] . . . . . . 1
#> 
#> $seq$Pseq[[79]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . . . 1
#> [2,] . . 1 . . . .
#> [3,] . . . . . 1 .
#> [4,] . 1 . . . . .
#> [5,] . . . 1 . . .
#> [6,] 1 . . . . . .
#> [7,] . . . . 1 . .
#> 
#> $seq$Pseq[[80]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . 1 . .
#> [2,] . 1 . . . . .
#> [3,] 1 . . . . . .
#> [4,] . . . 1 . . .
#> [5,] . . . . . 1 .
#> [6,] . . 1 . . . .
#> [7,] . . . . . . 1
#> 
#> $seq$Pseq[[81]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . 1 . . .
#> [2,] 1 . . . . . .
#> [3,] . . . . . . 1
#> [4,] . . 1 . . . .
#> [5,] . . . . . 1 .
#> [6,] . 1 . . . . .
#> [7,] . . . . 1 . .
#> 
#> $seq$Pseq[[82]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . . . 1
#> [2,] . 1 . . . . .
#> [3,] . . . . . 1 .
#> [4,] . . . . 1 . .
#> [5,] . . . 1 . . .
#> [6,] 1 . . . . . .
#> [7,] . . 1 . . . .
#> 
#> $seq$Pseq[[83]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . 1 . . . . .
#> [2,] . . 1 . . . .
#> [3,] . . . . . 1 .
#> [4,] . . . . . . 1
#> [5,] 1 . . . . . .
#> [6,] . . . . 1 . .
#> [7,] . . . 1 . . .
#> 
#> $seq$Pseq[[84]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . 1 . . .
#> [2,] . . . . 1 . .
#> [3,] . . . . . . 1
#> [4,] . 1 . . . . .
#> [5,] . . . . . 1 .
#> [6,] 1 . . . . . .
#> [7,] . . 1 . . . .
#> 
#> $seq$Pseq[[85]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . 1 . . .
#> [2,] . 1 . . . . .
#> [3,] 1 . . . . . .
#> [4,] . . 1 . . . .
#> [5,] . . . . . 1 .
#> [6,] . . . . 1 . .
#> [7,] . . . . . . 1
#> 
#> $seq$Pseq[[86]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . . . 1
#> [2,] . . 1 . . . .
#> [3,] . . . . . 1 .
#> [4,] . . . 1 . . .
#> [5,] 1 . . . . . .
#> [6,] . 1 . . . . .
#> [7,] . . . . 1 . .
#> 
#> $seq$Pseq[[87]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . 1 . . . .
#> [2,] . 1 . . . . .
#> [3,] . . . . . . 1
#> [4,] . . . . 1 . .
#> [5,] . . . . . 1 .
#> [6,] 1 . . . . . .
#> [7,] . . . 1 . . .
#> 
#> $seq$Pseq[[88]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . 1 . .
#> [2,] . 1 . . . . .
#> [3,] . . . . . 1 .
#> [4,] . . 1 . . . .
#> [5,] . . . 1 . . .
#> [6,] 1 . . . . . .
#> [7,] . . . . . . 1
#> 
#> $seq$Pseq[[89]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . . . 1
#> [2,] . . 1 . . . .
#> [3,] 1 . . . . . .
#> [4,] . . . 1 . . .
#> [5,] . . . . . 1 .
#> [6,] . 1 . . . . .
#> [7,] . . . . 1 . .
#> 
#> $seq$Pseq[[90]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . . . 1
#> [2,] . . . . 1 . .
#> [3,] . . . . . 1 .
#> [4,] . 1 . . . . .
#> [5,] 1 . . . . . .
#> [6,] . . 1 . . . .
#> [7,] . . . 1 . . .
#> 
#> $seq$Pseq[[91]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . 1 . . .
#> [2,] . 1 . . . . .
#> [3,] . . . . 1 . .
#> [4,] . . 1 . . . .
#> [5,] . . . . . 1 .
#> [6,] 1 . . . . . .
#> [7,] . . . . . . 1
#> 
#> $seq$Pseq[[92]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . 1 . . . . .
#> [2,] . . 1 . . . .
#> [3,] 1 . . . . . .
#> [4,] . . . . . . 1
#> [5,] . . . . . 1 .
#> [6,] . . . . 1 . .
#> [7,] . . . 1 . . .
#> 
#> $seq$Pseq[[93]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . 1 . . . .
#> [2,] . . . 1 . . .
#> [3,] . . . . . . 1
#> [4,] . 1 . . . . .
#> [5,] . . . . . 1 .
#> [6,] 1 . . . . . .
#> [7,] . . . . 1 . .
#> 
#> $seq$Pseq[[94]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . . . 1
#> [2,] 1 . . . . . .
#> [3,] . . . . . 1 .
#> [4,] . . 1 . . . .
#> [5,] . . . 1 . . .
#> [6,] . 1 . . . . .
#> [7,] . . . . 1 . .
#> 
#> $seq$Pseq[[95]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . 1 . . .
#> [2,] . 1 . . . . .
#> [3,] . . . . . . 1
#> [4,] . . . . 1 . .
#> [5,] . . . . . 1 .
#> [6,] 1 . . . . . .
#> [7,] . . 1 . . . .
#> 
#> $seq$Pseq[[96]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . 1 . . . . .
#> [2,] . . 1 . . . .
#> [3,] . . . . . 1 .
#> [4,] . . . 1 . . .
#> [5,] 1 . . . . . .
#> [6,] . . . . 1 . .
#> [7,] . . . . . . 1
#> 
#> $seq$Pseq[[97]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . . . 1
#> [2,] . . . . 1 . .
#> [3,] . . . . . 1 .
#> [4,] . 1 . . . . .
#> [5,] . . . 1 . . .
#> [6,] 1 . . . . . .
#> [7,] . . 1 . . . .
#> 
#> $seq$Pseq[[98]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . . 1 . .
#> [2,] . 1 . . . . .
#> [3,] 1 . . . . . .
#> [4,] . . 1 . . . .
#> [5,] . . . . . 1 .
#> [6,] . . . . . . 1
#> [7,] . . . 1 . . .
#> 
#> $seq$Pseq[[99]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . 1 . . . . .
#> [2,] . . . . . 1 .
#> [3,] . . 1 . . . .
#> [4,] . . . . 1 . .
#> [5,] . . . 1 . . .
#> [6,] 1 . . . . . .
#> [7,] . . . . . . 1
#> 
#> $seq$Pseq[[100]]
#> 7 x 7 sparse Matrix of class "dgCMatrix"
#>                   
#> [1,] . . . 1 . . .
#> [2,] . . 1 . . . .
#> [3,] . . . . . 1 .
#> [4,] . . . . . . 1
#> [5,] 1 . . . . . .
#> [6,] . 1 . . . . .
#> [7,] . . . . 1 . .
#> 
#> 
#> 

hard_seeds <- matrix(c(4,6,5,4),2)
seeds <- rbind(as.matrix(check_seeds(seeds, 10)$seeds),hard_seeds)
if (FALSE) {
# for some reason this fails in check
graph_match_convex(g1, g2, seeds)
}

# match G_1 & G_2 using PATH algorithm
graph_match_PATH(g1, g2)

#> Error in .local(x, i, j, ..., value): RHS 'value' (class dsparseVector) matches 'ANY', but must match matrix class splrMatrix

# match G_1 & G_2 using percolation graph matching method
seeds <- 1:5
graph_match_percolation(g1, g2, seeds, r = 2)

#> $call
#> graph_match_percolation(A = g1, B = g2, seeds = seeds, r = 2)
#> 
#> $corr
#>   corr_A corr_B
#> 1      1      1
#> 2      2      2
#> 3      3      3
#> 4      4      4
#> 5      5      5
#> 6      6      8
#> 7      8      7
#> 8      9      9
#> 9     10     10
#> 
#> $ns
#> [1] 5
#> 
#> $order
#> [1] 1 2 3 4 5 8 7 9 6
#> 

# match G_1 & G_2 using Expand When Stuck graph matching method
seeds <- 1:5
graph_match_ExpandWhenStuck(g1, g2, seeds, r = 2)

#> $call
#> graph_match_ExpandWhenStuck(A = g1, B = g2, seeds = seeds, r = 2)
#> 
#> $corr
#>   corr_A corr_B
#> 1      1      1
#> 2      2      2
#> 3      3      3
#> 4      4      4
#> 5      5      5
#> 6      6      8
#> 7      8      7
#> 8      9      9
#> 9     10     10
#> 
#> $ns
#> [1] 5
#> 
#> $order
#> [1] 1 2 3 4 5 8 6 9 7
#> 

# match G_1 & G_2 using soft percolation graph matching method
seeds <- 1:5
graph_match_soft_percolation(g1, g2, seeds, r = 2, max_iter = 2)

#> $call
#> graph_match_soft_percolation(A = g1, B = g2, seeds = seeds, r = 2, 
#>     max_iter = 2)
#> 
#> $corr
#>   corr_A corr_B
#> 1      1      1
#> 2      2      2
#> 3      3      3
#> 4      4      4
#> 5      5      5
#> 9      6      8
#> 6      8      7
#> 8      9      9
#> 7     10     10
#> 
#> $ns
#> [1] 5
#> 
#> $order
#> [1] 1 2 3 4 5 9 6 8 7
#> 

# match G_1 & G_2 using IsoRank algorithm
startm <- matrix(0, 10, 10)
diag(startm)[1:4] <- 1
GM_IsoRank <- graph_match_IsoRank(g1, g2, startm, alpha = .3, method = "greedy")

# match G_1 & G_2 using Umeyama algorithm
G <- sample_correlated_gnp_pair(10, .9, .5)
G1 <- G$graph1
G2 <- G$graph2
GM_U <- graph_match_Umeyama(G1, G2, startm, alpha = .3)