Initialization of the start matrix

Initialize the start matrix for graph matching iteration.

init_start(start, nns, ns = 0, soft_seeds = NULL, ...)

Arguments

start
A matrix or a character. Any nns-by-nns doubly stochastic matrix or start method like "bari", "convex" or "rds" to initialize the start matrix.

nns
An integer. Number of non-seeds.

ns
An integer. Number of seeds.

soft_seeds
A vector, a matrix or a data frame. If there is no error in the soft seeds, input can be a vector of soft seed indices in \(G_1\). Or if there is error in soft seeds, input should be in the form of a matrix or a data frame, with the first column being the indices of \(G_1\) and the second column being the corresponding indices of \(G_2\). Note that if there are seeds in graphs, seeds should be put before non-seeds.

...
Arguments passed to other start functions

Value

init_start returns a nns-by-nns doubly stochastic matrix as the start matrix in the graph matching iteration. If conduct a soft seeding graph matching, returns a nns-by-nns doubly stochastic matrix with 1's corresponding to the soft seeds and values at the other places are derived by different start method.

Examples

ss <- matrix(c(5, 4, 4, 3), nrow = 2)
# initialize start matrix without soft seeds
init_start(start = "bari", nns = 5)

#> Sparse part
#> 5 x 5 sparse Matrix of class "dsCMatrix"
#>               
#> [1,] . . . . .
#> [2,] . . . . .
#> [3,] . . . . .
#> [4,] . . . . .
#> [5,] . . . . .
#> plus left factor
#> 5 x 1 Matrix of class "dgeMatrix"
#>      [,1]
#> [1,]    1
#> [2,]    1
#> [3,]    1
#> [4,]    1
#> [5,]    1
#> times right factor transpose
#> 5 x 1 Matrix of class "dgeMatrix"
#>      [,1]
#> [1,]  0.2
#> [2,]  0.2
#> [3,]  0.2
#> [4,]  0.2
#> [5,]  0.2

init_start(start = "rds", nns = 3)

#> 3 x 3 Matrix of class "dgeMatrix"
#>            [,1]      [,2]      [,3]
#> [1,] 0.64077248 0.1763260 0.1829015
#> [2,] 0.04542403 0.5471491 0.4074268
#> [3,] 0.31380315 0.2765251 0.4096718

init_start(start = "rds_perm_bari", nns = 5)

#> Sparse part
#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>                                                       
#> [1,] .         .         0.1115607 .         .        
#> [2,] .         .         .         0.1115607 .        
#> [3,] .         0.1115607 .         .         .        
#> [4,] .         .         .         .         0.1115607
#> [5,] 0.1115607 .         .         .         .        
#> plus left factor
#> 5 x 1 Matrix of class "dgeMatrix"
#>           [,1]
#> [1,] 0.8884393
#> [2,] 0.8884393
#> [3,] 0.8884393
#> [4,] 0.8884393
#> [5,] 0.8884393
#> times right factor transpose
#> 5 x 1 Matrix of class "dgeMatrix"
#>      [,1]
#> [1,]  0.2
#> [2,]  0.2
#> [3,]  0.2
#> [4,]  0.2
#> [5,]  0.2

# initialize start matrix with soft seeds
init_start(start = "bari", nns = 5, ns = 3, soft_seeds = c(5, 7, 8))

#> Sparse part
#> 5 x 5 sparse Matrix of class "dsCMatrix"
#>               
#> [1,] . . . . .
#> [2,] . 1 . . .
#> [3,] . . . . .
#> [4,] . . . 1 .
#> [5,] . . . . 1
#> plus left factor
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>       
#> [1,] 1
#> [2,] .
#> [3,] 1
#> [4,] .
#> [5,] .
#> times right factor transpose
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>         
#> [1,] 0.5
#> [2,] .  
#> [3,] 0.5
#> [4,] .  
#> [5,] .  

init_start(start = "rds", nns = 5, soft_seeds = ss)

#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>                                           
#> [1,] 0.2467334 0.75276096 . . 0.0005056175
#> [2,] 0.3658244 0.16235676 . . 0.4718187980
#> [3,] 0.3874998 0.08460401 . . 0.5278961803
#> [4,] .         .          1 . .           
#> [5,] .         .          . 1 .           

init_start(start = "rds_perm_bari", nns = 5, soft_seeds = ss)

#> Sparse part
#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>                                          
#> [1,] .          0.06963822 . . .         
#> [2,] .          .          . . 0.06963822
#> [3,] 0.06963822 .          . . .         
#> [4,] .          .          1 . .         
#> [5,] .          .          . 1 .         
#> plus left factor
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>               
#> [1,] 0.9303618
#> [2,] 0.9303618
#> [3,] 0.9303618
#> [4,] .        
#> [5,] .        
#> times right factor transpose
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>               
#> [1,] 0.3333333
#> [2,] 0.3333333
#> [3,] .        
#> [4,] .        
#> [5,] 0.3333333

# initialize start matrix for convex graph matching
cgnp_pair <- sample_correlated_gnp_pair(n = 10, corr =  0.3, p =  0.5)
g1 <- cgnp_pair$graph1
g2 <- cgnp_pair$graph2
seeds <- 1:10 <= 2
init_start(start = "convex", nns = 8, A = g1, B = g2, seeds = seeds)

#> Sparse part
#> 6 x 6 sparse Matrix of class "dgCMatrix"
#>                                                                     
#> [1,] 0.36291383 0.06019802 .          0.17651139 .         0.2315202
#> [2,] 0.23152019 0.17651139 0.20759913 .          .         0.2175319
#> [3,] .          .          0.17651139 0.06019802 0.2075991 .        
#> [4,] 0.03463272 0.23152019 0.15733390 0.20759913 0.1808800 0.1765114
#> [5,] 0.02556530 0.20759913 0.23152019 0.15733390 0.2111441 0.1553147
#> [6,] 0.17651139 0.15531470 0.06019802 0.23152019 0.1573339 0.2075991
#> plus left factor
#> 6 x 1 sparse Matrix of class "dgCMatrix"
#>                
#> [1,] 0.01152266
#> [2,] 0.01152266
#> [3,] 0.01152266
#> [4,] 0.01152266
#> [5,] 0.01152266
#> [6,] 0.01152266
#> times right factor transpose
#> 6 x 1 sparse Matrix of class "dgCMatrix"
#>           
#> [1,] 0.125
#> [2,] 0.125
#> [3,] 0.125
#> [4,] 0.125
#> [5,] 0.125
#> [6,] 0.125

# FW graph matching with incorrect seeds to start at convex start
init_start(start = "convex", nns = 8, ns = 2, soft_seeds = ss, A = g1, B = g2, seeds = seeds)

#> Sparse part
#> 8 x 8 sparse Matrix of class "dgCMatrix"
#>                                                                       
#> [1,] 0.51977511 0.01341221 0.26447749 0.03056273 .          0.16080231
#> [2,] 0.28578275 .          0.08501762 0.04942849 .          0.37449512
#> [3,] 0.13027064 0.01097015 0.35178329 0.04150701 0.03591476 .         
#> [4,] 0.04150701 0.37555692 0.13109196 0.03890272 0.22040351 0.18156772
#> [5,] .          0.24618509 .          0.33843665 0.29342651 0.04872369
#> [6,] .          0.22350339 0.02034067 0.33854267 0.20888409 0.17509455
#> [7,] 0.02266448 .          0.11486616 0.08012357 0.05814788 .         
#> [8,] .          0.13037223 0.02145266 0.07152602 0.17225309 0.04834646
#>                           
#> [1,] .          .         
#> [2,] 0.16358945 0.04168657
#> [3,] 0.18263005 0.24692410
#> [4,] .          .         
#> [5,] 0.06225790 .         
#> [6,] 0.02266448 .         
#> [7,] 0.35392886 0.35929890
#> [8,] 0.20395911 0.34112028
#> plus left factor
#> 8 x 1 sparse Matrix of class "dgCMatrix"
#>                
#> [1,] 0.01097015
#> [2,] .         
#> [3,] .         
#> [4,] 0.01097015
#> [5,] 0.01097015
#> [6,] 0.01097015
#> [7,] 0.01097015
#> [8,] 0.01097015
#> times right factor transpose
#> 8 x 1 sparse Matrix of class "dgCMatrix"
#>               
#> [1,] .        
#> [2,] .        
#> [3,] 0.1666667
#> [4,] 0.1666667
#> [5,] 0.1666667
#> [6,] 0.1666667
#> [7,] 0.1666667
#> [8,] 0.1666667

#