Initialization of the start matrix

Initialize the start matrix for graph matching iteration.

init_start(start, nns, ns = 0, soft_seeds = NULL, ...)

Arguments

start
A matrix or a character. Any nns-by-nns doubly stochastic matrix or start method like "bari", "convex" or "rds" to initialize the start matrix.

nns
An integer. Number of non-seeds.

ns
An integer. Number of seeds.

soft_seeds
A vector, a matrix or a data frame. If there is no error in the soft seeds, input can be a vector of soft seed indices in \(G_1\). Or if there is error in soft seeds, input should be in the form of a matrix or a data frame, with the first column being the indices of \(G_1\) and the second column being the corresponding indices of \(G_2\). Note that if there are seeds in graphs, seeds should be put before non-seeds.

...
Arguments passed to other start functions

Value

init_start returns a nns-by-nns doubly stochastic matrix as the start matrix in the graph matching iteration. If conduct a soft seeding graph matching, returns a nns-by-nns doubly stochastic matrix with 1's corresponding to the soft seeds and values at the other places are derived by different start method.

Examples

ss <- matrix(c(5, 4, 4, 3), nrow = 2)
# initialize start matrix without soft seeds
init_start(start = "bari", nns = 5)

#> Sparse part
#> 5 x 5 sparse Matrix of class "dsCMatrix"
#>               
#> [1,] . . . . .
#> [2,] . . . . .
#> [3,] . . . . .
#> [4,] . . . . .
#> [5,] . . . . .
#> plus left factor
#> 5 x 1 Matrix of class "dgeMatrix"
#>      [,1]
#> [1,]    1
#> [2,]    1
#> [3,]    1
#> [4,]    1
#> [5,]    1
#> times right factor transpose
#> 5 x 1 Matrix of class "dgeMatrix"
#>      [,1]
#> [1,]  0.2
#> [2,]  0.2
#> [3,]  0.2
#> [4,]  0.2
#> [5,]  0.2

init_start(start = "rds", nns = 3)

#> 3 x 3 Matrix of class "dgeMatrix"
#>            [,1]      [,2]      [,3]
#> [1,] 0.46591328 0.1440870 0.3899997
#> [2,] 0.06273292 0.5234524 0.4138147
#> [3,] 0.47135384 0.3324606 0.1961856

init_start(start = "rds_perm_bari", nns = 5)

#> Sparse part
#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>                                                       
#> [1,] .         .         .         0.2406651 .        
#> [2,] .         0.2406651 .         .         .        
#> [3,] .         .         0.2406651 .         .        
#> [4,] 0.2406651 .         .         .         .        
#> [5,] .         .         .         .         0.2406651
#> plus left factor
#> 5 x 1 Matrix of class "dgeMatrix"
#>           [,1]
#> [1,] 0.7593349
#> [2,] 0.7593349
#> [3,] 0.7593349
#> [4,] 0.7593349
#> [5,] 0.7593349
#> times right factor transpose
#> 5 x 1 Matrix of class "dgeMatrix"
#>      [,1]
#> [1,]  0.2
#> [2,]  0.2
#> [3,]  0.2
#> [4,]  0.2
#> [5,]  0.2

# initialize start matrix with soft seeds
init_start(start = "bari", nns = 5, ns = 3, soft_seeds = c(5, 7, 8))

#> Sparse part
#> 5 x 5 sparse Matrix of class "dsCMatrix"
#>               
#> [1,] . . . . .
#> [2,] . 1 . . .
#> [3,] . . . . .
#> [4,] . . . 1 .
#> [5,] . . . . 1
#> plus left factor
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>       
#> [1,] 1
#> [2,] .
#> [3,] 1
#> [4,] .
#> [5,] .
#> times right factor transpose
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>         
#> [1,] 0.5
#> [2,] .  
#> [3,] 0.5
#> [4,] .  
#> [5,] .  

init_start(start = "rds", nns = 5, soft_seeds = ss)

#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>                                        
#> [1,] 0.01790314 0.4195558 . . 0.5625410
#> [2,] 0.60986526 0.1073521 . . 0.2827826
#> [3,] 0.37223063 0.4730925 . . 0.1546769
#> [4,] .          .         1 . .        
#> [5,] .          .         . 1 .        

init_start(start = "rds_perm_bari", nns = 5, soft_seeds = ss)

#> Sparse part
#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>                                          
#> [1,] 0.08287218 .          . . .         
#> [2,] .          .          . . 0.08287218
#> [3,] .          0.08287218 . . .         
#> [4,] .          .          1 . .         
#> [5,] .          .          . 1 .         
#> plus left factor
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>               
#> [1,] 0.9171278
#> [2,] 0.9171278
#> [3,] 0.9171278
#> [4,] .        
#> [5,] .        
#> times right factor transpose
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>               
#> [1,] 0.3333333
#> [2,] 0.3333333
#> [3,] .        
#> [4,] .        
#> [5,] 0.3333333

# initialize start matrix for convex graph matching
cgnp_pair <- sample_correlated_gnp_pair(n = 10, corr =  0.3, p =  0.5)
g1 <- cgnp_pair$graph1
g2 <- cgnp_pair$graph2
seeds <- 1:10 <= 2
init_start(start = "convex", nns = 8, A = g1, B = g2, seeds = seeds)

#> Sparse part
#> 6 x 6 sparse Matrix of class "dgCMatrix"
#>                                                                         
#> [1,] 0.492118258 .          0.12725601 0.02192094 0.01726841 0.265267221
#> [2,] .           0.38389808 0.26406848 0.12482148 .          0.047357953
#> [3,] 0.007284786 0.19020817 0.12019102 0.24484767 0.10219889 0.029794612
#> [4,] 0.067096462 0.18255193 0.05179577 0.36596951 0.01492448 0.009895003
#> [5,] 0.038542780 0.05417552 .          .          0.21516831 0.411250599
#> [6,] .           0.08597412 0.33594596 .          .          0.003601050
#> plus left factor
#> 6 x 1 sparse Matrix of class "dgCMatrix"
#>                 
#> [1,] 0.003845304
#> [2,] 0.003845304
#> [3,] 0.003845304
#> [4,] 0.003845304
#> [5,] 0.003845304
#> [6,] 0.003845304
#> times right factor transpose
#> 6 x 1 sparse Matrix of class "dgCMatrix"
#>           
#> [1,] 0.125
#> [2,] 0.125
#> [3,] 0.125
#> [4,] 0.125
#> [5,] 0.125
#> [6,] 0.125

# FW graph matching with incorrect seeds to start at convex start
init_start(start = "convex", nns = 8, ns = 2, soft_seeds = ss, A = g1, B = g2, seeds = seeds)

#> Sparse part
#> 8 x 8 sparse Matrix of class "dgCMatrix"
#>                                                                       
#> [1,] 0.50118224 .          0.12188410 0.04482034 .          0.26711032
#> [2,] 0.01259950 0.37382197 0.25421318 0.12101900 .          0.04482034
#> [3,] .          0.19543318 0.12092324 0.22307208 0.09996741 0.05010837
#> [4,] 0.07872498 0.18568827 0.06406459 0.37481268 0.01251864 .         
#> [5,] 0.04432658 0.05010837 .          .          0.22561893 0.39397070
#> [6,] .          0.11604503 0.32993099 .          .          0.02117906
#> [7,] 0.27381447 0.06793044 .          0.07199571 0.35440980 0.17266880
#> [8,] 0.08935223 0.01097274 0.09638440 0.15168069 0.29488573 0.03754291
#>                           
#> [1,] 0.02792995 0.02447355
#> [2,] .          0.19352601
#> [3,] .          0.31049572
#> [4,] .          0.27159135
#> [5,] 0.20463905 0.06873687
#> [6,] 0.52024541 .         
#> [7,] .          0.04658129
#> [8,] 0.23458609 0.07199571
#> plus left factor
#> 8 x 1 sparse Matrix of class "dgCMatrix"
#>               
#> [1,] 0.0125995
#> [2,] .        
#> [3,] .        
#> [4,] 0.0125995
#> [5,] 0.0125995
#> [6,] 0.0125995
#> [7,] 0.0125995
#> [8,] 0.0125995
#> times right factor transpose
#> 8 x 1 sparse Matrix of class "dgCMatrix"
#>               
#> [1,] .        
#> [2,] .        
#> [3,] 0.1666667
#> [4,] 0.1666667
#> [5,] 0.1666667
#> [6,] 0.1666667
#> [7,] 0.1666667
#> [8,] 0.1666667

#