Initialization of the start matrix

Initialize the start matrix for graph matching iteration.

init_start(start, nns, ns = 0, soft_seeds = NULL, ...)

Arguments

start
A matrix or a character. Any nns-by-nns doubly stochastic matrix or start method like "bari", "convex" or "rds" to initialize the start matrix.

nns
An integer. Number of non-seeds.

ns
An integer. Number of seeds.

soft_seeds
A vector, a matrix or a data frame. If there is no error in the soft seeds, input can be a vector of soft seed indices in \(G_1\). Or if there is error in soft seeds, input should be in the form of a matrix or a data frame, with the first column being the indices of \(G_1\) and the second column being the corresponding indices of \(G_2\). Note that if there are seeds in graphs, seeds should be put before non-seeds.

...
Arguments passed to other start functions

Value

init_start returns a nns-by-nns doubly stochastic matrix as the start matrix in the graph matching iteration. If conduct a soft seeding graph matching, returns a nns-by-nns doubly stochastic matrix with 1's corresponding to the soft seeds and values at the other places are derived by different start method.

Examples

ss <- matrix(c(5, 4, 4, 3), nrow = 2)
# initialize start matrix without soft seeds
init_start(start = "bari", nns = 5)

#> Sparse part
#> 5 x 5 sparse Matrix of class "dsCMatrix"
#>               
#> [1,] . . . . .
#> [2,] . . . . .
#> [3,] . . . . .
#> [4,] . . . . .
#> [5,] . . . . .
#> plus left factor
#> 5 x 1 Matrix of class "dgeMatrix"
#>      [,1]
#> [1,]    1
#> [2,]    1
#> [3,]    1
#> [4,]    1
#> [5,]    1
#> times right factor transpose
#> 5 x 1 Matrix of class "dgeMatrix"
#>      [,1]
#> [1,]  0.2
#> [2,]  0.2
#> [3,]  0.2
#> [4,]  0.2
#> [5,]  0.2

init_start(start = "rds", nns = 3)

#> 3 x 3 Matrix of class "dgeMatrix"
#>           [,1]      [,2]      [,3]
#> [1,] 0.3109804 0.5640669 0.1249527
#> [2,] 0.3698186 0.1917983 0.4383831
#> [3,] 0.3192010 0.2441348 0.4366642

init_start(start = "rds_perm_bari", nns = 5)

#> Sparse part
#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>                                                       
#> [1,] 0.9931905 .         .         .         .        
#> [2,] .         .         0.9931905 .         .        
#> [3,] .         .         .         .         0.9931905
#> [4,] .         .         .         0.9931905 .        
#> [5,] .         0.9931905 .         .         .        
#> plus left factor
#> 5 x 1 Matrix of class "dgeMatrix"
#>           [,1]
#> [1,] 0.0068095
#> [2,] 0.0068095
#> [3,] 0.0068095
#> [4,] 0.0068095
#> [5,] 0.0068095
#> times right factor transpose
#> 5 x 1 Matrix of class "dgeMatrix"
#>      [,1]
#> [1,]  0.2
#> [2,]  0.2
#> [3,]  0.2
#> [4,]  0.2
#> [5,]  0.2

# initialize start matrix with soft seeds
init_start(start = "bari", nns = 5, ns = 3, soft_seeds = c(5, 7, 8))

#> Sparse part
#> 5 x 5 sparse Matrix of class "dsCMatrix"
#>               
#> [1,] . . . . .
#> [2,] . 1 . . .
#> [3,] . . . . .
#> [4,] . . . 1 .
#> [5,] . . . . 1
#> plus left factor
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>       
#> [1,] 1
#> [2,] .
#> [3,] 1
#> [4,] .
#> [5,] .
#> times right factor transpose
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>         
#> [1,] 0.5
#> [2,] .  
#> [3,] 0.5
#> [4,] .  
#> [5,] .  

init_start(start = "rds", nns = 5, soft_seeds = ss)

#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>                                        
#> [1,] 0.2106996 0.1406841 . . 0.64861631
#> [2,] 0.5666020 0.1116064 . . 0.32179158
#> [3,] 0.2226989 0.7477070 . . 0.02959413
#> [4,] .         .         1 . .         
#> [5,] .         .         . 1 .         

init_start(start = "rds_perm_bari", nns = 5, soft_seeds = ss)

#> Sparse part
#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>                                       
#> [1,] 0.4972358 .         . . .        
#> [2,] .         .         . . 0.4972358
#> [3,] .         0.4972358 . . .        
#> [4,] .         .         1 . .        
#> [5,] .         .         . 1 .        
#> plus left factor
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>               
#> [1,] 0.5027642
#> [2,] 0.5027642
#> [3,] 0.5027642
#> [4,] .        
#> [5,] .        
#> times right factor transpose
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>               
#> [1,] 0.3333333
#> [2,] 0.3333333
#> [3,] .        
#> [4,] .        
#> [5,] 0.3333333

# initialize start matrix for convex graph matching
cgnp_pair <- sample_correlated_gnp_pair(n = 10, corr =  0.3, p =  0.5)
g1 <- cgnp_pair$graph1
g2 <- cgnp_pair$graph2
seeds <- 1:10 <= 2
init_start(start = "convex", nns = 8, A = g1, B = g2, seeds = seeds)

#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                                                                     
#>  [1,] 1 .  .           .           .           .           .        
#>  [2,] . 1  .           .           .           .           .        
#>  [3,] . .  0.25900705  0.27353260 -0.32730555  0.20750821  0.2081927
#>  [4,] . .  0.29998966  0.04270292  0.28525112  0.05917776  0.1163970
#>  [5,] . . -0.07205621  0.29073404  0.15524510  0.02861319  0.1299984
#>  [6,] . .  0.22497471  0.05403809  0.15298117  0.22841139  0.2797368
#>  [7,] . . -0.24756823 -0.01136870  0.30928743  0.29106059  0.3092874
#>  [8,] . .  0.15619811 -0.22836247  0.23805921  0.29803475  0.3092874
#>  [9,] . .  0.30928743  0.29134432 -0.09394638 -0.07205621 -0.1000042
#> [10,] . .  0.07016749  0.28737921  0.28042791 -0.04074968 -0.2528955
#>                                         
#>  [1,]  .           .          .         
#>  [2,]  .           .          .         
#>  [3,]  0.10765049  0.1552451  0.11616941
#>  [4,]  0.30928743 -0.1668440  0.05403809
#>  [5,] -0.04774113  0.2781481  0.23705851
#>  [6,]  0.25928750  0.1284481 -0.32787777
#>  [7,] -0.13357470  0.3092874  0.17358875
#>  [8,]  0.12460522 -0.0262704  0.12844815
#>  [9,]  0.13527788  0.2208098  0.30928743
#> [10,]  0.24520730  0.1011758  0.30928743

# FW graph matching with incorrect seeds to start at convex start
init_start(start = "convex", nns = 8, ns = 2, soft_seeds = ss, A = g1, B = g2, seeds = seeds)

#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                                                                         
#>  [1,] 1 . . .  .           .           .           .           .        
#>  [2,] . 1 . .  .           .           .           .           .        
#>  [3,] . . . .  0.05605747 -0.08824271  0.06762521  0.20889250  0.3519912
#>  [4,] . . 1 .  .           .           .           .           .        
#>  [5,] . . . 1  .           .           .           .           .        
#>  [6,] . . . .  0.34218427  0.33619579  0.39669692 -0.08824271 -0.1740249
#>  [7,] . . . .  0.08796762  0.41888912  0.25926356 -0.06345349  0.4893459
#>  [8,] . . . .  0.43645581  0.05605747  0.39444692  0.36730849 -0.3210995
#>  [9,] . . . .  0.25926356  0.08589306 -0.17409008  0.33619579  0.4366802
#> [10,] . . . . -0.18192873  0.19120727  0.05605747  0.23929943  0.2171071
#>                  
#>  [1,]  .         
#>  [2,]  .         
#>  [3,]  0.40367634
#>  [4,]  .         
#>  [5,]  .         
#>  [6,]  0.18719065
#>  [7,] -0.19201274
#>  [8,]  0.06683085
#>  [9,]  0.05605747
#> [10,]  0.47825744

#