Initialization of the start matrix

Initialize the start matrix for graph matching iteration.

init_start(start, nns, ns = 0, soft_seeds = NULL, ...)

Arguments

start
A matrix or a character. Any nns-by-nns doubly stochastic matrix or start method like "bari", "convex" or "rds" to initialize the start matrix.

nns
An integer. Number of non-seeds.

ns
An integer. Number of seeds.

soft_seeds
A vector, a matrix or a data frame. If there is no error in the soft seeds, input can be a vector of soft seed indices in \(G_1\). Or if there is error in soft seeds, input should be in the form of a matrix or a data frame, with the first column being the indices of \(G_1\) and the second column being the corresponding indices of \(G_2\). Note that if there are seeds in graphs, seeds should be put before non-seeds.

...
Arguments passed to other start functions

Value

init_start returns a nns-by-nns doubly stochastic matrix as the start matrix in the graph matching iteration. If conduct a soft seeding graph matching, returns a nns-by-nns doubly stochastic matrix with 1's corresponding to the soft seeds and values at the other places are derived by different start method.

Examples

ss <- matrix(c(5, 4, 4, 3), nrow = 2)
# initialize start matrix without soft seeds
init_start(start = "bari", nns = 5)

#> Sparse part
#> 5 x 5 sparse Matrix of class "dsCMatrix"
#>               
#> [1,] . . . . .
#> [2,] . . . . .
#> [3,] . . . . .
#> [4,] . . . . .
#> [5,] . . . . .
#> plus left factor
#> 5 x 1 Matrix of class "dgeMatrix"
#>      [,1]
#> [1,]    1
#> [2,]    1
#> [3,]    1
#> [4,]    1
#> [5,]    1
#> times right factor transpose
#> 5 x 1 Matrix of class "dgeMatrix"
#>      [,1]
#> [1,]  0.2
#> [2,]  0.2
#> [3,]  0.2
#> [4,]  0.2
#> [5,]  0.2

init_start(start = "rds", nns = 3)

#> 3 x 3 Matrix of class "dgeMatrix"
#>             [,1]      [,2]       [,3]
#> [1,] 0.004768706 0.2355898 0.75964148
#> [2,] 0.354060240 0.4316473 0.21429245
#> [3,] 0.641172719 0.3327632 0.02606407

init_start(start = "rds_perm_bari", nns = 5)

#> Sparse part
#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>                                                       
#> [1,] .         0.2271598 .         .         .        
#> [2,] .         .         .         0.2271598 .        
#> [3,] .         .         0.2271598 .         .        
#> [4,] 0.2271598 .         .         .         .        
#> [5,] .         .         .         .         0.2271598
#> plus left factor
#> 5 x 1 Matrix of class "dgeMatrix"
#>           [,1]
#> [1,] 0.7728402
#> [2,] 0.7728402
#> [3,] 0.7728402
#> [4,] 0.7728402
#> [5,] 0.7728402
#> times right factor transpose
#> 5 x 1 Matrix of class "dgeMatrix"
#>      [,1]
#> [1,]  0.2
#> [2,]  0.2
#> [3,]  0.2
#> [4,]  0.2
#> [5,]  0.2

# initialize start matrix with soft seeds
init_start(start = "bari", nns = 5, ns = 3, soft_seeds = c(5, 7, 8))

#> Sparse part
#> 5 x 5 sparse Matrix of class "dsCMatrix"
#>               
#> [1,] . . . . .
#> [2,] . 1 . . .
#> [3,] . . . . .
#> [4,] . . . 1 .
#> [5,] . . . . 1
#> plus left factor
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>       
#> [1,] 1
#> [2,] .
#> [3,] 1
#> [4,] .
#> [5,] .
#> times right factor transpose
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>         
#> [1,] 0.5
#> [2,] .  
#> [3,] 0.5
#> [4,] .  
#> [5,] .  

init_start(start = "rds", nns = 5, soft_seeds = ss)

#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>                                       
#> [1,] 0.2730231 0.4608101 . . 0.2661669
#> [2,] 0.4221339 0.3551028 . . 0.2227633
#> [3,] 0.3048431 0.1840871 . . 0.5110699
#> [4,] .         .         1 . .        
#> [5,] .         .         . 1 .        

init_start(start = "rds_perm_bari", nns = 5, soft_seeds = ss)

#> Sparse part
#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>                                       
#> [1,] .         0.3182345 . . .        
#> [2,] .         .         . . 0.3182345
#> [3,] 0.3182345 .         . . .        
#> [4,] .         .         1 . .        
#> [5,] .         .         . 1 .        
#> plus left factor
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>               
#> [1,] 0.6817655
#> [2,] 0.6817655
#> [3,] 0.6817655
#> [4,] .        
#> [5,] .        
#> times right factor transpose
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>               
#> [1,] 0.3333333
#> [2,] 0.3333333
#> [3,] .        
#> [4,] .        
#> [5,] 0.3333333

# initialize start matrix for convex graph matching
cgnp_pair <- sample_correlated_gnp_pair(n = 10, corr =  0.3, p =  0.5)
g1 <- cgnp_pair$graph1
g2 <- cgnp_pair$graph2
seeds <- 1:10 <= 2
init_start(start = "convex", nns = 8, A = g1, B = g2, seeds = seeds)

#> Sparse part
#> 6 x 6 sparse Matrix of class "dgCMatrix"
#>                                                                     
#> [1,] 0.27761180 0.1753132 .          0.01676391 0.18188123 0.1943676
#> [2,] 0.20314547 0.1045559 0.14246265 0.11573324 0.15087983 .        
#> [3,] .          0.2581959 0.08779202 .          0.22048567 .        
#> [4,] 0.14246265 .         0.17531323 0.26051562 .          0.3223169
#> [5,] 0.20771232 0.2093780 .          0.02443340 0.08779202 0.1424627
#> [6,] 0.03066589 .         0.19547597 0.27761180 0.14246265 0.2264509
#> plus left factor
#> 6 x 1 sparse Matrix of class "dgCMatrix"
#>                
#> [1,] 0.01159955
#> [2,] 0.01159955
#> [3,] 0.01159955
#> [4,] 0.01159955
#> [5,] 0.01159955
#> [6,] 0.01159955
#> times right factor transpose
#> 6 x 1 sparse Matrix of class "dgCMatrix"
#>           
#> [1,] 0.125
#> [2,] 0.125
#> [3,] 0.125
#> [4,] 0.125
#> [5,] 0.125
#> [6,] 0.125

# FW graph matching with incorrect seeds to start at convex start
init_start(start = "convex", nns = 8, ns = 2, soft_seeds = ss, A = g1, B = g2, seeds = seeds)

#> Sparse part
#> 8 x 8 sparse Matrix of class "dgCMatrix"
#>                                                                       
#> [1,] 0.17200892 0.08473046 .          0.01119119 0.27806336 0.12018736
#> [2,] 0.20359053 0.32136491 0.06602492 0.09143439 .          0.10590838
#> [3,] .          0.22998321 0.10827426 .          0.20113757 .         
#> [4,] 0.04841014 0.05759191 0.28176139 0.43708345 .          0.12841344
#> [5,] 0.25487198 0.11852747 0.03959074 0.05759191 0.23404022 0.07096884
#> [6,] 0.11299203 .          0.05345183 0.15822196 0.01799445 0.36905336
#> [7,] .          .          0.34656527 .          0.14910976 0.04573691
#> [8,] 0.20812640 0.18780204 0.05759191 0.19773742 0.07291496 0.11299203
#>                          
#> [1,] 0.1152498 0.17182924
#> [2,] 0.1269464 0.08473046
#> [3,] 0.3073453 0.15325967
#> [4,] .         .         
#> [5,] 0.1129920 0.06467713
#> [6,] .         0.24154668
#> [7,] 0.1867850 0.22506341
#> [8,] 0.1039418 0.01215373
#> plus left factor
#> 8 x 1 sparse Matrix of class "dgCMatrix"
#>                
#> [1,] 0.04673968
#> [2,] .         
#> [3,] .         
#> [4,] 0.04673968
#> [5,] 0.04673968
#> [6,] 0.04673968
#> [7,] 0.04673968
#> [8,] 0.04673968
#> times right factor transpose
#> 8 x 1 sparse Matrix of class "dgCMatrix"
#>               
#> [1,] .        
#> [2,] .        
#> [3,] 0.1666667
#> [4,] 0.1666667
#> [5,] 0.1666667
#> [6,] 0.1666667
#> [7,] 0.1666667
#> [8,] 0.1666667

#