Initialization of the start matrix

Initialize the start matrix for graph matching iteration.

init_start(
  start,
  nns,
  ns = 0,
  soft_seeds = NULL,
  A = NULL,
  B = NULL,
  seeds = NULL,
  g = 1
)

Arguments

start
A matrix or a character. Any nns-by-nns doubly stochastic matrix or start method like "bari", "convex" or "rds" to initialize the start matrix.

nns
An integer. Number of non-seeds.

ns
An integer. Number of seeds.

soft_seeds
A vector, a matrix or a data frame. If there is no error in the soft seeds, input can be a vector of soft seed indices in \(G_1\). Or if there is error in soft seeds, input should be in the form of a matrix or a data frame, with the first column being the indices of \(G_1\) and the second column being the corresponding indices of \(G_2\). Note that if there are seeds in graphs, seeds should be put before non-seeds.

A
A matrix or an igraph object. Adjacency matrix of \(G_1\). Needed only when start is convex.

B
A matrix or an igraph object. Adjacency matrix of \(G_2\). Needed only when start is convex.

seeds
A logical vector. TRUE indicates the corresponding vertex is a seed. Needed only when start is convex.

g
A number. Specified in the range of [0,1] to set weights to random permutaion matrix and barycenter matrix.

Value

init_start returns a nns-by-nns doubly stochastic matrix as the start matrix in the graph matching iteration. If conduct a soft seeding graph matching, returns a nns-by-nns doubly stochastic matrix with 1's corresponding to the soft seeds and values at the other places are derived by different start method.

Examples

ss <- matrix(c(5,4,4,3), nrow = 2)
# initialize start matrix without soft seeds
init_start(start = "bari", nns = 5)

#> Sparse part
#> 5 x 5 sparse Matrix of class "dsCMatrix"
#>               
#> [1,] . . . . .
#> [2,] . . . . .
#> [3,] . . . . .
#> [4,] . . . . .
#> [5,] . . . . .
#> plus left factor
#> 5 x 1 Matrix of class "dgeMatrix"
#>      [,1]
#> [1,]    1
#> [2,]    1
#> [3,]    1
#> [4,]    1
#> [5,]    1
#> times right factor transpose
#> 5 x 1 Matrix of class "dgeMatrix"
#>      [,1]
#> [1,]  0.2
#> [2,]  0.2
#> [3,]  0.2
#> [4,]  0.2
#> [5,]  0.2

init_start(start = "rds", nns = 3)

#>           [,1]       [,2]       [,3]
#> [1,] 0.4372804 0.09380979 0.46890980
#> [2,] 0.1850818 0.34303726 0.47188098
#> [3,] 0.3776378 0.56315295 0.05920921

init_start(start = "rds_perm_bari", nns = 5)

#> Sparse part
#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>                                                  
#> [1,] .        0.312055 .        .        .       
#> [2,] .        .        0.312055 .        .       
#> [3,] .        .        .        .        0.312055
#> [4,] 0.312055 .        .        .        .       
#> [5,] .        .        .        0.312055 .       
#> plus left factor
#> 5 x 1 Matrix of class "dgeMatrix"
#>          [,1]
#> [1,] 0.687945
#> [2,] 0.687945
#> [3,] 0.687945
#> [4,] 0.687945
#> [5,] 0.687945
#> times right factor transpose
#> 5 x 1 Matrix of class "dgeMatrix"
#>      [,1]
#> [1,]  0.2
#> [2,]  0.2
#> [3,]  0.2
#> [4,]  0.2
#> [5,]  0.2

# initialize start matrix with soft seeds
init_start(start = "bari", nns = 5, ns = 3, soft_seeds = c(5,7,8))

#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]  0.5    0  0.5    0    0
#> [2,]  0.0    1  0.0    0    0
#> [3,]  0.5    0  0.5    0    0
#> [4,]  0.0    0  0.0    1    0
#> [5,]  0.0    0  0.0    0    1

init_start(start = "rds", nns = 5, soft_seeds = ss)

#>           [,1]      [,2] [,3] [,4]      [,5]
#> [1,] 0.4888451 0.2553411    0    0 0.2558132
#> [2,] 0.1734544 0.6735817    0    0 0.1529675
#> [3,] 0.3377005 0.0710772    0    0 0.5912193
#> [4,] 0.0000000 0.0000000    1    0 0.0000000
#> [5,] 0.0000000 0.0000000    0    1 0.0000000

init_start(start = "rds_perm_bari", nns = 5, soft_seeds = ss)

#> Sparse part
#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>                                       
#> [1,] .         0.8909552 . . .        
#> [2,] 0.8909552 .         . . .        
#> [3,] .         .         . . 0.8909552
#> [4,] .         .         1 . .        
#> [5,] .         .         . 1 .        
#> plus left factor
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>               
#> [1,] 0.1090448
#> [2,] 0.1090448
#> [3,] 0.1090448
#> [4,] .        
#> [5,] .        
#> times right factor transpose
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>               
#> [1,] 0.3333333
#> [2,] 0.3333333
#> [3,] .        
#> [4,] .        
#> [5,] 0.3333333

# initialize start matrix for convex graph matching
cgnp_pair <- sample_correlated_gnp_pair(n = 10, corr =  0.3, p =  0.5)
g1 <- cgnp_pair$graph1
g2 <- cgnp_pair$graph2
seeds <- 1:10 <= 2
init_start(start = "convex", nns = 8, A = g1, B = g2, seeds = seeds)

#> 8 x 8 sparse Matrix of class "dgCMatrix"
#>                                                                           
#> [1,] 0.1187865 0.1684945 0.1187865 0.1187865 0.1187865 0.1187865 0.1187865
#> [2,] 0.1187865 0.1187865 0.1187865 0.1187865 0.1187865 0.1684945 0.1187865
#> [3,] 0.1187865 0.1187865 0.1187865 0.1187865 0.1187865 0.1187865 0.1187865
#> [4,] 0.1187865 0.1187865 0.1187865 0.1684945 0.1187865 0.1187865 0.1187865
#> [5,] 0.1684945 0.1187865 0.1187865 0.1187865 0.1187865 0.1187865 0.1187865
#> [6,] 0.1187865 0.1187865 0.1684945 0.1187865 0.1187865 0.1187865 0.1187865
#> [7,] 0.1187865 0.1187865 0.1187865 0.1187865 0.1187865 0.1187865 0.1684945
#> [8,] 0.1187865 0.1187865 0.1187865 0.1187865 0.1684945 0.1187865 0.1187865
#>               
#> [1,] 0.1187865
#> [2,] 0.1187865
#> [3,] 0.1684945
#> [4,] 0.1187865
#> [5,] 0.1187865
#> [6,] 0.1187865
#> [7,] 0.1187865
#> [8,] 0.1187865

# FW graph matching with incorrect seeds to start at convex start
init_start(start = "convex", nns = 8, ns = 2, soft_seeds = ss, A = g1, B = g2, seeds = seeds)

#> 8 x 8 sparse Matrix of class "dgCMatrix"
#>                                                                              
#> [1,] . .  0.09953393  0.3651559  0.08739513  0.2519072 -0.1392863  0.33529408
#> [2,] 1 .  .           .          .           .          .          .         
#> [3,] . 1  .           .          .           .          .          .         
#> [4,] . .  0.25190724 -0.1885914  0.13295067  0.3479042  0.3651559  0.09067349
#> [5,] . .  0.36515588  0.2360188  0.35118252 -0.4137686  0.3651559  0.09625557
#> [6,] . . -0.13928626  0.3524093  0.20260206  0.3492674 -0.1301484  0.36515588
#> [7,] . .  0.09067349  0.3606507  0.36515588  0.1488391  0.1739671 -0.13928626
#> [8,] . .  0.33201572 -0.1256433 -0.13928626  0.3158507  0.3651559  0.25190724

#