Initialization of the start matrix

Initialize the start matrix for graph matching iteration.

init_start(start, nns, ns = 0, soft_seeds = NULL, ...)

Arguments

start
A matrix or a character. Any nns-by-nns doubly stochastic matrix or start method like "bari", "convex" or "rds" to initialize the start matrix.

nns
An integer. Number of non-seeds.

ns
An integer. Number of seeds.

soft_seeds
A vector, a matrix or a data frame. If there is no error in the soft seeds, input can be a vector of soft seed indices in \(G_1\). Or if there is error in soft seeds, input should be in the form of a matrix or a data frame, with the first column being the indices of \(G_1\) and the second column being the corresponding indices of \(G_2\). Note that if there are seeds in graphs, seeds should be put before non-seeds.

...
Arguments passed to other start functions

Value

init_start returns a nns-by-nns doubly stochastic matrix as the start matrix in the graph matching iteration. If conduct a soft seeding graph matching, returns a nns-by-nns doubly stochastic matrix with 1's corresponding to the soft seeds and values at the other places are derived by different start method.

Examples

ss <- matrix(c(5, 4, 4, 3), nrow = 2)
# initialize start matrix without soft seeds
init_start(start = "bari", nns = 5)

#> Sparse part
#> 5 x 5 sparse Matrix of class "dsCMatrix"
#>               
#> [1,] . . . . .
#> [2,] . . . . .
#> [3,] . . . . .
#> [4,] . . . . .
#> [5,] . . . . .
#> plus left factor
#> 5 x 1 Matrix of class "dgeMatrix"
#>      [,1]
#> [1,]    1
#> [2,]    1
#> [3,]    1
#> [4,]    1
#> [5,]    1
#> times right factor transpose
#> 5 x 1 Matrix of class "dgeMatrix"
#>      [,1]
#> [1,]  0.2
#> [2,]  0.2
#> [3,]  0.2
#> [4,]  0.2
#> [5,]  0.2

init_start(start = "rds", nns = 3)

#> 3 x 3 Matrix of class "dgeMatrix"
#>           [,1]      [,2]      [,3]
#> [1,] 0.3109804 0.5640669 0.1249527
#> [2,] 0.3698186 0.1917983 0.4383831
#> [3,] 0.3192010 0.2441348 0.4366642

init_start(start = "rds_perm_bari", nns = 5)

#> Sparse part
#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>                                                       
#> [1,] 0.9931905 .         .         .         .        
#> [2,] .         .         0.9931905 .         .        
#> [3,] .         .         .         .         0.9931905
#> [4,] .         .         .         0.9931905 .        
#> [5,] .         0.9931905 .         .         .        
#> plus left factor
#> 5 x 1 Matrix of class "dgeMatrix"
#>           [,1]
#> [1,] 0.0068095
#> [2,] 0.0068095
#> [3,] 0.0068095
#> [4,] 0.0068095
#> [5,] 0.0068095
#> times right factor transpose
#> 5 x 1 Matrix of class "dgeMatrix"
#>      [,1]
#> [1,]  0.2
#> [2,]  0.2
#> [3,]  0.2
#> [4,]  0.2
#> [5,]  0.2

# initialize start matrix with soft seeds
init_start(start = "bari", nns = 5, ns = 3, soft_seeds = c(5, 7, 8))

#> Sparse part
#> 5 x 5 sparse Matrix of class "dsCMatrix"
#>               
#> [1,] . . . . .
#> [2,] . 1 . . .
#> [3,] . . . . .
#> [4,] . . . 1 .
#> [5,] . . . . 1
#> plus left factor
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>       
#> [1,] 1
#> [2,] .
#> [3,] 1
#> [4,] .
#> [5,] .
#> times right factor transpose
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>         
#> [1,] 0.5
#> [2,] .  
#> [3,] 0.5
#> [4,] .  
#> [5,] .  

init_start(start = "rds", nns = 5, soft_seeds = ss)

#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>                                        
#> [1,] 0.2106996 0.1406841 . . 0.64861631
#> [2,] 0.5666020 0.1116064 . . 0.32179158
#> [3,] 0.2226989 0.7477070 . . 0.02959413
#> [4,] .         .         1 . .         
#> [5,] .         .         . 1 .         

init_start(start = "rds_perm_bari", nns = 5, soft_seeds = ss)

#> Sparse part
#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>                                       
#> [1,] 0.4972358 .         . . .        
#> [2,] .         .         . . 0.4972358
#> [3,] .         0.4972358 . . .        
#> [4,] .         .         1 . .        
#> [5,] .         .         . 1 .        
#> plus left factor
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>               
#> [1,] 0.5027642
#> [2,] 0.5027642
#> [3,] 0.5027642
#> [4,] .        
#> [5,] .        
#> times right factor transpose
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>               
#> [1,] 0.3333333
#> [2,] 0.3333333
#> [3,] .        
#> [4,] .        
#> [5,] 0.3333333

# initialize start matrix for convex graph matching
cgnp_pair <- sample_correlated_gnp_pair(n = 10, corr =  0.3, p =  0.5)
g1 <- cgnp_pair$graph1
g2 <- cgnp_pair$graph2
seeds <- 1:10 <= 2
init_start(start = "convex", nns = 8, A = g1, B = g2, seeds = seeds)

#> Sparse part
#> 6 x 6 sparse Matrix of class "dgCMatrix"
#>                                                                        
#> [1,] 0.13328194 0.19611679 0.08885970 0.02731295 .          0.167949552
#> [2,] 0.39248191 0.01715787 0.12085055 0.03884697 0.10428926 0.306033039
#> [3,] 0.03521968 0.15047160 0.16857130 .          0.02480547 0.009796253
#> [4,] 0.10322582 .          0.17849120 0.26527377 0.32283142 0.097672783
#> [5,] .          .          0.07672746 0.14552725 0.20427538 0.044711954
#> [6,] 0.07518620 .          0.30437658 0.22398246 0.21431266 0.178571215
#> plus left factor
#> 6 x 1 sparse Matrix of class "dgCMatrix"
#>                
#> [1,] 0.00357089
#> [2,] 0.00357089
#> [3,] 0.00357089
#> [4,] 0.00357089
#> [5,] 0.00357089
#> [6,] 0.00357089
#> times right factor transpose
#> 6 x 1 sparse Matrix of class "dgCMatrix"
#>           
#> [1,] 0.125
#> [2,] 0.125
#> [3,] 0.125
#> [4,] 0.125
#> [5,] 0.125
#> [6,] 0.125

# FW graph matching with incorrect seeds to start at convex start
init_start(start = "convex", nns = 8, ns = 2, soft_seeds = ss, A = g1, B = g2, seeds = seeds)

#> Sparse part
#> 8 x 8 sparse Matrix of class "dgCMatrix"
#>                                                                       
#> [1,] 0.14437522 0.19444813 0.09762770 0.03226265 0.03150764 0.14688151
#> [2,] 0.39080085 0.02405738 0.10762906 0.05954192 0.09506118 0.32290961
#> [3,] .          0.16035117 0.15212285 0.04781938 0.01467472 0.04585005
#> [4,] 0.07483836 .          0.18914549 0.26234580 0.31719357 0.11097040
#> [5,] 0.04781938 .          0.05839288 0.14133562 0.22098769 .         
#> [6,] 0.09706474 .          0.31115161 0.19834137 0.22348101 0.15671751
#> [7,] 0.10605107 0.36659236 0.02405738 0.12538966 0.05979307 0.07857364
#> [8,] 0.13905038 0.25455096 0.04662928 0.11971986 0.02405738 0.12485353
#>                          
#> [1,] 0.12196164 0.2176918
#> [2,] .          .        
#> [3,] 0.23681286 0.3423690
#> [4,] 0.03226265 .        
#> [5,] 0.51822070 .        
#> [6,] .          .        
#> [7,] 0.05451394 0.1717851
#> [8,] 0.02298447 0.2549104
#> plus left factor
#> 8 x 1 sparse Matrix of class "dgCMatrix"
#>                
#> [1,] 0.01324374
#> [2,] .         
#> [3,] .         
#> [4,] 0.01324374
#> [5,] 0.01324374
#> [6,] 0.01324374
#> [7,] 0.01324374
#> [8,] 0.01324374
#> times right factor transpose
#> 8 x 1 sparse Matrix of class "dgCMatrix"
#>               
#> [1,] .        
#> [2,] .        
#> [3,] 0.1666667
#> [4,] 0.1666667
#> [5,] 0.1666667
#> [6,] 0.1666667
#> [7,] 0.1666667
#> [8,] 0.1666667

#