Initialization of the start matrix

Initialize the start matrix for graph matching iteration.

init_start(start, nns, ns = 0, soft_seeds = NULL, ...)

Arguments

start
A matrix or a character. Any nns-by-nns doubly stochastic matrix or start method like "bari", "convex" or "rds" to initialize the start matrix.

nns
An integer. Number of non-seeds.

ns
An integer. Number of seeds.

soft_seeds
A vector, a matrix or a data frame. If there is no error in the soft seeds, input can be a vector of soft seed indices in \(G_1\). Or if there is error in soft seeds, input should be in the form of a matrix or a data frame, with the first column being the indices of \(G_1\) and the second column being the corresponding indices of \(G_2\). Note that if there are seeds in graphs, seeds should be put before non-seeds.

...
Arguments passed to other start functions

Value

init_start returns a nns-by-nns doubly stochastic matrix as the start matrix in the graph matching iteration. If conduct a soft seeding graph matching, returns a nns-by-nns doubly stochastic matrix with 1's corresponding to the soft seeds and values at the other places are derived by different start method.

Examples

ss <- matrix(c(5, 4, 4, 3), nrow = 2)
# initialize start matrix without soft seeds
init_start(start = "bari", nns = 5)

#> Sparse part
#> 5 x 5 sparse Matrix of class "dsCMatrix"
#>               
#> [1,] . . . . .
#> [2,] . . . . .
#> [3,] . . . . .
#> [4,] . . . . .
#> [5,] . . . . .
#> plus left factor
#> 5 x 1 Matrix of class "dgeMatrix"
#>      [,1]
#> [1,]    1
#> [2,]    1
#> [3,]    1
#> [4,]    1
#> [5,]    1
#> times right factor transpose
#> 5 x 1 Matrix of class "dgeMatrix"
#>      [,1]
#> [1,]  0.2
#> [2,]  0.2
#> [3,]  0.2
#> [4,]  0.2
#> [5,]  0.2

init_start(start = "rds", nns = 3)

#> 3 x 3 Matrix of class "dgeMatrix"
#>           [,1]      [,2]      [,3]
#> [1,] 0.2671236 0.4163905 0.3164859
#> [2,] 0.5750521 0.1504994 0.2744485
#> [3,] 0.1578242 0.4331101 0.4090657

init_start(start = "rds_perm_bari", nns = 5)

#> Sparse part
#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>                                                       
#> [1,] .         .         .         .         0.4331351
#> [2,] 0.4331351 .         .         .         .        
#> [3,] .         0.4331351 .         .         .        
#> [4,] .         .         .         0.4331351 .        
#> [5,] .         .         0.4331351 .         .        
#> plus left factor
#> 5 x 1 Matrix of class "dgeMatrix"
#>           [,1]
#> [1,] 0.5668649
#> [2,] 0.5668649
#> [3,] 0.5668649
#> [4,] 0.5668649
#> [5,] 0.5668649
#> times right factor transpose
#> 5 x 1 Matrix of class "dgeMatrix"
#>      [,1]
#> [1,]  0.2
#> [2,]  0.2
#> [3,]  0.2
#> [4,]  0.2
#> [5,]  0.2

# initialize start matrix with soft seeds
init_start(start = "bari", nns = 5, ns = 3, soft_seeds = c(5, 7, 8))

#> Sparse part
#> 5 x 5 sparse Matrix of class "dsCMatrix"
#>               
#> [1,] . . . . .
#> [2,] . 1 . . .
#> [3,] . . . . .
#> [4,] . . . 1 .
#> [5,] . . . . 1
#> plus left factor
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>       
#> [1,] 1
#> [2,] .
#> [3,] 1
#> [4,] .
#> [5,] .
#> times right factor transpose
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>         
#> [1,] 0.5
#> [2,] .  
#> [3,] 0.5
#> [4,] .  
#> [5,] .  

init_start(start = "rds", nns = 5, soft_seeds = ss)

#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>                                         
#> [1,] 0.52249361 0.3813915 . . 0.09611489
#> [2,] 0.09292396 0.4388959 . . 0.46818013
#> [3,] 0.38458242 0.1797126 . . 0.43570498
#> [4,] .          .         1 . .         
#> [5,] .          .         . 1 .         

init_start(start = "rds_perm_bari", nns = 5, soft_seeds = ss)

#> Sparse part
#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>                                       
#> [1,] .         0.4612725 . . .        
#> [2,] .         .         . . 0.4612725
#> [3,] 0.4612725 .         . . .        
#> [4,] .         .         1 . .        
#> [5,] .         .         . 1 .        
#> plus left factor
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>               
#> [1,] 0.5387275
#> [2,] 0.5387275
#> [3,] 0.5387275
#> [4,] .        
#> [5,] .        
#> times right factor transpose
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>               
#> [1,] 0.3333333
#> [2,] 0.3333333
#> [3,] .        
#> [4,] .        
#> [5,] 0.3333333

# initialize start matrix for convex graph matching
cgnp_pair <- sample_correlated_gnp_pair(n = 10, corr =  0.3, p =  0.5)
g1 <- cgnp_pair$graph1
g2 <- cgnp_pair$graph2
seeds <- 1:10 <= 2
if (FALSE) {
 init_start(start = "convex", nns = 8, A = g1, B = g2, seeds = seeds)
 # FW graph matching with incorrect seeds to start at convex start
 init_start(start = "convex", nns = 8, ns = 2, soft_seeds = ss, A = g1, B = g2, seeds = seeds)
}