Find the largest common connected subgraph(LCCS)

Assume two aligned graphs, find the largest common connect subgraph of these two graphs, which is an induced connected subgraph of both graphs that has as many vertices as possible.

largest_common_cc(A, B, min_degree = 1)

Arguments

A
A matrix or an igraph object. Adjacency matrix of \(G_1\).

B
A matrix or an igraph object. Adjacency matrix of \(G_2\).

min_degree
A number. Defines the level of connectness of the obtained largest common connected subgraph. The induced subgraph is an graph with a minimum degree of vertices equal to min_degree.

Value

largest_common_cc returns the common largest connected subgraphs of two aligned graphs in the igraph object form and a logical vector indicating which vertices in the original graphs remain in the induced subgraph.

Examples

cgnp_pair <- sample_correlated_gnp_pair(n = 10, corr =  0.7, p =  0.2)
g1 <- cgnp_pair$graph1
g2 <- cgnp_pair$graph2
# put no constraint on the minimum degree of the common largest conncect subgraph
lccs1 <- largest_common_cc(g1, g2, min_degree = 1)
# induced subgraph
lccs1$g1

#> IGRAPH f74c80a U--- 10 11 -- Erdos renyi (gnp) graph
#> + attr: name (g/c), type (g/c), loops (g/l), p (g/n)
#> + edges from f74c80a:
#>  [1] 1-- 3 2-- 3 1-- 4 3-- 4 5-- 8 6-- 8 7-- 8 2-- 9 4--10 6--10 9--10

lccs1$g2

#> IGRAPH 1d1adbb U--- 10 12 -- Erdos renyi (gnp) graph
#> + attr: name_1 (g/c), name_2 (g/c), type_1 (g/c), type_2 (g/c), loops_1
#> | (g/l), loops_2 (g/l), p_1 (g/n), p_2 (g/n), name (g/c), type (g/c),
#> | loops (g/l), p (g/n)
#> + edges from 1d1adbb:
#>  [1] 9--10 7-- 9 7-- 8 6--10 6-- 9 6-- 8 5-- 6 4-- 7 3-- 5 3-- 4 2-- 9 1-- 3

# label of vertices of the induced subgraph in the original graph
igraph::V(g1)[lccs1$keep]

#> + 10/10 vertices, from f74c80a:
#>  [1]  1  2  3  4  5  6  7  8  9 10

# obtain a common largest connect subgraph with each vertex having a minimum degree of 3
lccs3 <- largest_common_cc(g1, g2, min_degree = 3)