Measure functions

Measures for computing the goodness of matching for each vertex.

row_cor(g1, g2)

row_diff(g1, g2)

row_perm_stat(g1, g2, exact = TRUE)

Arguments

g1
A matrix or an igraph object. Adjacency matrix of \(G_1\).

g2
A matrix or an igraph object. Adjacency matrix of \(G_2\) after adjusting rows and columns according to the correlation of matching between two graphs.

exact
A logical. If g1 and g2 are binary, then set exact=TRUE, if g1 and g2 are weighted graphs, then set exact=FALSE.

Value

row_cor returns a vector, each element is 1 minus the row correlation value for the corresponding vertex.

row_diff returns a vector, each element is the row difference value for the corresponding vertex.

row_perm_stat returns a vector, each element is the row permutation statistics value for the corresponding vertex.

Examples

cgnp_pair <- sample_correlated_gnp_pair(n = 50, corr =  0.3,p =  0.5)
g1 <- cgnp_pair$graph1
g2 <- cgnp_pair$graph2
match <- graph_match_FW(g1, g2)
g2m <- g2[match$corr$corr_B, match$corr$corr_B]
g1 <- g1[]
row_cor(g1, g2m)

#>  [1] 0.4830918 0.5600000 0.6538719 0.6863354 0.4409622 0.7622483 0.5996796
#>  [8] 0.5458174 0.7179487 0.8442707 0.5658330 0.6016053 0.6421865 0.5951795
#> [15] 0.4078435 0.6404538 0.6669334 0.5165062 0.6887336 0.4845283 0.5951795
#> [22] 0.7313978 0.6734014 0.3178135 0.6068215 0.4845283 0.4025765 0.5722379
#> [29] 0.6076389 0.5570540 0.6863354 0.6471377 0.4387257 0.6776708 0.5636071
#> [36] 0.5212838 0.6819268 0.4783280 0.4577922 0.4845283 0.6419280 0.6016053
#> [43] 0.6471377 0.7305195 0.6404538 0.7163457 0.4845283 0.7190997 0.5212838
#> [50] 0.7808578

row_diff(g1, g2m)

#>  [1] 12 14 17 17 11 19 15 13 18 21 14 15 17 15 10 16 17 13 17 12 15 20 17  8 15
#> [26] 12 10 15 14 13 17 16 11 17 14 13 17 12 12 12 16 15 16 18 16 18 12 18 13 19

row_perm_stat(g1, g2m)

#>  [1] -3.5709604 -2.9934050 -2.1535907 -2.1335262 -3.8521230 -1.5702803
#>  [7] -2.7112016 -3.2778606 -1.8499960 -0.9998196 -2.9900235 -2.7044863
#> [13] -2.1340101 -2.7044863 -4.1470343 -2.4259713 -2.1323650 -3.2786354
#> [19] -2.1319191 -3.5709604 -2.7044863 -1.2712147 -2.1377140 -4.7122143
#> [25] -2.7312805 -3.5709604 -4.1318864 -2.7038257 -2.9967699 -3.2806123
#> [31] -2.1335262 -2.4297064 -3.8482191 -2.1377140 -2.9899663 -3.2754463
#> [37] -2.1407410 -3.5609264 -3.5610169 -3.5709604 -2.4206230 -2.7044863
#> [43] -2.4178931 -1.8466346 -2.4452153 -1.8524837 -3.5822425 -1.8524837
#> [49] -3.2856469 -1.5777141