Measure functions

Measures for computing the goodness of matching for each vertex.

row_cor(g1, g2)

row_diff(g1, g2)

row_perm_stat(g1, g2, exact = TRUE, ...)

Arguments

g1
A matrix or an igraph object. Adjacency matrix of \(G_1\).

g2
A matrix or an igraph object. Adjacency matrix of \(G_2\) after adjusting rows and columns according to the correlation of matching between two graphs.

exact
A logical. If g1 and g2 are binary, then set exact=TRUE, if g1 and g2 are weighted graphs, then set exact=FALSE.

Value

row_cor returns a vector, each element is 1 minus the row correlation value for the corresponding vertex.

row_diff returns a vector, each element is the row difference value for the corresponding vertex.

row_perm_stat returns a vector, each element is the row permutation statistics value for the corresponding vertex.

Examples

cgnp_pair <- sample_correlated_gnp_pair(n = 50, corr =  0.3,p =  0.5)
g1 <- cgnp_pair$graph1
g2 <- cgnp_pair$graph2
match <- graph_match_FW(g1, g2)
g2m <- g2[match$corr$corr_B, match$corr$corr_B]
g1 <- g1[]
row_cor(g1, g2m)

#>  [1] 0.6755011 0.6016053 0.5651372 0.7163457 0.5516036 0.5951795 0.6421865
#>  [8] 0.5970885 0.4783280 0.7612489 0.5996796 0.5723314 0.7284424 0.6887336
#> [15] 0.4395515 0.4908249 0.4345365 0.6016053 0.6734014 0.7922705 0.6353016
#> [22] 0.5970885 0.6797437 0.5101021 0.6690753 0.4540792 0.8097470 0.7684729
#> [29] 0.7225749 0.6568144 0.6404538 0.6478107 0.6863354 0.6776708 0.4933290
#> [36] 0.4926108 0.5398155 0.6493506 0.4806344 0.8856683 0.4395515 0.6419280
#> [43] 0.4025765 0.6559688 0.4900874 0.5636071 0.8097470 0.4102322 0.5833333
#> [50] 0.8254050

row_diff(g1, g2m)

#>  [1] 17 15 14 18 14 15 17 15 12 18 15 14 18 17 11 11 11 15 17 20 16 15 17 13 17
#> [26] 11 20 20 18 16 16 16 17 17 13 12 14 16 12 22 11 16 10 17 12 14 20 11 15 20

row_perm_stat(g1, g2m)

#>  [1] -2.1117383 -2.7234444 -2.9179212 -1.9057376 -2.9828890 -2.6837404
#>  [7] -2.2325882 -2.6706478 -3.5947816 -1.9057376 -2.7234444 -2.9307628
#> [13] -1.7994600 -2.1117383 -3.8540496 -3.7637336 -3.8807673 -2.6837404
#> [19] -2.1381816 -1.3068962 -2.4241596 -2.7101376 -2.1249320 -3.2168560
#> [25] -2.0855148 -3.8018438 -1.2273503 -1.3336895 -1.8389463 -2.3715697
#> [31] -2.4109176 -2.3977393 -2.1117383 -2.1782773 -3.1536878 -3.5157890
#> [37] -2.9697414 -2.3846237 -3.6082825 -0.6817331 -3.8540496 -2.4374664
#> [43] -4.1667530 -2.2053076 -3.6082825 -2.9697414 -1.3068962 -3.7890426
#> [49] -2.6189605 -1.3202742