Measure functions

Measures for computing the goodness of matching for each vertex.

row_cor(g1, g2)

row_diff(g1, g2)

row_perm_stat(g1, g2, exact = TRUE)

Arguments

g1
A matrix or an igraph object. Adjacency matrix of \(G_1\).

g2
A matrix or an igraph object. Adjacency matrix of \(G_2\) after adjusting rows and columns according to the correlation of matching between two graphs.

exact
A logical. If g1 and g2 are binary, then set exact=TRUE, if g1 and g2 are weighted graphs, then set exact=FALSE.

Value

row_cor returns a vector, each element is 1 minus the row correlation value for the corresponding vertex.

row_diff returns a vector, each element is the row difference value for the corresponding vertex.

row_perm_stat returns a vector, each element is the row permutation statistics value for the corresponding vertex.

Examples

cgnp_pair <- sample_correlated_gnp_pair(n = 50, corr =  0.3,p =  0.5)
g1 <- cgnp_pair$graph1
g2 <- cgnp_pair$graph2
match <- graph_match_FW(g1, g2)
g2m <- g2[match$corr$corr_B, match$corr$corr_B]
g1 <- g1[]
row_cor(g1, g2m)

#>  [1] 0.6703657 0.4387257 0.4806344 0.5576923 0.5735853 0.5612812 0.5612812
#>  [8] 0.4395515 0.5367944 0.6068215 0.5860920 0.5212838 0.8361441 0.7225749
#> [15] 0.4540792 0.5600000 0.6322504 0.6669334 0.5550750 0.5052506 0.4057056
#> [22] 0.8312933 0.4216483 0.6419280 0.4900874 0.7225749 0.4244092 0.6409719
#> [29] 0.5212838 0.6114222 0.7190997 0.7190997 0.7624732 0.5860920 0.6068215
#> [36] 0.8442707 0.7980769 0.7022250 0.6388424 0.5052506 0.4702093 0.5658330
#> [43] 0.5398155 0.5297882 0.7993569 0.7922705 0.5304982 0.6819268 0.4089699
#> [50] 0.4032236

row_diff(g1, g2m)

#>  [1] 19 11 12 14 14 14 14 11 13 15 15 13 21 18 11 14 16 17 15 13 10 21 10 16 12
#> [26] 18 11 16 13 15 18 18 19 15 15 21 20 17 16 13 12 14 14 13 20 20 13 17 11 10

row_perm_stat(g1, g2m)

#>  [1] -1.6913163 -3.8660861 -3.5105020 -2.9708336 -3.1575776 -3.0528563
#>  [7] -2.9708336 -3.8462660 -3.2099452 -2.6220684 -2.7668335 -3.2396468
#> [13] -0.9514638 -1.9087649 -3.9629531 -2.9893380 -2.3648836 -2.1947877
#> [19] -2.7668335 -3.3160895 -4.3109231 -1.0719715 -4.0436160 -2.4808106
#> [25] -3.6760372 -1.8277567 -3.8876280 -2.2759713 -3.2396468 -2.7903990
#> [31] -1.8467673 -1.8872434 -1.5809447 -2.7668335 -2.6676010 -0.9704490
#> [37] -1.2953021 -2.0622815 -2.4180527 -3.3388792 -3.6018588 -3.0768036
#> [43] -3.0303203 -3.3632082 -1.3157049 -1.3157049 -3.2240974 -2.1135259
#> [49] -3.9109249 -4.0977154