Measure functions

Measures for computing the goodness of matching for each vertex.

row_cor(g1, g2)

row_diff(g1, g2)

row_perm_stat(g1, g2, exact = TRUE)

Arguments

g1
A matrix or an igraph object. Adjacency matrix of \(G_1\).

g2
A matrix or an igraph object. Adjacency matrix of \(G_2\) after adjusting rows and columns according to the correlation of matching between two graphs.

exact
A logical. If g1 and g2 are binary, then set exact=TRUE, if g1 and g2 are weighted graphs, then set exact=FALSE.

Value

row_cor returns a vector, each element is 1 minus the row correlation value for the corresponding vertex.

row_diff returns a vector, each element is the row difference value for the corresponding vertex.

row_perm_stat returns a vector, each element is the row permutation statistics value for the corresponding vertex.

Examples

cgnp_pair <- sample_correlated_gnp_pair(n = 50, corr =  0.3,p =  0.5)
g1 <- cgnp_pair$graph1
g2 <- cgnp_pair$graph2
match <- graph_match_FW(g1, g2)
g2m <- g2[match$corr$corr_B, match$corr$corr_B]
g1 <- g1[]
row_cor(g1, g2m)

#>  [1] 0.7284424 0.4284524 0.6211202 0.5585852 0.4709567 0.6819268 0.6378205
#>  [8] 0.7618552 0.5196155 0.4257483 0.5297882 0.6441224 0.7812500 0.6016053
#> [15] 0.7128973 0.4775641 0.5516036 0.4284524 0.5229850 0.6478107 0.5398155
#> [22] 0.5304982 0.7115700 0.7246377 0.5000000 0.7117552 0.5246525 0.6776708
#> [29] 0.8254050 0.5651372 0.5951795 0.6441224 0.6930562 0.6353016 0.5507246
#> [36] 0.5148580 0.7622483 0.5507246 0.5860920 0.6016053 0.6016053 0.5636071
#> [43] 0.5542575 0.4870130 0.7706458 0.6797437 0.6568144 0.7973898 0.4845283
#> [50] 0.6930562

row_diff(g1, g2m)

#>  [1] 18 11 16 14 12 17 16 18 13 11 13 16 18 15 18 12 14 11 14 16 14 13 18 18 12
#> [26] 18 13 17 20 14 15 16 17 16 14 13 19 14 15 15 15 14 14 12 19 17 16 20 12 17

row_perm_stat(g1, g2m)

#>  [1] -1.859079 -3.849096 -2.439321 -2.989966 -3.566893 -2.140741 -2.425971
#>  [8] -1.845434 -3.276563 -3.852123 -3.276089 -2.419006 -1.846423 -2.704486
#> [15] -1.863194 -3.566893 -2.989556 -3.849096 -2.991375 -2.429706 -2.989966
#> [22] -3.289109 -1.852484 -1.859079 -3.589476 -1.859079 -3.275446 -2.132365
#> [29] -1.271250 -2.993620 -2.704486 -2.429706 -2.131828 -2.422944 -2.989966
#> [36] -3.275446 -1.570280 -2.989966 -2.703826 -2.715020 -2.705936 -2.989966
#> [43] -2.990024 -3.561017 -1.587020 -2.137714 -2.439321 -1.282023 -3.570960
#> [50] -2.148674