Measure functions

Measures for computing the goodness of matching for each vertex.

row_cor(g1, g2)

row_diff(g1, g2)

row_perm_stat(g1, g2, exact = TRUE)

Arguments

g1
A matrix or an igraph object. Adjacency matrix of \(G_1\).

g2
A matrix or an igraph object. Adjacency matrix of \(G_2\) after adjusting rows and columns according to the correlation of matching between two graphs.

exact
A logical. If g1 and g2 are binary, then set exact=TRUE, if g1 and g2 are weighted graphs, then set exact=FALSE.

Value

row_cor returns a vector, each element is 1 minus the row correlation value for the corresponding vertex.

row_diff returns a vector, each element is the row difference value for the corresponding vertex.

row_perm_stat returns a vector, each element is the row permutation statistics value for the corresponding vertex.

Examples

cgnp_pair <- sample_correlated_gnp_pair(n = 50, corr =  0.3,p =  0.5)
g1 <- cgnp_pair$graph1
g2 <- cgnp_pair$graph2
match <- graph_match_FW(g1, g2)

#> Error in .local(x, i, j, ..., value): RHS 'value' (class dsparseVector) matches 'ANY', but must match matrix class splrMatrix

g2m <- g2[match$corr$corr_B, match$corr$corr_B]

#> Error in h(simpleError(msg, call)): error in evaluating the argument 'i' in selecting a method for function '[': object of type 'closure' is not subsettable

g1 <- g1[]
row_cor(g1, g2m)

#> Error in row_cor(g1, g2m): object 'g2m' not found

row_diff(g1, g2m)

#> Error in row_diff(g1, g2m): object 'g2m' not found

row_perm_stat(g1, g2m)

#> Error in row_perm_stat(g1, g2m): object 'g2m' not found