Measure functions

Measures for computing the goodness of matching for each vertex.

row_cor(g1, g2)

row_diff(g1, g2)

row_perm_stat(g1, g2, exact = TRUE)

Arguments

g1
A matrix or an igraph object. Adjacency matrix of \(G_1\).

g2
A matrix or an igraph object. Adjacency matrix of \(G_2\) after adjusting rows and columns according to the correlation of matching between two graphs.

exact
A logical. If g1 and g2 are binary, then set exact=TRUE, if g1 and g2 are weighted graphs, then set exact=FALSE.

Value

row_cor returns a vector, each element is 1 minus the row correlation value for the corresponding vertex.

row_diff returns a vector, each element is the row difference value for the corresponding vertex.

row_perm_stat returns a vector, each element is the row permutation statistics value for the corresponding vertex.

Examples

cgnp_pair <- sample_correlated_gnp_pair(n = 50, corr =  0.3,p =  0.5)
g1 <- cgnp_pair$graph1
g2 <- cgnp_pair$graph2
match <- graph_match_FW(g1, g2)
g2m <- g2[match$corr$corr_B, match$corr$corr_B]
g1 <- g1[]
row_cor(g1, g2m)

#>  [1] 0.4800000 0.6441224 0.6211202 0.4395515 0.6755011 0.7128973 0.4974258
#>  [8] 0.6054939 0.7746902 0.6114222 0.5196155 0.7225749 0.5842492 0.5434484
#> [15] 0.6441224 0.6410256 0.7368193 0.3111252 0.8131578 0.5879663 0.4345365
#> [22] 0.5068573 0.3999875 0.5612812 0.5516036 0.5715878 0.4852143 0.5658330
#> [29] 0.5636071 0.6410256 0.6669334 0.6051003 0.5843961 0.7163457 0.6054019
#> [36] 0.5068573 0.6538719 0.7875409 0.7684729 0.5585852 0.5196155 0.6797437
#> [43] 0.7368193 0.5423957 0.5254539 0.4732134 0.6545754 0.6690753 0.6299253
#> [50] 0.5951795

row_diff(g1, g2m)

#>  [1] 12 16 16 11 17 18 12 15 19 15 13 18 16 14 16 16 18  8 20 15 11 13 10 14 14
#> [26] 14 12 14 14 16 17 14 14 18 15 13 17 20 20 14 13 17 18 14 14 12 16 17 15 15

row_perm_stat(g1, g2m)

#>  [1] -3.569459 -2.435408 -2.417722 -3.855193 -2.136376 -1.841854 -3.587370
#>  [8] -2.703463 -1.556106 -2.694760 -3.279322 -1.859487 -2.400363 -2.993585
#> [15] -2.417722 -2.422112 -1.833131 -4.712396 -1.274756 -2.707849 -3.864090
#> [22] -3.301622 -4.127843 -3.011364 -2.984849 -2.959242 -3.610455 -3.015873
#> [29] -3.006881 -2.430955 -2.127603 -2.971933 -2.955061 -1.855055 -2.721144
#> [36] -3.266277 -2.127603 -1.288019 -1.265968 -2.997991 -3.279322 -2.140789
#> [43] -1.868398 -2.976213 -2.971933 -3.569459 -2.448891 -2.123243 -2.739199
#> [50] -2.703463