Measure functions

Measures for computing the goodness of matching for each vertex.

row_cor(g1, g2)

row_diff(g1, g2)

row_perm_stat(g1, g2, exact = TRUE)

Arguments

g1
A matrix or an igraph object. Adjacency matrix of \(G_1\).

g2
A matrix or an igraph object. Adjacency matrix of \(G_2\) after adjusting rows and columns according to the correlation of matching between two graphs.

exact
A logical. If g1 and g2 are binary, then set exact=TRUE, if g1 and g2 are weighted graphs, then set exact=FALSE.

Value

row_cor returns a vector, each element is 1 minus the row correlation value for the corresponding vertex.

row_diff returns a vector, each element is the row difference value for the corresponding vertex.

row_perm_stat returns a vector, each element is the row permutation statistics value for the corresponding vertex.

Examples

cgnp_pair <- sample_correlated_gnp_pair(n = 50, corr =  0.3,p =  0.5)
g1 <- cgnp_pair$graph1
g2 <- cgnp_pair$graph2
match <- graph_match_FW(g1, g2)
g2m <- g2[match$corr$corr_B, match$corr$corr_B]
g1 <- g1[]
row_cor(g1, g2m)

#>  [1] 0.6776708 0.5297882 0.4952556 0.5318327 0.7582531 0.7304842 0.4807692
#>  [8] 0.4438111 0.6471377 0.4438111 0.6698075 0.4775641 0.5879663 0.5996796
#> [15] 0.5212838 0.6575252 0.5970885 0.5587387 0.5542575 0.8104012 0.7368193
#> [22] 0.5367944 0.5951795 0.4488243 0.5297882 0.7128973 0.5951795 0.5367944
#> [29] 0.5212838 0.8361441 0.5570540 0.7190997 0.6797437 0.2821283 0.5612812
#> [36] 0.6322504 0.5860920 0.5722379 0.5879663 0.6054939 0.5842654 0.5970885
#> [43] 0.6291615 0.5658330 0.4702093 0.4807692 0.6908009 0.6755011 0.6114222
#> [50] 0.7999183

row_diff(g1, g2m)

#>  [1] 17 13 13 13 19 18 12 11 16 11 16 12 15 15 13 16 15 14 14 19 18 13 15 11 13
#> [26] 18 15 13 13 21 13 18 17  7 14 16 15 15 15 15 14 15 16 14 12 12 18 17 15 18

row_perm_stat(g1, g2m)

#>  [1] -2.124456 -3.302064 -3.293680 -3.302064 -1.571267 -1.816127 -3.579854
#>  [8] -3.847156 -2.402571 -3.839044 -2.447778 -3.579854 -2.709941 -2.715960
#> [15] -3.275218 -2.444158 -2.715960 -2.964679 -2.973557 -1.515325 -1.824060
#> [22] -3.249834 -2.703281 -3.874002 -3.259050 -1.838691 -2.721334 -3.249834
#> [29] -3.298267 -1.004293 -3.229199 -1.867741 -2.143614 -4.982103 -2.996046
#> [36] -2.444158 -2.695985 -2.695985 -2.733543 -2.695985 -3.019308 -2.730127
#> [43] -2.429787 -2.973557 -3.584372 -3.568256 -1.845374 -2.148987 -2.688064
#> [50] -1.885629