Sample correlated G(n,p) random graphs

Sample a pair of correlated G(n,p) random graphs with correlation between two graphs being rho and edge probability being p.

sample_correlated_gnp_pair(n, corr, p, permutation = 1:n, ...)

sample_correlated_gnp_pair_w_junk(
  n,
  corr,
  p,
  ncore = n,
  permutation = 1:n,
  ...
)

Arguments

n
An integer. Number of total vertices for the sampled graphs.

corr
A number. The target Pearson correlation between the adjacency matrices of the generated graphs. It must be in open (0,1) interval.

p
A number. Edge probability between two vertices. It must be in open (0,1) interval.

permutation
A numeric vector,permute second graph.

...
Passed to sample_correlated_gnp_pair and sample_correlated_gnp_pair_w_junk.

ncore
An integer. Number of core vertices.

Value

sample_correlated_gnp_pair returns a list of two igraph object, named graph1 and graph2, which are two graphs whose adjacency matrix entries correlated with rho.

sample_correlated_gnp_pair_w_junk returns a list of two igraph object, named graph1 and graph2, which are two graphs whose adjacency matrix entries correlated with rho and with first ncore vertices being core vertices and the rest being junk vertices.

Examples

sample_correlated_gnp_pair(50, 0.3, 0.5)

#> $graph1
#> IGRAPH b7c453b U--- 50 637 -- Erdos renyi (gnp) graph
#> + attr: name (g/c), type (g/c), loops (g/l), p (g/n)
#> + edges from b7c453b:
#>  [1]  1-- 2  1-- 3  2-- 5  3-- 5  4-- 5  2-- 6  3-- 6  4-- 6  5-- 6  1-- 7
#> [11]  2-- 7  3-- 7  4-- 7  1-- 8  3-- 8  6-- 8  7-- 8  1-- 9  3-- 9  4-- 9
#> [21]  6-- 9  4--10  2--11  3--11  4--11  5--11  6--11  7--11  3--12  5--12
#> [31]  9--12 11--12  1--13  2--13  3--13  4--13  6--13  7--13 10--13 12--13
#> [41]  1--14  3--14  5--14  7--14  8--14  9--14 10--14 12--14 13--14  1--15
#> [51]  2--15  3--15  4--15  5--15  8--15  9--15 11--15 13--15  2--16  4--16
#> [61]  5--16  8--16 10--16 11--16 12--16 13--16 14--16  3--17  4--17  7--17
#> [71]  9--17 11--17 12--17 13--17  1--18  4--18 11--18 12--18 13--18 16--18
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH a8f9923 U--- 50 597 -- Erdos renyi (gnp) graph
#> + attr: name_1 (g/c), name_2 (g/c), type_1 (g/c), type_2 (g/c), loops_1
#> | (g/l), loops_2 (g/l), p_1 (g/n), p_2 (g/n), name (g/c), type (g/c),
#> | loops (g/l), p (g/n)
#> + edges from a8f9923:
#>  [1] 48--50 47--49 46--50 46--49 46--48 46--47 45--50 45--48 44--46 43--50
#> [11] 43--48 43--45 42--47 42--44 41--46 41--42 40--50 40--47 40--45 40--44
#> [21] 39--49 39--48 39--46 39--45 39--44 39--43 39--42 39--40 38--49 38--42
#> [31] 38--41 38--40 37--50 37--49 37--46 37--45 37--44 37--40 37--39 37--38
#> [41] 36--49 36--48 36--44 36--41 36--37 35--50 35--49 35--48 35--46 35--43
#> [51] 35--41 35--38 34--50 34--49 34--48 34--44 34--43 34--42 34--38 34--35
#> + ... omitted several edges
#> 

sample_correlated_gnp_pair_w_junk(50, 0.3, 0.5, 40)

#> $graph1
#> IGRAPH 0850b03 U--- 50 650 -- 
#> + attr: name_1 (g/c), name_2 (g/c), loops_1 (g/l), loops_2 (g/l), type
#> | (g/c), p (g/n)
#> + edges from 0850b03:
#>  [1] 49--50 48--50 48--49 47--48 45--50 45--49 45--46 44--50 44--49 44--48
#> [11] 44--45 43--48 43--45 43--44 42--50 42--49 42--46 42--45 42--43 41--50
#> [21] 41--46 41--44 41--43 40--49 40--48 40--47 40--45 40--44 40--43 40--42
#> [31] 40--41 39--50 39--48 39--47 39--44 39--43 38--50 38--49 38--48 38--44
#> [41] 38--42 38--41 37--50 37--46 37--45 37--43 37--39 37--38 36--49 36--48
#> [51] 36--47 36--46 36--45 36--44 36--43 36--42 36--41 36--37 35--50 35--49
#> [61] 35--48 35--46 35--45 35--44 35--42 35--41 35--40 35--39 35--38 35--36
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH a680656 U--- 50 627 -- 
#> + attr: name_1 (g/c), name_2 (g/c), loops_1 (g/l), loops_2 (g/l),
#> | type_1 (g/c), type_2 (g/c), p_1 (g/n), p_2 (g/n), type (g/c), p (g/n)
#> + edges from a680656:
#>  [1] 48--49 47--50 46--50 46--48 46--47 45--48 44--49 44--47 44--46 43--47
#> [11] 43--46 43--45 42--49 42--48 42--47 42--44 41--48 40--49 40--46 40--43
#> [21] 40--41 39--50 39--49 39--47 39--46 39--45 39--41 38--41 38--40 38--39
#> [31] 37--49 37--47 37--46 37--43 37--42 37--39 37--38 36--50 36--49 36--47
#> [41] 36--44 36--43 36--42 36--40 36--37 35--48 35--44 35--42 35--40 35--39
#> [51] 35--38 35--36 34--50 34--48 34--47 34--45 34--41 34--39 34--38 33--48
#> [61] 33--46 33--45 33--41 33--39 33--38 33--36 32--49 32--47 32--46 32--45
#> + ... omitted several edges
#>