Sample correlated G(n,p) random graphs

Sample a pair of correlated G(n,p) random graphs with correlation between two graphs being rho and edge probability being p.

sample_correlated_gnp_pair(n, corr, p, permutation = 1:n, ...)

sample_correlated_gnp_pair_w_junk(
  n,
  corr,
  p,
  ncore = n,
  permutation = 1:n,
  ...
)

Arguments

n
An integer. Number of total vertices for the sampled graphs.

corr
A number. The target Pearson correlation between the adjacency matrices of the generated graphs. It must be in open (0,1) interval.

p
A number. Edge probability between two vertices. It must be in open (0,1) interval.

permutation
A numeric vector,permute second graph.

...
Passed to sample_correlated_gnp_pair and sample_correlated_gnp_pair_w_junk.

ncore
An integer. Number of core vertices.

Value

sample_correlated_gnp_pair returns a list of two igraph object, named graph1 and graph2, which are two graphs whose adjacency matrix entries correlated with rho.

sample_correlated_gnp_pair_w_junk returns a list of two igraph object, named graph1 and graph2, which are two graphs whose adjacency matrix entries correlated with rho and with first ncore vertices being core vertices and the rest being junk vertices.

Examples

sample_correlated_gnp_pair(50, 0.3, 0.5)

#> $graph1
#> IGRAPH 2923480 U--- 50 628 -- Erdos renyi (gnp) graph
#> + attr: name (g/c), type (g/c), loops (g/l), p (g/n)
#> + edges from 2923480:
#>  [1]  1-- 2  2-- 3  1-- 4  2-- 5  1-- 6  2-- 6  3-- 6  4-- 6  5-- 6  1-- 7
#> [11]  2-- 7  3-- 7  4-- 7  5-- 7  2-- 8  4-- 8  5-- 8  6-- 8  7-- 8  3-- 9
#> [21]  5-- 9  6-- 9  8-- 9  3--10  5--10  6--10  8--10  9--10  4--11  9--11
#> [31] 10--11  1--12  4--12  5--12  6--12  7--12  8--12  2--13  4--13  8--13
#> [41] 12--13  3--14  4--14  8--14 10--14 13--14  1--15  4--15  5--15  6--15
#> [51]  7--15  8--15 10--15 13--15  1--16  3--16  4--16  9--16 12--16 13--16
#> [61] 14--16  1--17  2--17  3--17  7--17  9--17 12--17 13--17 14--17 15--17
#> [71] 16--17  6--18  8--18  9--18 10--18 11--18 13--18 15--18  2--19  3--19
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH 1a21fa9 U--- 50 591 -- Erdos renyi (gnp) graph
#> + attr: name_1 (g/c), name_2 (g/c), type_1 (g/c), type_2 (g/c), loops_1
#> | (g/l), loops_2 (g/l), p_1 (g/n), p_2 (g/n), name (g/c), type (g/c),
#> | loops (g/l), p (g/n)
#> + edges from 1a21fa9:
#>  [1] 48--49 47--50 47--49 47--48 46--50 46--48 46--47 45--50 45--48 44--48
#> [11] 44--47 44--46 43--49 43--47 43--44 42--50 42--49 42--48 42--47 42--46
#> [21] 42--44 41--49 41--47 41--45 41--44 41--43 40--47 40--43 40--42 39--50
#> [31] 39--49 39--48 39--46 39--42 39--40 38--48 38--46 38--45 38--44 38--42
#> [41] 38--39 37--50 37--48 37--47 37--46 37--44 37--43 37--42 37--41 37--39
#> [51] 36--50 36--45 36--42 36--39 35--49 35--47 35--42 35--41 35--40 35--38
#> + ... omitted several edges
#> 

sample_correlated_gnp_pair_w_junk(50, 0.3, 0.5, 40)

#> $graph1
#> IGRAPH d8e8193 U--- 50 611 -- 
#> + attr: name_1 (g/c), name_2 (g/c), loops_1 (g/l), loops_2 (g/l), type
#> | (g/c), p (g/n)
#> + edges from d8e8193:
#>  [1] 49--50 48--49 46--50 46--47 45--49 45--48 45--46 44--50 43--49 42--50
#> [11] 42--49 42--48 42--46 41--50 41--49 41--47 41--46 41--45 41--44 41--43
#> [21] 41--42 40--50 40--49 40--48 40--46 40--41 39--48 39--47 39--44 39--40
#> [31] 38--50 38--48 38--46 38--45 38--44 38--42 37--50 37--48 37--46 37--45
#> [41] 37--44 37--42 37--40 37--39 37--38 36--50 36--48 36--47 36--46 36--45
#> [51] 36--43 36--41 36--40 36--38 36--37 35--49 35--45 35--44 35--43 35--39
#> [61] 35--37 35--36 34--50 34--47 34--46 34--43 34--41 34--39 34--37 34--36
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH 4d29043 U--- 50 639 -- 
#> + attr: name_1 (g/c), name_2 (g/c), loops_1 (g/l), loops_2 (g/l),
#> | type_1 (g/c), type_2 (g/c), p_1 (g/n), p_2 (g/n), type (g/c), p (g/n)
#> + edges from 4d29043:
#>  [1] 49--50 47--50 46--50 46--49 46--48 45--49 45--48 45--47 44--50 44--49
#> [11] 44--48 44--47 44--46 44--45 43--50 43--47 43--46 43--45 42--50 42--45
#> [21] 42--43 41--50 41--49 41--45 41--43 41--42 40--50 40--48 40--46 40--45
#> [31] 40--44 40--43 40--42 40--41 39--50 39--48 39--47 39--43 39--42 39--41
#> [41] 39--40 38--50 38--49 38--47 38--46 38--42 38--39 37--48 37--47 37--45
#> [51] 37--44 37--39 37--38 36--50 36--49 36--44 36--43 36--42 36--39 36--38
#> [61] 35--50 35--47 35--45 35--44 35--43 35--42 35--40 35--39 35--38 35--37
#> + ... omitted several edges
#>