Sample correlated G(n,p) random graphs

Sample a pair of correlated G(n,p) random graphs with correlation between two graphs being rho and edge probability being p.

sample_correlated_gnp_pair(n, corr, p, permutation = 1:n, ...)

sample_correlated_gnp_pair_w_junk(
  n,
  corr,
  p,
  ncore = n,
  permutation = 1:n,
  ...
)

Arguments

n
An integer. Number of total vertices for the sampled graphs.

corr
A number. The target Pearson correlation between the adjacency matrices of the generated graphs. It must be in open (0,1) interval.

p
A number. Edge probability between two vertices. It must be in open (0,1) interval.

permutation
A numeric vector,permute second graph.

...
Passed to sample_correlated_gnp_pair and sample_correlated_gnp_pair_w_junk.

ncore
An integer. Number of core vertices.

Value

sample_correlated_gnp_pair returns a list of two igraph object, named graph1 and graph2, which are two graphs whose adjacency matrix entries correlated with rho.

sample_correlated_gnp_pair_w_junk returns a list of two igraph object, named graph1 and graph2, which are two graphs whose adjacency matrix entries correlated with rho and with first ncore vertices being core vertices and the rest being junk vertices.

Examples

sample_correlated_gnp_pair(50, 0.3, 0.5)

#> $graph1
#> IGRAPH 2a59738 U--- 50 626 -- Erdos renyi (gnp) graph
#> + attr: name (g/c), type (g/c), loops (g/l), p (g/n)
#> + edges from 2a59738:
#>  [1]  1-- 2  2-- 4  3-- 4  3-- 5  1-- 6  3-- 6  4-- 6  5-- 6  4-- 7  1-- 8
#> [11]  1-- 9  2-- 9  3-- 9  6-- 9  7-- 9  4--10  6--10  7--10  8--10  3--11
#> [21]  4--11  8--11  1--12  2--12  5--12  7--12  8--12 11--12  2--13  6--13
#> [31]  9--13 10--13 11--13 12--13  1--14  3--14  4--14  5--14  8--14  9--14
#> [41] 10--14 12--14  2--15  6--15  8--15 13--15 14--15  2--16  5--16  6--16
#> [51]  8--16 11--16 12--16 13--16 14--16  3--17  4--17 10--17 11--17 13--17
#> [61] 14--17 15--17  3--18  4--18  5--18  6--18  8--18  9--18 11--18 14--18
#> [71] 16--18  3--19  4--19  5--19  6--19  7--19  8--19  9--19 10--19 13--19
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH 649563d U--- 50 620 -- Erdos renyi (gnp) graph
#> + attr: name_1 (g/c), name_2 (g/c), type_1 (g/c), type_2 (g/c), loops_1
#> | (g/l), loops_2 (g/l), p_1 (g/n), p_2 (g/n), name (g/c), type (g/c),
#> | loops (g/l), p (g/n)
#> + edges from 649563d:
#>  [1] 49--50 48--49 47--49 46--49 46--48 45--47 45--46 44--50 44--49 44--46
#> [11] 44--45 43--49 43--48 43--46 43--45 42--46 42--45 41--50 41--46 41--45
#> [21] 41--43 40--50 40--49 40--46 40--45 40--43 40--41 39--50 39--49 39--47
#> [31] 39--46 39--44 39--42 39--41 38--49 38--48 38--47 38--44 38--42 38--41
#> [41] 38--40 38--39 37--48 37--47 37--45 37--43 37--42 37--39 36--49 36--48
#> [51] 36--47 36--45 36--44 35--47 35--46 35--45 35--44 35--43 35--41 35--39
#> + ... omitted several edges
#> 

sample_correlated_gnp_pair_w_junk(50, 0.3, 0.5, 40)

#> $graph1
#> IGRAPH 812a1ff U--- 50 659 -- 
#> + attr: name_1 (g/c), name_2 (g/c), loops_1 (g/l), loops_2 (g/l), type
#> | (g/c), p (g/n)
#> + edges from 812a1ff:
#>  [1] 47--49 47--48 46--49 46--48 46--47 45--50 45--47 45--46 44--50 44--46
#> [11] 43--49 43--46 43--44 42--50 42--48 42--45 42--43 41--49 41--48 41--47
#> [21] 41--45 41--42 40--49 40--48 40--46 40--45 40--44 40--43 40--42 40--41
#> [31] 39--49 39--48 39--47 39--46 39--45 39--43 39--42 39--41 38--49 38--46
#> [41] 38--45 38--44 38--43 38--40 37--50 37--49 37--48 37--46 37--45 37--41
#> [51] 37--39 36--50 36--46 36--44 36--43 36--42 36--41 36--40 36--39 36--38
#> [61] 36--37 35--49 35--48 35--46 35--42 35--40 35--39 35--37 35--36 34--50
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH 627d597 U--- 50 627 -- 
#> + attr: name_1 (g/c), name_2 (g/c), loops_1 (g/l), loops_2 (g/l),
#> | type_1 (g/c), type_2 (g/c), p_1 (g/n), p_2 (g/n), type (g/c), p (g/n)
#> + edges from 627d597:
#>  [1] 49--50 48--49 47--50 47--48 46--50 46--49 46--47 45--49 45--48 45--47
#> [11] 45--46 44--50 44--49 44--47 44--46 44--45 43--50 43--49 43--48 43--46
#> [21] 43--45 43--44 42--44 41--49 41--47 41--46 41--45 41--43 40--49 40--47
#> [31] 40--46 40--44 40--41 39--50 39--45 39--41 39--40 38--50 38--49 38--47
#> [41] 38--45 38--44 38--42 38--40 38--39 37--48 37--45 37--43 37--42 37--39
#> [51] 36--50 36--46 36--44 36--43 36--42 36--41 36--39 36--38 36--37 35--49
#> [61] 35--48 35--46 35--45 35--44 35--43 35--42 35--40 35--39 35--38 35--37
#> + ... omitted several edges
#>