Sample correlated G(n,p) random graphs

Sample a pair of correlated G(n,p) random graphs with correlation between two graphs being rho and edge probability being p.

sample_correlated_gnp_pair(n, corr, p, permutation = 1:n, ...)

sample_correlated_gnp_pair_w_junk(n, corr, p, ncore = n,
  permutation = 1:n, ...)

Arguments

n
An integer. Number of total vertices for the sampled graphs.

corr
A number. The target Pearson correlation between the adjacency matrices of the generated graphs. It must be in open (0,1) interval.

p
A number. Edge probability between two vertices. It must be in open (0,1) interval.

permutation
A numeric vector,permute second graph.

...
Passed to sample_correlated_gnp_pair.

ncore
An integer. Number of core vertices.

Value

sample_correlated_gnp_pair returns a list of two igraph object, named graph1 and graph2, which are two graphs whose adjacency matrix entries correlated with rho.

sample_correlated_gnp_pair_w_junk returns a list of two igraph object, named graph1 and graph2, which are two graphs whose adjacency matrix entries correlated with rho and with first ncore vertices being core vertices and the rest being junk vertices.

Examples

sample_correlated_gnp_pair(50, 0.3, 0.5)

#> $graph1
#> IGRAPH 4455bed U--- 50 626 -- Erdos renyi (gnp) graph
#> + attr: name (g/c), type (g/c), loops (g/l), p (g/n)
#> + edges from 4455bed:
#>  [1]  1-- 2  1-- 3  2-- 5  3-- 5  4-- 5  1-- 6  2-- 6  3-- 6  1-- 7  3-- 7
#> [11]  4-- 7  5-- 7  6-- 7  1-- 8  2-- 8  4-- 8  6-- 8  2-- 9  6-- 9  7-- 9
#> [21]  3--10  7--10  8--10  1--11  4--11  5--11  6--11  8--11 10--11  3--12
#> [31]  4--12  5--12  6--12  7--12  9--12 10--12  1--13  2--13  8--13  9--13
#> [41] 10--13  2--14  3--14  4--14  5--14  7--14  8--14  9--14 10--14  2--15
#> [51]  4--15  5--15  6--15  8--15  9--15 10--15 13--15  2--16  3--16  4--16
#> [61]  7--16  9--16 10--16 14--16 15--16  3--17  7--17  8--17 11--17 12--17
#> [71] 13--17 14--17 15--17  3--18  5--18  7--18  8--18 10--18 12--18 13--18
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH af11176 U--- 50 593 -- Erdos renyi (gnp) graph
#> + attr: name_1 (g/c), name_2 (g/c), type_1 (g/c), type_2 (g/c), loops_1
#> | (g/l), loops_2 (g/l), p_1 (g/n), p_2 (g/n), name (g/c), type (g/c),
#> | loops (g/l), p (g/n)
#> + edges from af11176:
#>  [1] 47--49 47--48 46--48 45--50 45--49 45--48 44--47 44--46 44--45 43--49
#> [11] 43--47 42--50 42--49 42--48 42--47 42--45 41--50 41--48 41--47 41--44
#> [21] 40--50 40--46 40--43 40--41 39--49 39--41 39--40 38--50 38--49 38--48
#> [31] 38--47 38--43 38--42 38--41 37--49 37--47 37--46 37--42 37--39 37--38
#> [41] 36--50 36--48 36--46 36--45 36--43 36--41 36--40 36--38 35--49 35--46
#> [51] 35--43 35--42 35--40 35--39 34--50 34--49 34--47 34--46 34--44 34--41
#> + ... omitted several edges
#> 

sample_correlated_gnp_pair_w_junk(50, 0.3, 0.5, 40)

#> $graph1
#> IGRAPH 2cf17a2 U--- 50 621 -- 
#> + attr: name_1 (g/c), name_2 (g/c), loops_1 (g/l), loops_2 (g/l), type
#> | (g/c), p (g/n)
#> + edges from 2cf17a2:
#>  [1] 48--50 48--49 47--50 47--49 45--49 45--46 44--48 43--49 43--48 43--46
#> [11] 43--44 42--50 42--44 41--50 41--49 41--47 41--42 40--50 40--48 40--44
#> [21] 40--43 39--48 39--47 39--46 39--45 39--42 39--41 38--50 38--49 38--48
#> [31] 38--46 38--45 38--39 37--50 37--49 37--48 37--46 37--42 37--40 36--50
#> [41] 36--49 36--45 36--43 36--40 36--39 36--38 35--49 35--48 35--47 35--44
#> [51] 35--43 35--40 35--37 35--36 34--50 34--49 34--46 34--45 34--44 34--41
#> [61] 34--39 34--38 34--36 33--50 33--48 33--47 33--46 33--45 33--44 33--43
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH cb9f3d7 U--- 50 630 -- 
#> + attr: name_1 (g/c), name_2 (g/c), loops_1 (g/l), loops_2 (g/l),
#> | type_1 (g/c), type_2 (g/c), p_1 (g/n), p_2 (g/n), type (g/c), p (g/n)
#> + edges from cb9f3d7:
#>  [1] 46--48 46--47 45--46 44--50 44--49 44--48 44--47 43--50 43--48 43--47
#> [11] 43--44 42--48 42--47 42--46 42--44 42--43 41--46 41--44 41--42 40--49
#> [21] 40--48 40--47 40--44 39--47 39--45 39--44 39--43 39--41 39--40 38--48
#> [31] 38--46 38--45 38--44 38--43 38--41 38--40 38--39 37--50 37--49 37--48
#> [41] 37--47 37--46 37--45 37--41 37--40 37--39 36--49 36--48 36--46 36--42
#> [51] 36--39 36--37 35--50 35--45 35--42 35--41 35--40 35--37 35--36 34--49
#> [61] 34--48 34--47 34--45 34--42 34--41 34--40 34--38 34--35 33--45 33--44
#> + ... omitted several edges
#>