Sample correlated G(n,p) random graphs

Sample a pair of correlated G(n,p) random graphs with correlation between two graphs being rho and edge probability being p.

sample_correlated_gnp_pair(n, corr, p, permutation = 1:n, ...)

sample_correlated_gnp_pair_w_junk(
  n,
  corr,
  p,
  ncore = n,
  permutation = 1:n,
  ...
)

Arguments

n
An integer. Number of total vertices for the sampled graphs.

corr
A number. The target Pearson correlation between the adjacency matrices of the generated graphs. It must be in open (0,1) interval.

p
A number. Edge probability between two vertices. It must be in open (0,1) interval.

permutation
A numeric vector,permute second graph.

...
Passed to sample_correlated_gnp_pair and sample_correlated_gnp_pair_w_junk.

ncore
An integer. Number of core vertices.

Value

sample_correlated_gnp_pair returns a list of two igraph object, named graph1 and graph2, which are two graphs whose adjacency matrix entries correlated with rho.

sample_correlated_gnp_pair_w_junk returns a list of two igraph object, named graph1 and graph2, which are two graphs whose adjacency matrix entries correlated with rho and with first ncore vertices being core vertices and the rest being junk vertices.

Examples

sample_correlated_gnp_pair(50, 0.3, 0.5)

#> $graph1
#> IGRAPH 65b178f U--- 50 616 -- Erdos renyi (gnp) graph
#> + attr: name (g/c), type (g/c), loops (g/l), p (g/n)
#> + edges from 65b178f:
#>  [1]  1-- 2  1-- 4  2-- 4  3-- 4  1-- 5  3-- 5  4-- 5  3-- 6  4-- 6  1-- 7
#> [11]  2-- 7  3-- 7  6-- 7  1-- 8  4-- 8  5-- 8  2-- 9  3-- 9  4-- 9  7-- 9
#> [21]  1--10  3--10  4--10  7--10  8--10  2--11  4--11  5--11  6--11  8--11
#> [31]  1--12  2--12  3--12  4--12  6--12  7--12  8--12  9--12 10--12 11--12
#> [41]  4--13  5--13  9--13 12--13  1--14  3--14  4--14  5--14  6--14  7--14
#> [51] 11--14 13--14  2--15  6--15  7--15  8--15 10--15 11--15 13--15  1--16
#> [61]  5--16  6--16  9--16 11--16 12--16 13--16 14--16  1--17  3--17  4--17
#> [71]  8--17 13--17  3--18  5--18  6--18  7--18  8--18 11--18 12--18 17--18
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH 2d6edfd U--- 50 620 -- Erdos renyi (gnp) graph
#> + attr: name_1 (g/c), name_2 (g/c), type_1 (g/c), type_2 (g/c), loops_1
#> | (g/l), loops_2 (g/l), p_1 (g/n), p_2 (g/n), name (g/c), type (g/c),
#> | loops (g/l), p (g/n)
#> + edges from 2d6edfd:
#>  [1] 49--50 48--50 47--50 47--49 46--50 46--49 46--48 46--47 45--48 45--47
#> [11] 45--46 44--49 44--46 44--45 43--48 43--46 43--44 42--49 42--44 42--43
#> [21] 41--49 41--46 41--45 41--44 40--50 40--48 40--46 40--45 40--44 40--43
#> [31] 40--42 39--50 39--48 39--43 39--41 38--50 38--49 38--48 38--47 38--43
#> [41] 38--42 38--39 37--47 37--45 37--42 37--38 36--49 36--45 36--42 36--41
#> [51] 36--38 35--49 35--48 35--47 35--46 35--45 35--44 35--43 35--42 35--40
#> + ... omitted several edges
#> 

sample_correlated_gnp_pair_w_junk(50, 0.3, 0.5, 40)

#> $graph1
#> IGRAPH decbb8b U--- 50 635 -- 
#> + attr: name_1 (g/c), name_2 (g/c), loops_1 (g/l), loops_2 (g/l), type
#> | (g/c), p (g/n)
#> + edges from decbb8b:
#>  [1] 49--50 48--50 47--50 47--48 46--50 46--49 46--48 45--48 44--49 44--48
#> [11] 44--46 43--49 43--48 42--48 42--47 42--43 41--46 41--45 41--44 41--43
#> [21] 41--42 40--49 40--48 40--46 40--43 40--41 39--49 39--46 39--45 39--44
#> [31] 38--50 38--49 38--47 38--45 38--43 38--39 37--49 37--48 37--47 37--44
#> [41] 36--47 36--45 36--44 36--43 36--41 36--40 36--38 35--50 35--48 35--47
#> [51] 35--46 35--45 35--41 35--40 35--38 35--37 35--36 34--50 34--47 34--46
#> [61] 34--44 34--38 34--37 34--36 33--50 33--49 33--48 33--47 33--46 33--42
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH 6e3306d U--- 50 614 -- 
#> + attr: name_1 (g/c), name_2 (g/c), loops_1 (g/l), loops_2 (g/l),
#> | type_1 (g/c), type_2 (g/c), p_1 (g/n), p_2 (g/n), type (g/c), p (g/n)
#> + edges from 6e3306d:
#>  [1] 49--50 47--49 46--49 46--48 46--47 45--47 44--50 44--48 44--47 44--45
#> [11] 43--50 43--49 43--48 43--46 43--45 43--44 42--50 42--49 42--45 42--44
#> [21] 41--50 41--48 41--47 41--44 41--43 40--45 40--44 40--41 39--50 39--48
#> [31] 39--47 39--45 39--40 38--49 38--47 38--46 38--42 38--40 38--39 37--46
#> [41] 37--44 37--43 37--41 37--40 36--46 36--44 36--41 36--38 35--49 35--48
#> [51] 35--46 35--45 35--44 35--43 35--41 34--50 34--49 34--47 34--46 34--44
#> [61] 34--43 34--41 34--40 34--38 34--36 33--50 33--49 33--48 33--47 33--44
#> + ... omitted several edges
#>