Sample graphs from edge probability matrix and correlation matrix

Sample a pair of graphs with specified edge probability and correlation between each pair of vertices.

sample_correlated_ieg_pair(
  n,
  p_mat,
  c_mat,
  directed = FALSE,
  loops = FALSE,
  permutation = 1:n
)

sample_correlated_rdpg(X, rho, nc = nrow(X), ...)

Arguments

n
An integer. Number of total vertices for the sampled graphs.

p_mat
An n-by-n matrix. Edge probability matrix, each entry should be in the open (0,1) interval.

c_mat
An n-by-n matrix. The target Pearson correlation matrix, each entry should be in the open (0,1) interval.

directed
Logical scalar, whether to generate directed graphs.

loops
Logical scalar, whether self-loops are allowed in the graph.

permutation
A numeric vector,permute second graph.

X
A matrix. Dot products matrix, each entry must be in open (0,1) interval.

rho
A number. The target Pearson correlation between the adjacency matrices of the generated graphs. It must be in open (0,1) interval.

nc
An integer. Number of core vertices.

...
Passed to sample_correlated_rdpg_pair.

Value

sample_correlated_ieg_pair returns two igraph objects named graph1 and graph2.

sample_correlated_rdpg returns two igraph objects named graph1 and graph2 that are sampled from random dot product graphs model.

Examples

n <- 50
p_mat <- matrix(runif(n^2),n)
c_mat <- matrix(runif(n^2),n)
sample_correlated_ieg_pair(n,p_mat,c_mat)

#> $graph1
#> IGRAPH 1d86e51 U--- 50 576 -- 
#> + edges from 1d86e51:
#>   [1] 1-- 3 1-- 7 1-- 8 1--11 1--13 1--15 1--17 1--19 1--21 1--26 1--28 1--31
#>  [13] 1--34 1--36 1--38 1--39 1--41 1--42 1--43 1--46 1--48 1--49 1--50 2-- 3
#>  [25] 2-- 4 2-- 9 2--10 2--12 2--13 2--15 2--16 2--20 2--21 2--23 2--24 2--26
#>  [37] 2--29 2--30 2--34 2--35 2--38 2--42 2--43 2--45 2--48 2--50 3-- 6 3-- 7
#>  [49] 3-- 8 3--12 3--14 3--15 3--16 3--17 3--22 3--23 3--24 3--30 3--34 3--35
#>  [61] 3--36 3--39 3--42 3--44 3--48 4-- 5 4-- 9 4--10 4--11 4--12 4--13 4--15
#>  [73] 4--16 4--19 4--21 4--23 4--24 4--26 4--27 4--30 4--31 4--32 4--33 4--36
#>  [85] 4--37 4--39 4--40 4--41 4--42 4--44 4--45 4--46 4--47 4--48 5-- 8 5-- 9
#>  [97] 5--12 5--13 5--14 5--15 5--16 5--18 5--19 5--23 5--25 5--28 5--30 5--34
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH 9ff4bb3 U--- 50 578 -- 
#> + edges from 9ff4bb3:
#>   [1] 1-- 3 1-- 4 1-- 6 1-- 7 1-- 8 1--11 1--12 1--13 1--15 1--19 1--20 1--21
#>  [13] 1--26 1--28 1--31 1--34 1--36 1--38 1--39 1--42 1--43 1--46 1--48 1--49
#>  [25] 1--50 2-- 3 2-- 9 2--10 2--12 2--13 2--15 2--16 2--20 2--23 2--24 2--26
#>  [37] 2--29 2--32 2--35 2--38 2--42 2--43 2--44 2--45 2--48 2--50 3-- 6 3-- 8
#>  [49] 3--10 3--14 3--15 3--16 3--19 3--22 3--23 3--30 3--31 3--34 3--35 3--36
#>  [61] 3--42 3--44 3--48 3--49 4-- 6 4-- 9 4--10 4--13 4--16 4--19 4--20 4--21
#>  [73] 4--23 4--24 4--26 4--31 4--32 4--33 4--36 4--37 4--39 4--40 4--42 4--44
#>  [85] 4--45 4--46 4--48 5-- 8 5-- 9 5--10 5--12 5--13 5--14 5--15 5--16 5--19
#>  [97] 5--20 5--23 5--25 5--28 5--30 5--34 5--36 5--38 5--39 5--42 5--44 5--45
#> + ... omitted several edges
#> 

## sample a pair of igraph objects from random dot
## product graphs model with dimension 3 and scale 8
n <- 50
xdim <- 3
scale <- 8
X <- matrix(rgamma(n*(xdim+1),scale,1),n,xdim+1)
X <- X/rowSums(X)
X <- X[,1:xdim]
sample_correlated_rdpg(X,rho=0.5)

#> $graph1
#> IGRAPH e8fc080 U--- 50 229 -- 
#> + edges from e8fc080:
#>  [1]  1-- 3  1-- 7  1--12  1--17  1--25  1--30  1--35  1--36  1--37  1--39
#> [11]  1--49  2--11  2--16  2--17  2--20  2--29  2--32  3--15  3--21  3--22
#> [21]  3--24  3--38  3--44  3--46  3--48  4-- 5  4--14  4--28  4--30  4--33
#> [31]  4--36  4--39  4--44  5--11  5--17  5--21  5--26  5--33  5--50  6--10
#> [41]  6--22  6--28  6--29  6--31  6--34  6--35  6--36  6--37  6--40  6--42
#> [51]  6--49  7--10  7--15  7--24  7--26  7--30  7--31  7--32  7--43  8--15
#> [61]  8--20  8--23  8--25  8--34  9--29  9--38  9--39  9--41  9--45 10--11
#> [71] 10--12 10--14 10--16 10--19 10--21 10--23 10--24 10--31 10--36 10--41
#> [81] 10--43 11--12 11--27 11--32 11--38 11--41 11--43 12--22 12--23 12--39
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH 99d2fd1 U--- 50 240 -- 
#> + edges from 99d2fd1:
#>  [1]  1-- 2  1-- 3  1-- 4  1-- 7  1-- 9  1--12  1--15  1--17  1--18  1--30
#> [11]  1--35  1--36  1--37  1--50  2--10  2--16  2--17  2--20  2--22  2--39
#> [21]  3-- 6  3--21  3--22  3--25  3--35  3--42  3--44  3--46  3--48  4-- 5
#> [31]  4--14  4--16  4--33  4--39  4--44  4--47  5--10  5--11  5--15  5--17
#> [41]  5--23  5--32  5--39  5--42  5--43  5--45  5--50  6--20  6--29  6--34
#> [51]  6--37  6--42  6--49  7-- 8  7--10  7--15  7--19  7--24  7--31  7--32
#> [61]  7--42  7--50  8--23  8--24  8--30  8--34  8--42  8--49  9--17  9--19
#> [71]  9--20  9--27  9--37  9--39  9--41  9--45 10--11 10--14 10--16 10--18
#> [81] 10--23 10--31 10--40 10--41 10--43 11--12 11--17 11--27 11--36 11--43
#> + ... omitted several edges
#>