Sample graphs from edge probability matrix and correlation matrix

Sample a pair of graphs with specified edge probability and correlation between each pair of vertices.

sample_correlated_ieg_pair(
  n,
  p_mat,
  c_mat,
  directed = FALSE,
  loops = FALSE,
  permutation = 1:n
)

sample_correlated_rdpg(X, rho, nc = nrow(X), ...)

Arguments

n
An integer. Number of total vertices for the sampled graphs.

p_mat
An n-by-n matrix. Edge probability matrix, each entry should be in the open (0,1) interval.

c_mat
An n-by-n matrix. The target Pearson correlation matrix, each entry should be in the open (0,1) interval.

directed
Logical scalar, whether to generate directed graphs.

loops
Logical scalar, whether self-loops are allowed in the graph.

permutation
A numeric vector,permute second graph.

X
A matrix. Dot products matrix, each entry must be in open (0,1) interval.

rho
A number. The target Pearson correlation between the adjacency matrices of the generated graphs. It must be in open (0,1) interval.

nc
An integer. Number of core vertices.

...
Passed to sample_correlated_rdpg_pair.

Value

sample_correlated_ieg_pair returns two igraph objects named graph1 and graph2.

sample_correlated_rdpg returns two igraph objects named graph1 and graph2 that are sampled from random dot product graphs model.

Examples

n <- 50
p_mat <- matrix(runif(n^2),n)
c_mat <- matrix(runif(n^2),n)
sample_correlated_ieg_pair(n,p_mat,c_mat)

#> $graph1
#> IGRAPH 070156f U--- 50 607 -- 
#> + edges from 070156f:
#>   [1] 1-- 5 1-- 7 1-- 8 1--10 1--11 1--13 1--15 1--17 1--20 1--22 1--24 1--26
#>  [13] 1--28 1--31 1--32 1--34 1--38 1--41 1--42 1--46 1--47 1--48 1--49 2-- 4
#>  [25] 2-- 5 2-- 6 2-- 7 2--11 2--12 2--13 2--14 2--16 2--17 2--18 2--19 2--20
#>  [37] 2--21 2--23 2--25 2--26 2--27 2--29 2--31 2--32 2--33 2--34 2--35 2--36
#>  [49] 2--41 2--46 2--48 3-- 5 3-- 6 3-- 8 3--12 3--14 3--16 3--17 3--21 3--22
#>  [61] 3--24 3--25 3--28 3--30 3--31 3--32 3--34 3--35 3--37 3--38 3--42 3--44
#>  [73] 3--46 3--48 3--49 4-- 5 4-- 8 4--10 4--13 4--14 4--16 4--19 4--22 4--24
#>  [85] 4--26 4--28 4--30 4--31 4--32 4--35 4--36 4--38 4--39 4--41 4--42 4--47
#>  [97] 4--48 5-- 6 5-- 7 5--10 5--14 5--16 5--17 5--19 5--21 5--23 5--25 5--30
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH b2a8e9d U--- 50 621 -- 
#> + edges from b2a8e9d:
#>   [1] 1-- 2 1-- 5 1-- 6 1-- 7 1--10 1--11 1--13 1--14 1--15 1--17 1--20 1--22
#>  [13] 1--24 1--26 1--28 1--29 1--33 1--35 1--38 1--41 1--42 1--46 1--48 1--49
#>  [25] 2-- 3 2-- 4 2-- 5 2-- 6 2-- 7 2-- 8 2--11 2--12 2--13 2--14 2--16 2--18
#>  [37] 2--19 2--20 2--23 2--25 2--27 2--30 2--31 2--32 2--33 2--34 2--35 2--36
#>  [49] 2--39 2--40 2--41 2--44 2--48 3-- 6 3-- 8 3-- 9 3--12 3--14 3--16 3--17
#>  [61] 3--18 3--21 3--22 3--24 3--25 3--28 3--30 3--31 3--32 3--34 3--35 3--38
#>  [73] 3--39 3--42 3--44 3--45 3--46 3--48 3--49 4-- 5 4-- 8 4--10 4--11 4--13
#>  [85] 4--14 4--16 4--17 4--18 4--22 4--24 4--26 4--28 4--30 4--31 4--32 4--35
#>  [97] 4--36 4--38 4--39 4--41 4--42 4--46 4--47 4--48 5-- 6 5-- 7 5-- 8 5--10
#> + ... omitted several edges
#> 

## sample a pair of igraph objects from random dot
## product graphs model with dimension 3 and scale 8
n <- 50
xdim <- 3
scale <- 8
X <- matrix(rgamma(n*(xdim+1),scale,1),n,xdim+1)
X <- X/rowSums(X)
X <- X[,1:xdim]
sample_correlated_rdpg(X,rho=0.5)

#> $graph1
#> IGRAPH f00f60d U--- 50 241 -- 
#> + edges from f00f60d:
#>  [1]  1-- 4  1-- 5  1-- 8  1--13  1--16  1--25  1--34  1--40  1--43  2-- 5
#> [11]  2-- 6  2--12  2--18  2--21  2--50  3--10  3--13  3--15  3--16  3--22
#> [21]  3--24  3--29  3--33  3--34  3--42  4--11  4--12  4--13  4--15  4--17
#> [31]  4--18  4--25  4--26  4--27  4--31  4--32  4--38  4--47  5-- 6  5--13
#> [41]  5--14  5--20  5--27  5--32  5--36  5--40  6-- 7  6--26  6--45  7-- 8
#> [51]  7-- 9  7--18  7--24  7--26  7--33  7--39  7--41  7--46  7--50  8--13
#> [61]  8--14  8--17  8--32  9--13  9--22  9--23  9--25  9--28  9--32  9--33
#> [71]  9--41  9--44  9--46  9--48  9--50 10--12 10--15 10--18 10--24 10--27
#> [81] 10--34 10--36 10--41 10--42 10--43 10--45 10--50 11--14 11--17 11--27
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH 3b2f6a0 U--- 50 257 -- 
#> + edges from 3b2f6a0:
#>  [1]  1-- 4  1-- 8  1--13  1--16  1--22  1--25  1--34  1--40  1--41  1--43
#> [11]  1--44  2-- 5  2-- 6  2-- 9  2--21  2--24  2--29  2--30  3--10  3--13
#> [21]  3--15  3--16  3--22  3--29  3--32  3--33  3--35  3--39  4-- 8  4--11
#> [31]  4--12  4--15  4--17  4--25  4--30  4--31  4--40  4--44  4--47  5-- 6
#> [41]  5--10  5--14  5--18  5--20  5--27  5--32  5--36  5--40  6-- 7  6-- 8
#> [51]  6--17  6--32  6--34  6--36  6--45  7-- 9  7--18  7--24  7--25  7--26
#> [61]  7--30  7--40  7--41  7--45  7--50  8-- 9  8--13  8--17  8--24  8--28
#> [71]  8--35  8--39  8--48  9--13  9--17  9--19  9--25  9--29  9--32  9--46
#> [81]  9--48  9--50 10--11 10--15 10--18 10--19 10--29 10--40 10--41 10--43
#> + ... omitted several edges
#>