Sample graphs from edge probability matrix and correlation matrix

Sample a pair of graphs with specified edge probability and correlation between each pair of vertices.

sample_correlated_ieg_pair(
  n,
  p_mat,
  c_mat,
  directed = FALSE,
  loops = FALSE,
  permutation = 1:n
)

sample_correlated_rdpg(X, rho, nc = nrow(X), ...)

Arguments

n
An integer. Number of total vertices for the sampled graphs.

p_mat
An n-by-n matrix. Edge probability matrix, each entry should be in the open (0,1) interval.

c_mat
An n-by-n matrix. The target Pearson correlation matrix, each entry should be in the open (0,1) interval.

directed
Logical scalar, whether to generate directed graphs.

loops
Logical scalar, whether self-loops are allowed in the graph.

permutation
A numeric vector,permute second graph.

X
A matrix. Dot products matrix, each entry must be in open (0,1) interval.

rho
A number. The target Pearson correlation between the adjacency matrices of the generated graphs. It must be in open (0,1) interval.

nc
An integer. Number of core vertices.

...
Passed to sample_correlated_rdpg_pair.

Value

sample_correlated_ieg_pair returns two igraph objects named graph1 and graph2.

sample_correlated_rdpg returns two igraph objects named graph1 and graph2 that are sampled from random dot product graphs model.

Examples

n <- 50
p_mat <- matrix(runif(n^2),n)
c_mat <- matrix(runif(n^2),n)
sample_correlated_ieg_pair(n,p_mat,c_mat)

#> $graph1
#> IGRAPH bc2ce61 U--- 50 592 -- 
#> + edges from bc2ce61:
#>   [1] 1-- 2 1-- 3 1-- 4 1-- 5 1--10 1--11 1--12 1--13 1--16 1--18 1--20 1--23
#>  [13] 1--27 1--28 1--29 1--32 1--33 1--34 1--39 1--40 1--41 1--44 1--47 2-- 7
#>  [25] 2-- 9 2--10 2--13 2--14 2--15 2--16 2--17 2--20 2--26 2--28 2--29 2--30
#>  [37] 2--32 2--33 2--34 2--36 2--38 2--39 2--41 2--44 2--45 2--46 2--49 3-- 8
#>  [49] 3--11 3--13 3--14 3--18 3--19 3--20 3--21 3--26 3--27 3--28 3--29 3--30
#>  [61] 3--31 3--33 3--35 3--38 3--39 3--40 3--42 3--43 3--45 3--48 3--50 4-- 7
#>  [73] 4-- 9 4--11 4--13 4--15 4--17 4--22 4--24 4--27 4--30 4--32 4--34 4--35
#>  [85] 4--37 4--38 4--39 4--42 4--44 4--45 4--46 4--48 4--49 5-- 6 5-- 8 5-- 9
#>  [97] 5--11 5--12 5--16 5--17 5--19 5--20 5--22 5--25 5--26 5--30 5--31 5--34
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH 90792f4 U--- 50 599 -- 
#> + edges from 90792f4:
#>   [1] 1-- 2 1-- 3 1-- 4 1-- 5 1--10 1--13 1--16 1--20 1--23 1--25 1--26 1--27
#>  [13] 1--28 1--29 1--31 1--32 1--33 1--34 1--39 1--40 1--41 1--44 1--47 2-- 7
#>  [25] 2-- 9 2--10 2--12 2--13 2--14 2--15 2--16 2--17 2--23 2--24 2--26 2--28
#>  [37] 2--29 2--30 2--31 2--32 2--33 2--34 2--36 2--39 2--41 2--42 2--45 2--46
#>  [49] 2--49 3-- 5 3-- 8 3--11 3--14 3--18 3--19 3--20 3--21 3--26 3--27 3--28
#>  [61] 3--29 3--31 3--32 3--33 3--35 3--36 3--37 3--38 3--39 3--40 3--41 3--42
#>  [73] 3--43 3--45 3--48 3--50 4-- 7 4-- 8 4-- 9 4--11 4--15 4--16 4--17 4--22
#>  [85] 4--24 4--27 4--30 4--32 4--34 4--35 4--38 4--39 4--42 4--44 4--45 4--46
#>  [97] 4--48 4--49 5-- 6 5-- 8 5-- 9 5--11 5--12 5--16 5--19 5--20 5--22 5--25
#> + ... omitted several edges
#> 

## sample a pair of igraph objects from random dot
## product graphs model with dimension 3 and scale 8
n <- 50
xdim <- 3
scale <- 8
X <- matrix(rgamma(n*(xdim+1),scale,1),n,xdim+1)
X <- X/rowSums(X)
X <- X[,1:xdim]
sample_correlated_rdpg(X,rho=0.5)

#> $graph1
#> IGRAPH 654e6de U--- 50 236 -- 
#> + edges from 654e6de:
#>  [1]  1-- 3  1-- 8  1--13  1--21  1--26  1--31  1--32  1--33  1--35  1--45
#> [11]  1--47  1--48  1--50  2-- 7  2--12  2--13  2--16  2--25  2--28  2--47
#> [21]  3--17  3--20  3--34  3--40  3--42  3--44  4--10  4--11  4--24  4--26
#> [31]  4--29  4--32  4--35  4--40  4--44  5-- 7  5--13  5--17  5--22  5--29
#> [41]  5--46  5--47  5--48  6--24  6--25  6--27  6--31  6--32  6--33  6--36
#> [51]  6--38  6--45  7--11  7--20  7--22  7--26  7--27  7--28  7--39  7--46
#> [61]  8--11  8--16  8--19  8--21  8--24  8--30  8--38  8--42  8--48  8--49
#> [71]  9--17  9--25  9--34  9--35  9--37  9--41  9--42  9--47  9--50 10--15
#> [81] 10--17 10--19 10--20 10--27 10--32 10--37 10--39 10--47 11--28 11--34
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH a089ba9 U--- 50 233 -- 
#> + edges from a089ba9:
#>  [1]  1-- 3  1-- 5  1-- 8  1--11  1--13  1--14  1--26  1--31  1--32  1--33
#> [11]  1--46  1--47  1--48  2-- 6  2--12  2--13  2--16  2--18  2--35  2--47
#> [21]  3--17  3--21  3--31  3--38  3--40  3--42  3--44  4--10  4--11  4--12
#> [31]  4--24  4--29  4--35  4--38  4--40  4--43  4--44  5-- 7  5--11  5--13
#> [41]  5--19  5--35  5--39  5--41  5--47  5--49  6--25  6--33  6--38  6--45
#> [51]  7--11  7--15  7--27  7--28  7--38  7--46  7--49  8--19  8--20  8--24
#> [61]  8--26  8--29  8--30  8--38  8--42  8--45  8--48  8--49  9--13  9--15
#> [71]  9--16  9--17  9--23  9--33  9--35  9--37  9--41  9--47  9--50 10--14
#> [81] 10--19 10--27 10--37 10--39 10--47 10--48 11--13 11--32 11--39 11--48
#> + ... omitted several edges
#>