Sample graphs from edge probability matrix and correlation matrix

Sample a pair of graphs with specified edge probability and correlation between each pair of vertices.

sample_correlated_ieg_pair(
  n,
  p_mat,
  c_mat,
  directed = FALSE,
  loops = FALSE,
  permutation = 1:n
)

sample_correlated_rdpg(X, rho, nc = nrow(X), ...)

Arguments

n
An integer. Number of total vertices for the sampled graphs.

p_mat
An n-by-n matrix. Edge probability matrix, each entry should be in the open (0,1) interval.

c_mat
An n-by-n matrix. The target Pearson correlation matrix, each entry should be in the open (0,1) interval.

directed
Logical scalar, whether to generate directed graphs.

loops
Logical scalar, whether self-loops are allowed in the graph.

permutation
A numeric vector,permute second graph.

X
A matrix. Dot products matrix, each entry must be in open (0,1) interval.

rho
A number. The target Pearson correlation between the adjacency matrices of the generated graphs. It must be in open (0,1) interval.

nc
An integer. Number of core vertices.

...
Passed to sample_correlated_rdpg_pair.

Value

sample_correlated_ieg_pair returns two igraph objects named graph1 and graph2.

sample_correlated_rdpg returns two igraph objects named graph1 and graph2 that are sampled from random dot product graphs model.

Examples

n <- 50
p_mat <- matrix(runif(n^2),n)
c_mat <- matrix(runif(n^2),n)
sample_correlated_ieg_pair(n,p_mat,c_mat)

#> $graph1
#> IGRAPH 7762c69 U--- 50 597 -- 
#> + edges from 7762c69:
#>   [1] 1-- 3 1-- 4 1-- 6 1-- 7 1-- 8 1-- 9 1--10 1--11 1--12 1--17 1--18 1--19
#>  [13] 1--21 1--23 1--26 1--27 1--28 1--30 1--32 1--33 1--34 1--37 1--40 1--41
#>  [25] 1--42 1--43 1--44 1--45 1--46 1--48 2-- 4 2-- 6 2-- 8 2-- 9 2--13 2--14
#>  [37] 2--17 2--18 2--19 2--20 2--22 2--23 2--24 2--26 2--27 2--28 2--30 2--35
#>  [49] 2--38 2--41 2--44 2--45 2--46 2--49 2--50 3-- 5 3-- 6 3-- 9 3--10 3--11
#>  [61] 3--12 3--14 3--15 3--21 3--22 3--23 3--24 3--27 3--28 3--32 3--33 3--34
#>  [73] 3--36 3--37 3--42 3--50 4-- 9 4--11 4--12 4--14 4--18 4--19 4--20 4--21
#>  [85] 4--22 4--23 4--24 4--26 4--27 4--29 4--35 4--37 4--40 4--44 4--45 4--46
#>  [97] 4--47 4--48 4--49 5--11 5--14 5--15 5--16 5--19 5--23 5--24 5--25 5--27
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH 213802b U--- 50 588 -- 
#> + edges from 213802b:
#>   [1] 1-- 6 1-- 8 1-- 9 1--10 1--11 1--12 1--17 1--18 1--19 1--21 1--23 1--25
#>  [13] 1--26 1--27 1--28 1--29 1--30 1--31 1--32 1--34 1--37 1--40 1--41 1--42
#>  [25] 1--44 1--45 1--46 2-- 4 2-- 8 2--13 2--14 2--17 2--18 2--19 2--20 2--22
#>  [37] 2--23 2--24 2--25 2--26 2--28 2--30 2--35 2--36 2--38 2--41 2--45 2--46
#>  [49] 2--49 2--50 3-- 5 3-- 6 3-- 9 3--10 3--11 3--12 3--14 3--15 3--17 3--21
#>  [61] 3--22 3--23 3--24 3--28 3--32 3--33 3--34 3--36 3--37 3--38 3--39 3--42
#>  [73] 3--43 3--50 4-- 9 4--10 4--11 4--14 4--18 4--19 4--20 4--21 4--22 4--23
#>  [85] 4--24 4--26 4--27 4--29 4--31 4--34 4--35 4--40 4--44 4--45 4--46 4--47
#>  [97] 4--49 5--14 5--15 5--16 5--19 5--23 5--25 5--28 5--30 5--33 5--36 5--37
#> + ... omitted several edges
#> 

## sample a pair of igraph objects from random dot
## product graphs model with dimension 3 and scale 8
n <- 50
xdim <- 3
scale <- 8
X <- matrix(rgamma(n*(xdim+1),scale,1),n,xdim+1)
X <- X/rowSums(X)
X <- X[,1:xdim]
sample_correlated_rdpg(X,rho=0.5)

#> $graph1
#> IGRAPH 6b9478d U--- 50 231 -- 
#> + edges from 6b9478d:
#>  [1]  1-- 2  1-- 9  1--11  1--13  1--40  1--50  2-- 5  2-- 6  2--17  2--23
#> [11]  2--31  2--32  2--35  2--41  2--44  2--48  2--49  2--50  3--10  3--11
#> [21]  3--18  3--23  3--34  3--44  3--46  4--10  4--16  4--19  4--29  4--36
#> [31]  4--39  4--41  4--43  4--50  5-- 9  5--18  5--24  5--36  5--46  6--16
#> [41]  6--18  6--25  6--31  6--36  6--37  6--41  6--49  7--19  7--23  7--27
#> [51]  7--31  7--33  7--34  7--36  7--39  7--43  8-- 9  8--10  8--11  8--15
#> [61]  8--32  8--35  8--44  8--47  8--50  9--11  9--13  9--22  9--25  9--30
#> [71]  9--36  9--37  9--41 10--11 10--14 10--21 10--25 10--31 10--38 10--45
#> [81] 10--47 11--14 11--16 11--20 11--24 11--25 11--27 11--31 11--33 11--35
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH f723091 U--- 50 231 -- 
#> + edges from f723091:
#>  [1]  1-- 2  1-- 9  1--17  1--20  1--24  1--32  1--38  1--40  2-- 5  2-- 6
#> [11]  2--15  2--17  2--26  2--31  2--32  2--35  2--39  2--40  2--41  2--42
#> [21]  3--17  3--18  3--29  3--31  3--34  3--44  3--46  3--48  4--16  4--19
#> [31]  4--24  4--31  4--36  4--39  4--47  5--13  5--24  5--47  5--48  6--12
#> [41]  6--28  6--31  6--37  6--44  6--49  7--19  7--23  7--27  7--33  7--34
#> [51]  7--36  7--40  7--43  7--49  7--50  8--11  8--15  8--19  8--30  8--32
#> [61]  8--33  8--50  9--11  9--15  9--25  9--30  9--33  9--34  9--35  9--41
#> [71] 10--11 10--19 10--25 10--31 10--35 10--38 10--50 11--15 11--16 11--20
#> [81] 11--22 11--23 11--24 11--27 11--33 11--36 11--38 11--39 11--49 12--18
#> + ... omitted several edges
#>