Sample graphs pair from stochastic block model

Sample a pair of random graphs from stochastic block model with correlation between two graphs being rho and edge probability being p.

sample_correlated_sbm_pair(
  n,
  pref.matrix,
  block.sizes,
  rho,
  permutation = 1:n,
  ...
)

sample_correlated_sbm_pair_w_junk(
  n,
  pref.matrix,
  block.sizes,
  rho,
  core.block.sizes,
  permutation = 1:n,
  ...
)

Arguments

n
An integer. Number of vertices in the graph.

pref.matrix
The matrix giving the Bernoulli rates. This is a K-by-K matrix, where k is the number of groups. The probability of creating an edge between vertices from groups i and j is given by element i,j. For undirected graphs, this matrix must be symmetric.

block.sizes
A numeric vector. Give the number of vertices in each group. The sum of the vector must match the number of vertices.

rho
A number. The target Pearson correlation between the adjacency matrices of the generated graphs. It must be in open (0,1) interval.

permutation
A numeric vector, permute second graph.

...
Passed to sample_correlated_sbm_pair and sample_correlated_sbm_pair_w_junk.

core.block.sizes
A numeric vector. Give the number of core vertices in each group. Entries should be smaller than block.sizes and the vector length should be the same as block.sizes.

Value

A list of two igraph object, named graph1 and graph2.

Examples

pm <- cbind( c(.1, .001), c(.001, .05) )
sample_correlated_sbm_pair(1000, pref.matrix=pm, block.sizes=c(300,700), rho=0.5)

#> $graph1
#> IGRAPH 46d52c3 U--- 1000 16870 -- Stochastic block-model
#> + attr: name (g/c), loops (g/l)
#> + edges from 46d52c3:
#>  [1]  1-- 2  2-- 3  5-- 8  2-- 9  1--12  6--12  8--13  8--14  4--15  4--16
#> [11]  2--17  5--17  9--17 11--17  5--18 10--18  8--20  9--21 10--21  2--22
#> [21]  3--22  5--22  7--22 19--22 17--23 12--24 20--25  5--26 18--26 11--27
#> [31] 16--27 21--27  9--28  6--29 14--29 25--29 18--30 23--30 26--30  7--31
#> [41]  9--31 11--31 17--31 21--31  3--32 17--32 10--33 15--33 17--33 31--33
#> [51] 32--33 30--34 32--34 27--35 33--35  1--36  3--36  5--36 10--36  3--37
#> [61] 17--37 26--37  1--38  5--38 20--38 33--38  3--39 10--39 19--39 22--39
#> [71] 30--39  4--40 12--40 17--40 22--40 14--41 19--41 25--41 30--41 36--41
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH 2eca25a U--- 1000 16735 -- Stochastic block-model
#> + attr: name_1 (g/c), name_2 (g/c), loops_1 (g/l), loops_2 (g/l), name
#> | (g/c), loops (g/l)
#> + edges from 2eca25a:
#>  [1] 996-- 997 993-- 999 990--1000 988-- 993 987-- 995 984-- 985 983-- 991
#>  [8] 982-- 983 981-- 982 978-- 985 976-- 993 975-- 987 975-- 981 974-- 997
#> [15] 973-- 980 971-- 989 971-- 985 970-- 988 969-- 987 969-- 984 969-- 980
#> [22] 969-- 970 968--1000 967-- 975 967-- 971 966-- 996 965-- 973 965-- 970
#> [29] 965-- 966 964-- 979 963-- 999 963-- 978 963-- 972 962-- 988 960-- 984
#> [36] 960-- 983 960-- 979 960-- 978 959-- 997 959-- 993 959-- 969 958-- 991
#> [43] 958-- 979 957-- 979 956-- 981 955-- 991 955-- 977 954-- 966 954-- 964
#> + ... omitted several edges
#> 

sample_correlated_sbm_pair_w_junk(1000, pref.matrix=pm, block.sizes=c(300,700), rho=0.5,
core.block.sizes=c(200,500))

#> $graph1
#> IGRAPH 35bda8e U--- 1000 17081 -- Stochastic block-model
#> + attr: name (g/c), loops (g/l)
#> + edges from 35bda8e:
#>  [1]  4-- 6 10--11  2--12  4--13  4--14  5--14 11--15  4--17  1--18  6--18
#> [11] 17--18 13--19  5--20 10--20  8--21 12--21 10--22  2--23  4--23  8--23
#> [21]  9--23 10--23 19--23  3--24  2--25  5--25  7--25 13--25 17--25 22--25
#> [31] 11--26 17--26 22--27  2--28 10--28 12--28 16--28 27--28  7--29 27--29
#> [41]  1--30 10--30 17--30 19--30 20--30 24--30 28--30  4--31  5--31  8--31
#> [51] 10--31 18--31 19--31  1--32  2--32  7--32  8--32 10--32 24--33  4--34
#> [61]  8--34 26--34 29--34 32--34  1--35  3--35  9--35 13--35 30--36 31--36
#> [71]  4--37  6--37 13--37 14--37  4--38  8--38 29--38 30--38  3--39 13--39
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH 594bbe8 U--- 1000 16782 -- Stochastic block-model
#> + attr: name_1 (g/c), name_2 (g/c), loops_1 (g/l), loops_2 (g/l), name
#> | (g/c), loops (g/l)
#> + edges from 594bbe8:
#>  [1] 997-- 999 990-- 998 989-- 997 988-- 999 986-- 993 986-- 987 984-- 991
#>  [8] 983-- 986 982-- 990 980-- 987 980-- 982 980-- 981 979-- 987 978-- 981
#> [15] 977-- 998 976-- 995 975-- 999 975-- 996 975-- 984 975-- 977 973-- 983
#> [22] 972-- 997 971-- 980 971-- 979 969-- 990 969-- 984 969-- 977 968-- 970
#> [29] 967-- 992 966-- 975 965-- 993 965-- 972 965-- 970 964-- 983 964-- 970
#> [36] 963-- 997 963-- 965 962-- 976 962-- 964 961--1000 961-- 982 961-- 980
#> [43] 960-- 982 960-- 972 960-- 969 960-- 968 960-- 964 959-- 999 959-- 990
#> + ... omitted several edges
#>