Sample graphs pair from stochastic block model

Sample a pair of random graphs from stochastic block model with correlation between two graphs being rho and edge probability being p.

sample_correlated_sbm_pair(
  n,
  pref.matrix,
  block.sizes,
  rho,
  permutation = 1:n,
  ...
)

sample_correlated_sbm_pair_w_junk(
  n,
  pref.matrix,
  block.sizes,
  rho,
  core.block.sizes,
  permutation = 1:n,
  ...
)

Arguments

n
An integer. Number of vertices in the graph.

pref.matrix
The matrix giving the Bernoulli rates. This is a K-by-K matrix, where k is the number of groups. The probability of creating an edge between vertices from groups i and j is given by element i,j. For undirected graphs, this matrix must be symmetric.

block.sizes
A numeric vector. Give the number of vertices in each group. The sum of the vector must match the number of vertices.

rho
A number. The target Pearson correlation between the adjacency matrices of the generated graphs. It must be in open (0,1) interval.

permutation
A numeric vector, permute second graph.

...
Passed to sample_correlated_sbm_pair and sample_correlated_sbm_pair_w_junk.

core.block.sizes
A numeric vector. Give the number of core vertices in each group. Entries should be smaller than block.sizes and the vector length should be the same as block.sizes.

Value

A list of two igraph object, named graph1 and graph2.

Examples

pm <- cbind( c(.1, .001), c(.001, .05) )
sample_correlated_sbm_pair(1000, pref.matrix=pm, block.sizes=c(300,700), rho=0.5)

#> $graph1
#> IGRAPH 4d7f83e U--- 1000 16889 -- Stochastic block-model
#> + attr: name (g/c), loops (g/l)
#> + edges from 4d7f83e:
#>  [1]  2-- 4  4-- 5  5-- 7  2--11 10--11  8--12 10--12 11--12 10--13  8--14
#> [11]  9--14 11--14  8--16  6--17  9--17 14--17 16--17  3--18 12--19 18--19
#> [21]  5--20  6--21  5--22 14--22 18--22  2--23  9--23 16--23 17--23 21--24
#> [31]  2--25 15--25 24--25 15--26 23--26 25--26  5--27 10--27 16--27 23--27
#> [41] 17--28 21--28  3--29 16--29 17--29 19--29 20--29 25--29  5--30 12--30
#> [51] 14--30  2--31 15--31 18--31 21--31 22--31  7--32 10--32 26--32 28--32
#> [61] 14--33 26--33  4--34  6--34 12--34 16--34 17--34 24--34 27--34  2--35
#> [71] 22--35 30--35  2--36  7--36 13--36 15--36 17--36 20--36 29--36  9--37
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH cc071a4 U--- 1000 17204 -- Stochastic block-model
#> + attr: name_1 (g/c), name_2 (g/c), loops_1 (g/l), loops_2 (g/l), name
#> | (g/c), loops (g/l)
#> + edges from cc071a4:
#>  [1] 995-- 999 995-- 998 990--1000 989-- 991 989-- 990 988-- 996 986-- 997
#>  [8] 985-- 991 981--1000 981-- 994 981-- 993 981-- 991 980--1000 978-- 998
#> [15] 978-- 994 976-- 988 975-- 986 975-- 981 974-- 997 974-- 991 974-- 988
#> [22] 973-- 996 973-- 986 973-- 983 973-- 975 972-- 976 970-- 986 969-- 972
#> [29] 968-- 997 968-- 994 966-- 994 966-- 975 964-- 976 964-- 973 963-- 980
#> [36] 963-- 977 963-- 969 962-- 980 961-- 976 961-- 964 960-- 992 960-- 989
#> [43] 960-- 988 960-- 982 960-- 981 960-- 972 958-- 996 958-- 987 958-- 985
#> + ... omitted several edges
#> 

sample_correlated_sbm_pair_w_junk(1000, pref.matrix=pm, block.sizes=c(300,700), rho=0.5,
core.block.sizes=c(200,500))

#> $graph1
#> IGRAPH 99dfdc7 U--- 1000 16744 -- Stochastic block-model
#> + attr: name (g/c), loops (g/l)
#> + edges from 99dfdc7:
#>  [1]  1-- 2  1-- 3  1-- 4  2-- 4  4-- 7  4-- 9  8--10  9--10  8--12  9--12
#> [11] 10--12 10--13  4--14 13--14  7--15 12--15  1--16  4--16  8--16 12--16
#> [21]  4--17 13--18  5--19  7--19 11--20 12--20 15--20 17--20  5--22  8--22
#> [31]  9--22 10--22 13--22 21--22  1--23 11--23  3--24  4--24 16--24 22--24
#> [41]  7--25 12--25  1--26  7--26  9--26  4--27  7--27  8--27 10--27 14--27
#> [51] 19--27  6--28  7--28 15--28 16--28 18--28 20--28 16--29 19--29 20--29
#> [61] 22--29  6--30 16--30  5--31 14--31  6--32  7--32 15--32  1--33  4--33
#> [71] 10--33 13--33 25--33 11--34  7--35 18--35 22--35 14--36 16--36 18--36
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH 2e70e2a U--- 1000 16956 -- Stochastic block-model
#> + attr: name_1 (g/c), name_2 (g/c), loops_1 (g/l), loops_2 (g/l), name
#> | (g/c), loops (g/l)
#> + edges from 2e70e2a:
#>  [1] 987-- 988 984-- 990 983-- 994 982-- 995 981-- 994 980-- 984 979-- 995
#>  [8] 978-- 997 977-- 978 976-- 994 976-- 987 975-- 980 974-- 989 974-- 982
#> [15] 973--1000 973-- 976 972-- 983 971-- 994 970--1000 970-- 976 969-- 970
#> [22] 968-- 997 967-- 994 967-- 980 965-- 986 964-- 999 964-- 985 963-- 999
#> [29] 962-- 973 961-- 983 961-- 974 960-- 998 960-- 983 960-- 971 959-- 996
#> [36] 959-- 990 959-- 981 958-- 997 958-- 983 958-- 973 957-- 988 957-- 987
#> [43] 957-- 975 956-- 987 956-- 973 956-- 966 955-- 963 954-- 983 954-- 982
#> + ... omitted several edges
#>