Sample graphs pair from stochastic block model

Sample a pair of random graphs from stochastic block model with correlation between two graphs being rho and edge probability being p.

sample_correlated_sbm_pair(
  n,
  pref.matrix,
  block.sizes,
  rho,
  permutation = 1:n,
  ...
)

sample_correlated_sbm_pair_w_junk(
  n,
  pref.matrix,
  block.sizes,
  rho,
  core.block.sizes,
  permutation = 1:n,
  ...
)

Arguments

n
An integer. Number of vertices in the graph.

pref.matrix
The matrix giving the Bernoulli rates. This is a K-by-K matrix, where k is the number of groups. The probability of creating an edge between vertices from groups i and j is given by element i,j. For undirected graphs, this matrix must be symmetric.

block.sizes
A numeric vector. Give the number of vertices in each group. The sum of the vector must match the number of vertices.

rho
A number. The target Pearson correlation between the adjacency matrices of the generated graphs. It must be in open (0,1) interval.

permutation
A numeric vector, permute second graph.

...
Passed to sample_correlated_sbm_pair and sample_correlated_sbm_pair_w_junk.

core.block.sizes
A numeric vector. Give the number of core vertices in each group. Entries should be smaller than block.sizes and the vector length should be the same as block.sizes.

Value

A list of two igraph object, named graph1 and graph2.

Examples

pm <- cbind( c(.1, .001), c(.001, .05) )
sample_correlated_sbm_pair(1000, pref.matrix=pm, block.sizes=c(300,700), rho=0.5)

#> $graph1
#> IGRAPH 4a65b51 U--- 1000 17013 -- Stochastic block-model
#> + attr: name (g/c), loops (g/l)
#> + edges from 4a65b51:
#>  [1]  2-- 3  4-- 6  2-- 7  7-- 9  8--11  6--13  8--13  5--15 11--15  7--16
#> [11] 11--16  6--17  9--17 11--17  8--18 15--18 17--18  6--19 10--19 16--19
#> [21] 17--19  3--20  8--20  6--21 11--21 16--21  2--22  8--22 13--22 21--22
#> [31]  1--23  5--23  7--23 13--24 16--24  6--25  8--25 11--25 22--25 13--26
#> [41] 14--26 20--26  2--27  9--27 26--27 15--28 18--28 19--29 22--29 19--30
#> [51] 21--30 29--31  2--32  8--32  2--33  5--33 21--33 11--34 25--34  1--35
#> [61]  4--35  7--35 27--35 34--35  8--36 13--36 19--36 23--36 24--36 28--36
#> [71] 30--36  2--37 15--37 20--37 21--37 32--37  4--38  5--38  8--38 24--38
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH a4f5028 U--- 1000 17071 -- Stochastic block-model
#> + attr: name_1 (g/c), name_2 (g/c), loops_1 (g/l), loops_2 (g/l), name
#> | (g/c), loops (g/l)
#> + edges from a4f5028:
#>  [1] 993-- 999 993-- 996 992-- 997 991--1000 989-- 997 989-- 994 987-- 995
#>  [8] 981-- 987 980-- 991 977-- 999 977-- 990 976-- 984 975-- 990 974-- 998
#> [15] 974-- 982 973-- 974 971-- 996 970-- 996 970-- 994 970-- 977 969-- 996
#> [22] 968-- 992 967-- 978 967-- 973 966-- 976 965-- 987 963-- 996 963-- 991
#> [29] 962--1000 962-- 981 962-- 969 961-- 989 961-- 985 961-- 979 961-- 962
#> [36] 960-- 998 960-- 991 960-- 974 960-- 961 959-- 997 958-- 999 958-- 997
#> [43] 958-- 970 958-- 961 957-- 986 957-- 976 956-- 980 956-- 971 956-- 967
#> + ... omitted several edges
#> 

sample_correlated_sbm_pair_w_junk(1000, pref.matrix=pm, block.sizes=c(300,700), rho=0.5,
core.block.sizes=c(200,500))

#> $graph1
#> IGRAPH 6308b35 U--- 1000 16757 -- Stochastic block-model
#> + attr: name (g/c), loops (g/l)
#> + edges from 6308b35:
#>  [1]  4-- 6  5-- 6  4-- 7  3--10  1--12 10--12  8--13 11--14 10--16  9--17
#> [11]  1--18 11--19  8--20 14--20 18--20  9--21 16--21 12--22 13--22 10--23
#> [21]  3--24  6--24 10--25 15--25  1--27  4--27  5--27 10--27 25--27  6--28
#> [31] 20--28 27--28 11--29 15--29 21--29  3--30 15--30 22--30 26--30  2--31
#> [41]  6--31  3--32 18--32  3--34  5--34 13--34 15--34 19--34 33--34 15--35
#> [51] 19--35 21--35 27--35 33--35  1--36 22--36 27--36 28--36 35--36  7--37
#> [61] 15--37 16--37 17--37  1--38 26--38 37--38 12--39 26--39 33--39 38--40
#> [71] 14--41 18--41 22--41 32--41 34--41 41--42  2--43 16--43 26--43 32--43
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH f32d140 U--- 1000 16893 -- Stochastic block-model
#> + attr: name_1 (g/c), name_2 (g/c), loops_1 (g/l), loops_2 (g/l), name
#> | (g/c), loops (g/l)
#> + edges from f32d140:
#>  [1] 998--1000 991-- 992 986-- 992 984-- 998 984-- 990 983-- 993 982-- 993
#>  [8] 982-- 990 978-- 995 977-- 998 974-- 982 972-- 993 972-- 982 970-- 980
#> [15] 969-- 976 968-- 975 967-- 987 965-- 993 964--1000 964-- 970 962-- 988
#> [22] 962-- 977 961-- 985 960-- 985 960-- 984 959-- 967 958-- 995 958-- 987
#> [29] 958-- 967 958-- 959 957-- 987 957-- 960 956--1000 956-- 986 956-- 969
#> [36] 956-- 961 955-- 996 954-- 992 953-- 999 953-- 994 952-- 993 952-- 981
#> [43] 952-- 965 951-- 973 950-- 993 950-- 983 950-- 969 950-- 965 949-- 980
#> + ... omitted several edges
#>