Sample graphs pair from stochastic block model

Sample a pair of random graphs from stochastic block model with correlation between two graphs being rho and edge probability being p.

sample_correlated_sbm_pair(
  n,
  pref.matrix,
  block.sizes,
  rho,
  permutation = 1:n,
  ...
)

sample_correlated_sbm_pair_w_junk(
  n,
  pref.matrix,
  block.sizes,
  rho,
  core.block.sizes,
  permutation = 1:n,
  ...
)

Arguments

n
An integer. Number of vertices in the graph.

pref.matrix
The matrix giving the Bernoulli rates. This is a K-by-K matrix, where k is the number of groups. The probability of creating an edge between vertices from groups i and j is given by element i,j. For undirected graphs, this matrix must be symmetric.

block.sizes
A numeric vector. Give the number of vertices in each group. The sum of the vector must match the number of vertices.

rho
A number. The target Pearson correlation between the adjacency matrices of the generated graphs. It must be in open (0,1) interval.

permutation
A numeric vector, permute second graph.

...
Passed to sample_correlated_sbm_pair and sample_correlated_sbm_pair_w_junk.

core.block.sizes
A numeric vector. Give the number of core vertices in each group. Entries should be smaller than block.sizes and the vector length should be the same as block.sizes.

Value

A list of two igraph object, named graph1 and graph2.

Examples

pm <- cbind( c(.1, .001), c(.001, .05) )
sample_correlated_sbm_pair(1000, pref.matrix=pm, block.sizes=c(300,700), rho=0.5)

#> $graph1
#> IGRAPH ef48da1 U--- 1000 16870 -- Stochastic block-model
#> + attr: name (g/c), loops (g/l)
#> + edges from ef48da1:
#>  [1]  2-- 3  1-- 6  3-- 9  3--10  9--10  5--11  1--12  3--12  5--12  8--12
#> [11]  6--13  9--14 13--14  9--16 15--16  4--17  5--17  2--18  7--18 11--18
#> [21]  3--19  7--19 17--19 16--20 12--21  2--22  9--22 10--22  2--23 18--24
#> [31]  5--25 23--26  9--27 16--27 20--27  9--28 26--28  4--29 13--29 22--29
#> [41] 26--29 25--30 27--30  2--31  4--31 16--31 20--31 21--31 24--31 19--32
#> [51] 24--32 29--32 31--32  3--33 28--34 31--34  1--35 20--35 30--35 35--36
#> [61]  3--37 13--37 16--37 18--37 26--37 30--37 36--37 11--38 23--38 37--38
#> [71] 11--39 23--39 31--39 32--39  6--40 14--40  3--41  7--41  9--41 16--41
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH 45a9278 U--- 1000 17017 -- Stochastic block-model
#> + attr: name_1 (g/c), name_2 (g/c), loops_1 (g/l), loops_2 (g/l), name
#> | (g/c), loops (g/l)
#> + edges from 45a9278:
#>  [1] 997-- 998 994-- 995 992-- 997 991--1000 988--1000 988-- 996 987-- 995
#>  [8] 987-- 992 986-- 997 986-- 993 986-- 990 985-- 989 982-- 993 982-- 990
#> [15] 982-- 988 980-- 999 980-- 990 980-- 988 979-- 994 976-- 995 975-- 980
#> [22] 974-- 987 974-- 982 972-- 999 972-- 977 972-- 976 971-- 997 971-- 974
#> [29] 970-- 975 968-- 982 967-- 996 966-- 987 966-- 985 966-- 982 965-- 975
#> [36] 965-- 973 963-- 995 963-- 988 963-- 987 962-- 999 962-- 994 962-- 991
#> [43] 962-- 987 962-- 982 962-- 980 961-- 994 960-- 994 959-- 976 959-- 972
#> + ... omitted several edges
#> 

sample_correlated_sbm_pair_w_junk(1000, pref.matrix=pm, block.sizes=c(300,700), rho=0.5,
core.block.sizes=c(200,500))

#> $graph1
#> IGRAPH 31f2c2b U--- 1000 16725 -- Stochastic block-model
#> + attr: name (g/c), loops (g/l)
#> + edges from 31f2c2b:
#>  [1]  4-- 6  3-- 9  6-- 9  4--10  7--10 10--11  7--13 11--15  8--16 12--16
#> [11]  7--18  9--18 11--18  9--20 12--20  3--21  8--21 16--21 20--21  2--22
#> [21]  3--22  9--22  3--23  4--23 21--23 22--23  1--24 14--24  9--25 13--26
#> [31] 22--26 11--27 21--27 24--27 12--28 19--28 21--28 27--28  2--29  5--29
#> [41]  9--29 12--29 23--29 13--30 23--30 27--30  1--31  3--32 17--32 30--32
#> [51]  2--33 25--33 32--33 11--34 15--34 20--34 27--34  5--35 15--35 21--35
#> [61] 33--35 31--36 32--36 36--37  9--38 37--38  6--40 20--40 24--40  5--41
#> [71]  9--41 12--41 33--41 37--41 12--42 32--42  5--43 11--43 12--43 14--43
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH e6d43be U--- 1000 16879 -- Stochastic block-model
#> + attr: name_1 (g/c), name_2 (g/c), loops_1 (g/l), loops_2 (g/l), name
#> | (g/c), loops (g/l)
#> + edges from e6d43be:
#>  [1] 997-- 998 995--1000 994-- 998 993-- 996 990-- 996 989-- 999 988-- 998
#>  [8] 986-- 993 984-- 998 982-- 991 982-- 988 981-- 984 980-- 995 980-- 985
#> [15] 979-- 995 979-- 990 977-- 990 976-- 993 976-- 983 976-- 982 972-- 994
#> [22] 972-- 990 972-- 985 969-- 983 969-- 973 968-- 985 968-- 977 968-- 973
#> [29] 967-- 990 967-- 984 967-- 976 967-- 975 966-- 998 966-- 993 965-- 983
#> [36] 965-- 980 965-- 979 965-- 973 964-- 986 963-- 995 963-- 992 963-- 990
#> [43] 963-- 977 962-- 994 962-- 979 961--1000 961-- 981 961-- 976 959-- 984
#> + ... omitted several edges
#>