Sample graphs pair from stochastic block model

Sample a pair of random graphs from stochastic block model with correlation between two graphs being rho and edge probability being p.

sample_correlated_sbm_pair(
  n,
  pref.matrix,
  block.sizes,
  rho,
  permutation = 1:n,
  ...
)

sample_correlated_sbm_pair_w_junk(
  n,
  pref.matrix,
  block.sizes,
  rho,
  core.block.sizes,
  permutation = 1:n,
  ...
)

Arguments

n
An integer. Number of vertices in the graph.

pref.matrix
The matrix giving the Bernoulli rates. This is a K-by-K matrix, where k is the number of groups. The probability of creating an edge between vertices from groups i and j is given by element i,j. For undirected graphs, this matrix must be symmetric.

block.sizes
A numeric vector. Give the number of vertices in each group. The sum of the vector must match the number of vertices.

rho
A number. The target Pearson correlation between the adjacency matrices of the generated graphs. It must be in open (0,1) interval.

permutation
A numeric vector, permute second graph.

...
Passed to sample_correlated_sbm_pair and sample_correlated_sbm_pair_w_junk.

core.block.sizes
A numeric vector. Give the number of core vertices in each group. Entries should be smaller than block.sizes and the vector length should be the same as block.sizes.

Value

A list of two igraph object, named graph1 and graph2.

Examples

pm <- cbind( c(.1, .001), c(.001, .05) )
sample_correlated_sbm_pair(1000, pref.matrix=pm, block.sizes=c(300,700), rho=0.5)

#> $graph1
#> IGRAPH 640b17a U--- 1000 17005 -- Stochastic block-model
#> + attr: name (g/c), loops (g/l)
#> + edges from 640b17a:
#>  [1]  1-- 5  7-- 8  6-- 9  1--10  2--10  5--10  2--13  2--15  5--15  8--16
#> [11] 14--16  3--17  7--17 11--17 13--17  4--18  5--18  9--18 14--18 18--19
#> [21] 16--20 17--20 14--22 17--23 23--24  3--25 12--25 23--25 24--25  9--26
#> [31] 13--26 20--26  6--27  7--27 17--27 18--27 12--28 17--28  9--29 20--29
#> [41] 21--30 25--30 26--30 11--31 23--31  1--32 20--32  8--33 20--33  7--34
#> [51] 21--34 23--34  2--35  7--35 13--36 30--36  4--37  9--37 24--37 32--37
#> [61] 35--37  1--38  3--38  7--38 11--38 16--38 25--38  9--39 20--39 23--39
#> [71] 25--39 27--39 31--39  1--40 11--40 15--40 24--40 27--40 33--40 12--41
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH d823257 U--- 1000 16960 -- Stochastic block-model
#> + attr: name_1 (g/c), name_2 (g/c), loops_1 (g/l), loops_2 (g/l), name
#> | (g/c), loops (g/l)
#> + edges from d823257:
#>  [1] 992-- 998 992-- 997 992-- 994 989--1000 989-- 999 988-- 995 988-- 992
#>  [8] 988-- 989 987--1000 984-- 999 984-- 997 983-- 990 983-- 987 979-- 996
#> [15] 979-- 990 979-- 985 979-- 981 978-- 992 977-- 983 974-- 984 972-- 999
#> [22] 972-- 994 972-- 993 970-- 984 968-- 994 968-- 987 966-- 994 966-- 977
#> [29] 965-- 983 965-- 977 963-- 969 961-- 996 961-- 993 961-- 977 960-- 977
#> [36] 959-- 989 958-- 990 958-- 976 957-- 990 957-- 983 957-- 980 957-- 976
#> [43] 957-- 969 957-- 964 956-- 994 956-- 991 955-- 990 955-- 978 955-- 973
#> + ... omitted several edges
#> 

sample_correlated_sbm_pair_w_junk(1000, pref.matrix=pm, block.sizes=c(300,700), rho=0.5,
core.block.sizes=c(200,500))

#> $graph1
#> IGRAPH ad2210b U--- 1000 16847 -- Stochastic block-model
#> + attr: name (g/c), loops (g/l)
#> + edges from ad2210b:
#>  [1]  1-- 3  4-- 7  6-- 7  4-- 8  1-- 9  4-- 9  3--11  4--11  7--12  1--14
#> [11]  7--14  5--15 12--15  2--16  6--17  7--17 12--17 17--18  9--19 16--19
#> [21]  5--20  2--21  4--21  5--21  7--21 13--21 16--22 17--22 11--23  3--24
#> [31]  6--24  8--24  4--25 13--25  3--26 15--26 17--26  4--27  9--27 12--27
#> [41] 24--27  7--28  8--28 23--28  3--29 11--29 22--29  9--30 18--30  2--32
#> [51]  9--32 14--32 21--32 16--34  7--35 27--35  3--36  6--36  7--36 10--36
#> [61] 20--36 25--36 35--36 12--37 10--38 13--38 18--38 37--38 12--39 14--39
#> [71] 20--39 25--39  1--40  9--40 27--40 31--40 33--40  3--41 20--41  4--42
#> + ... omitted several edges
#> 
#> $graph2
#> IGRAPH e9dd8ac U--- 1000 16904 -- Stochastic block-model
#> + attr: name_1 (g/c), name_2 (g/c), loops_1 (g/l), loops_2 (g/l), name
#> | (g/c), loops (g/l)
#> + edges from e9dd8ac:
#>  [1] 993--1000 991-- 992 990-- 998 990-- 993 989-- 997 988-- 990 984-- 987
#>  [8] 982-- 994 981-- 986 980-- 995 979-- 988 978-- 992 978-- 987 978-- 982
#> [15] 977-- 980 976-- 990 976-- 986 976-- 977 975-- 995 974-- 991 973-- 985
#> [22] 972-- 992 971-- 987 969-- 991 969-- 990 968-- 997 968-- 979 968-- 969
#> [29] 967-- 990 967-- 981 966-- 995 965-- 974 964-- 994 963-- 986 963-- 977
#> [36] 963-- 972 962-- 998 960-- 996 960-- 973 959-- 991 958-- 959 957-- 989
#> [43] 957-- 986 957-- 975 956-- 961 955-- 979 955-- 975 955-- 966 953-- 996
#> + ... omitted several edges
#>