Split an igraph object into aligned graphs by attribute

Given an igraph object and an edge attribute, this function finds all unique values of the edge attribute in the graph and returns a list of igraph objects on the same vertex set where each element of the list has a graph containing only those edges with specified attributed.

split_igraph(g, e_attr, strip_vertex_attr = FALSE)

Arguments

g
An igraph object

e_attr
the name of an edge attribute in g

strip_vertex_attr
Whether to remove all vertex attribute from the new graphs

Value

A named list of igraph objects

Examples

g <- igraph::sample_gnm(20, 60)
igraph::E(g)$color <- 
  sample(c("red", "green"), 60, replace = TRUE)
split_igraph(g, "color")

#> $green
#> IGRAPH 7daf4a2 U--- 20 25 -- Erdos renyi (gnm) graph
#> + attr: name (g/c), type (g/c), loops (g/l), m (g/n), color (e/c)
#> + edges from 7daf4a2:
#>  [1]  3-- 4  1-- 5  1-- 6  3-- 7  7-- 8  2-- 9  3--10  6--10  5--11 10--11
#> [11]  2--12  3--12 11--13  5--14  2--15 13--16  4--17  9--17  1--18 11--19
#> [21] 12--19 14--19 15--19 16--19  5--20
#> 
#> $red
#> IGRAPH a754e9d U--- 20 35 -- Erdos renyi (gnm) graph
#> + attr: name (g/c), type (g/c), loops (g/l), m (g/n), color (e/c)
#> + edges from a754e9d:
#>  [1]  1-- 3  2-- 5  4-- 6  5-- 6  2-- 8  3-- 8  6-- 9  8-- 9  5--10  6--12
#> [11]  8--12 10--12 11--12  7--13  8--13  3--14 10--14 11--14  2--16  4--16
#> [21]  6--16 10--16 12--17  3--18 10--18 13--18 15--18  8--19 17--19 18--19
#> [31]  3--20  4--20  9--20 10--20 16--20
#>