"SPLR" Methods

Methods for the splrMatrix class. Most behave like Matrix methods though things like output show the decomposition. Use as.matrix to see the computed dense matrix.

Usage

# S4 method for splrMatrix
show(object)

# S4 method for splrMatrix
print(x)

# S4 method for splrMatrix,splrMatrix
%*%(x, y)

# S4 method for splrMatrix,matrix_list
%*%(x, y)

# S4 method for matrix_list,splrMatrix
%*%(x, y)

# S4 method for Matrix,splrMatrix
%*%(x, y)

# S4 method for matrix,splrMatrix
%*%(x, y)

# S4 method for numeric,splrMatrix
%*%(x, y)

# S4 method for numLike,splrMatrix
%*%(x, y)

# S4 method for ANY,splrMatrix
%*%(x, y)

# S4 method for splrMatrix
dim(x)

# S4 method for splrMatrix
length(x)

# S4 method for splrMatrix,Matrix
%*%(x, y)

# S4 method for splrMatrix,matrix
%*%(x, y)

# S4 method for splrMatrix,numeric
%*%(x, y)

# S4 method for splrMatrix,numLike
%*%(x, y)

# S4 method for splrMatrix,ANY
%*%(x, y)

# S4 method for splrMatrix,splrMatrix
*(e1, e2)

# S4 method for Matrix,splrMatrix
*(e1, e2)

# S4 method for splrMatrix,ddiMatrix
*(e1, e2)

# S4 method for ddiMatrix,splrMatrix
*(e1, e2)

# S4 method for matrix,splrMatrix
*(e1, e2)

# S4 method for numeric,splrMatrix
*(e1, e2)

# S4 method for ANY,splrMatrix
*(e1, e2)

# S4 method for splrMatrix,matrix
*(e1, e2)

# S4 method for splrMatrix,Matrix
*(e1, e2)

# S4 method for splrMatrix,numeric
*(e1, e2)

# S4 method for splrMatrix,ANY
*(e1, e2)

# S4 method for splrMatrix,matrix
/(e1, e2)

# S4 method for splrMatrix,Matrix
/(e1, e2)

# S4 method for splrMatrix,ANY
/(e1, e2)

# S4 method for splrMatrix,splrMatrix
+(e1, e2)

# S4 method for splrMatrix,splrMatrix
-(e1, e2)

# S4 method for splrMatrix,Matrix
+(e1, e2)

# S4 method for splrMatrix,numeric
+(e1, e2)

# S4 method for splrMatrix,ANY
+(e1, e2)

# S4 method for splrMatrix,missing
-(e1, e2 = NULL)

# S4 method for splrMatrix,Matrix
-(e1, e2)

# S4 method for splrMatrix,ddiMatrix
-(e1, e2)

# S4 method for splrMatrix,numeric
-(e1, e2)

# S4 method for splrMatrix,ANY
-(e1, e2)

# S4 method for Matrix,splrMatrix
+(e1, e2)

# S4 method for numeric,splrMatrix
+(e1, e2)

# S4 method for ANY,splrMatrix
+(e1, e2)

# S4 method for Matrix,splrMatrix
-(e1, e2)

# S4 method for numeric,splrMatrix
-(e1, e2)

# S4 method for ANY,splrMatrix
-(e1, e2)

# S4 method for splrMatrix,character
norm(x, type, ...)

# S4 method for splrMatrix
rowSums(x, na.rm = FALSE, dims = 1, ...)

# S4 method for splrMatrix
colSums(x, na.rm = FALSE, dims = 1, ...)

# S4 method for splrMatrix
rowMeans(x, na.rm = FALSE, dims = 1, ...)

# S4 method for splrMatrix
colMeans(x, na.rm = FALSE, dims = 1, ...)

# S4 method for splrMatrix,ANY
sum(x, ..., na.rm = FALSE)

# S4 method for splrMatrix
mean(x, ...)

# S4 method for splrMatrix,missing,missing,missing
[(x, i = NULL, j = NULL, drop = NULL)

# S4 method for splrMatrix,numeric,numeric,logical
[(x, i, j, ..., drop = TRUE)

# S4 method for splrMatrix,missing,numeric,logical
[(x, i, j, ..., drop = TRUE)

# S4 method for splrMatrix,missing,numeric,missing
[(x, i, j, ..., drop = TRUE)

# S4 method for splrMatrix,missing,logical,logical
[(x, i, j, ..., drop = TRUE)

# S4 method for splrMatrix,missing,logical,missing
[(x, i, j, ..., drop = TRUE)

# S4 method for splrMatrix,numeric,missing,logical
[(x, i, j, ..., drop = TRUE)

# S4 method for splrMatrix,numeric,missing,missing
[(x, i, j, ..., drop = TRUE)

# S4 method for splrMatrix,logical,missing,logical
[(x, i, j, ..., drop = TRUE)

# S4 method for splrMatrix,logical,missing,missing
[(x, i, j, ..., drop = TRUE)

# S4 method for splrMatrix,numeric,ANY,logical
[(x, i, j, ..., drop = TRUE)

# S4 method for splrMatrix,numeric,logical,logical
[(x, i, j, ..., drop = TRUE)

# S4 method for splrMatrix,numeric,ANY,missing
[(x, i, j, ..., drop = FALSE)

# S4 method for splrMatrix,logical,ANY,ANY
[(x, i, j, ..., drop = TRUE)

# S4 method for splrMatrix,logical,ANY,missing
[(x, i, j, ..., drop = TRUE)

# S4 method for splrMatrix,logical,numeric,missing
[(x, i, j, ..., drop = TRUE)

# S4 method for splrMatrix,logical,numeric,logical
[(x, i, j, ..., drop = TRUE)

# S4 method for splrMatrix,missing,missing,missing
[(x, i = NULL, j = NULL, drop = NULL)

# S4 method for splrMatrix,matrix,missing,missing
[(x, i, j, ..., drop = TRUE)

# S4 method for splrMatrix,numeric,numeric,ANY
[(x, i, j, ...) <- value

# S4 method for splrMatrix,numeric,missing,ANY
[(x, i, j, ...) <- value

# S4 method for splrMatrix,missing,numeric,ANY
[(x, i, j, ...) <- value

# S4 method for Matrix,ANY,ANY,splrMatrix
[(x, i, j, ...) <- value

# S4 method for splrMatrix
dim(x)

# S4 method for splrMatrix
str(object)

# S4 method for splrMatrix
t(x)

# S4 method for splrMatrix
diag(x)

Arguments

object

As in "Matrix"

x

As in "Matrix"

y

As in "Matrix"

e1

As in "Matrix"

e2

As in "Matrix"

type

As in "Matrix"

...

As in "Matrix"

na.rm

As in "Matrix"

dims

As in "Matrix"

i

As in "Matrix"

j

As in "Matrix"

drop

As in "Matrix"

value

As in "Matrix"

Value

Results of matrix operations for splrMatrix objects. Attempts are made such that the returned object is stored efficiently, either as a splrMatrix or sparse Matrix.