Start matrix initialization

initialize the start matrix for graph matching iteration.

bari_start(nns, ns = 0, soft_seeds = NULL)

rds_sinkhorn_start(nns, ns = 0, soft_seeds = NULL, distribution = "runif")

rds_perm_bari_start(nns, ns = 0, soft_seeds = NULL, g = 1, is_splr = TRUE)

rds_from_sim_start(nns, ns = 0, soft_seeds = NULL, sim)

Arguments

nns
An integer. Number of non-seeds.

ns
An integer. Number of hard seeds.

soft_seeds
A vector, a matrix or a data frame. If there is no error in soft seeds, input can be a vector of soft seed indices in \(G_1\). Or if there is error in soft seeds, input in the form of a matrix or a data frame, with the first column being the indices of \(G_1\) and the second column being the corresponding indices of \(G_2\). Note that if there are seeds in graphs, seeds should be put before non-seeds.

distribution
A character. Specify the distribution from which the random doubly stochastic matrix is sampled. Should input the name of the function for generating random deviates from that distribution.

g
A number. Specified in the range of [0, 1] to set weights to random permutation matrix and barycenter matrix.

is_splr
should we return a splr matrix? (default = TRUE)

sim
nns x nns non-negative matrix.

Value

bari_start returns a nns-by-nns matrix with 1's corresponding to the adaptive seeds and being bari-centered at other places.

rds_sinkhorn_start returns a nns-by-nns doubly stochastic matrix with 1's corresponding to adaptive seeds.

rds_perm_bari returns a nns-by-nns doubly stochastic matrix with 1's corresponding to adaptive seeds.

rds_from_sim_start returns a doubly stochastic Matrix given by sinkhorn algorithm applied to a matrix of iid log-normal with mu=sim. Note, this ignores soft seeds.

Examples

## Case without soft seeds
bari_start(3)

#> Sparse part
#> 3 x 3 sparse Matrix of class "dsCMatrix"
#>           
#> [1,] . . .
#> [2,] . . .
#> [3,] . . .
#> plus left factor
#> 3 x 1 Matrix of class "dgeMatrix"
#>      [,1]
#> [1,]    1
#> [2,]    1
#> [3,]    1
#> times right factor transpose
#> 3 x 1 Matrix of class "dgeMatrix"
#>           [,1]
#> [1,] 0.3333333
#> [2,] 0.3333333
#> [3,] 0.3333333

## Case with correct soft seeds and input is a vector
bari_start(nns=5, ns=3, soft_seeds=c(5, 7, 8))

#> m[ <ij-matrix> ] <- v: inefficiently treating single elements

#> Sparse part
#> 5 x 5 sparse Matrix of class "dsCMatrix"
#>               
#> [1,] . . . . .
#> [2,] . 1 . . .
#> [3,] . . . . .
#> [4,] . . . 1 .
#> [5,] . . . . 1
#> plus left factor
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>       
#> [1,] 1
#> [2,] .
#> [3,] 1
#> [4,] .
#> [5,] .
#> times right factor transpose
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>         
#> [1,] 0.5
#> [2,] .  
#> [3,] 0.5
#> [4,] .  
#> [5,] .  

## Case with erroneous soft seeds and the input is a matrix
bari_start(nns=5, soft_seeds=matrix(c(2, 4, 2, 3), nrow=2))

#> Sparse part
#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>               
#> [1,] . . . . .
#> [2,] . 1 . . .
#> [3,] . . . . .
#> [4,] . . 1 . .
#> [5,] . . . . .
#> plus left factor
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>       
#> [1,] 1
#> [2,] .
#> [3,] 1
#> [4,] .
#> [5,] 1
#> times right factor transpose
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>               
#> [1,] 0.3333333
#> [2,] .        
#> [3,] .        
#> [4,] 0.3333333
#> [5,] 0.3333333

## Case without soft seeds
rds_sinkhorn_start(5)

#> 5 x 5 Matrix of class "dgeMatrix"
#>           [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.2964893 0.2694481 0.08613913 0.2408804 0.1070430
#> [2,] 0.1300198 0.0857632 0.29006326 0.1628831 0.3312706
#> [3,] 0.0548671 0.1681663 0.23534985 0.3047860 0.2368307
#> [4,] 0.3323903 0.1886302 0.25158475 0.1265853 0.1008094
#> [5,] 0.1862334 0.2879922 0.13686300 0.1648652 0.2240462

## Case with soft seeds and the input is a data frame
rds_sinkhorn_start(nns=5,
   soft_seeds = as.data.frame(matrix(c(2, 4, 2, 3), nrow=2)),
   distribution = "rnorm")

#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>                                        
#> [1,] 0.3078137 . . 0.07183856 0.6203477
#> [2,] .         1 . .          .        
#> [3,] 0.1786291 . . 0.62565311 0.1957178
#> [4,] .         . 1 .          .        
#> [5,] 0.5135573 . . 0.30250796 0.1839348

## Case without soft seeds
rds_perm_bari_start(nns=5)

#> Sparse part
#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>                                                  
#> [1,] 0.190982 .        .        .        .       
#> [2,] .        .        .        0.190982 .       
#> [3,] .        .        0.190982 .        .       
#> [4,] .        0.190982 .        .        .       
#> [5,] .        .        .        .        0.190982
#> plus left factor
#> 5 x 1 Matrix of class "dgeMatrix"
#>          [,1]
#> [1,] 0.809018
#> [2,] 0.809018
#> [3,] 0.809018
#> [4,] 0.809018
#> [5,] 0.809018
#> times right factor transpose
#> 5 x 1 Matrix of class "dgeMatrix"
#>      [,1]
#> [1,]  0.2
#> [2,]  0.2
#> [3,]  0.2
#> [4,]  0.2
#> [5,]  0.2

## Case with soft seeds and the input is a data frame
rds_perm_bari_start(nns=5, ns=0, soft_seeds=as.data.frame(matrix(c(2, 4, 2, 3), nrow=2)))

#> Sparse part
#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>                                       
#> [1,] .         . . 0.5910224 .        
#> [2,] .         1 . .         .        
#> [3,] 0.5910224 . . .         .        
#> [4,] .         . 1 .         .        
#> [5,] .         . . .         0.5910224
#> plus left factor
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>               
#> [1,] 0.4089776
#> [2,] .        
#> [3,] 0.4089776
#> [4,] .        
#> [5,] 0.4089776
#> times right factor transpose
#> 5 x 1 sparse Matrix of class "dgCMatrix"
#>               
#> [1,] 0.3333333
#> [2,] .        
#> [3,] .        
#> [4,] 0.3333333
#> [5,] 0.3333333

sim <- Matrix::rsparsematrix(10, 10, .4,
 rand.x = function(n) rep(1,n))
start_sparse <- rds_from_sim_start(10, sim = sim)
start_dense <- rds_from_sim_start(10, sim = as.matrix(sim))